==> 1_cleaning.Rmd <==
---
title: "Data_cleaning"
date: "21/10/2020"
output: 
  bookdown::html_document2:
    highlight: tango
    number_sections: yes
    theme: united
    code_folding: hide
    toc: yes
---

```{r setup2, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)

if (!require(pacman)) install.packages("pacman")

pacman::p_load(tidyverse, DT, knitr, lazyeval, labelled, 
               readxl, googledrive, kableExtra, estimatr, 
               ggforce, magrittr, RColorBrewer, readstata13, 
               metaplus, sjlabelled, purrr)

```


```{r import_studies_info, include = FALSE}
# Read in studies file with information by countries especially ids

studies <- read_excel("2_input_data/studies_info.xlsx", sheet = "sample")
```

```{r real_data, echo=FALSE, message=FALSE, warning=FALSE}

bfa <- 
  read.dta13("1_country_data/bfa/BF_vaccine_data.dta")  %>%
  dplyr::as_tibble() %>%
  dplyr::mutate(
    take_vaccine = recode(take_vaccine, 
                          "Ne sait pas" = "DK",
                          "Refus de répondre" = NA_character_,
                          "Oui" = "Yes", 
                          "Non" = "No"),
    study = "bfa_recovr",
    no_vaccine_666 = no_vaccine__666,
    yes_vaccine_666 = yes_vaccine__666,
    yes_vaccine_1 = ifelse(yes_vaccine_2 == 1, 1, yes_vaccine_1),
    yes_vaccine_2 = yes_vaccine_3,
    yes_vaccine_3 = yes_vaccine_4,
    yes_vaccine_4 = yes_vaccine_5,
    yes_vaccine_5 = yes_vaccine_6,
    trust_vaccine_1 = 
      ifelse(trust_vaccine == "Membres de la famille", 1, 0),
    trust_vaccine_2 = 
      ifelse(trust_vaccine == "Les amis que vous voyez et à qui vous parlez", 1, 0),
    trust_vaccine_2 = 
      ifelse(trust_vaccine == "Les amis que vous vous êtes faits en ligne", 1, trust_vaccine_2),
    trust_vaccine_3 = 
      ifelse(trust_vaccine == "Les chefs religieux", 1, 0),
    trust_vaccine_4 = 
      ifelse(trust_vaccine == "Personne célèbre" , 1, 0),
    trust_vaccine_5 = 
      ifelse(trust_vaccine == "Médecins ou autre personnel d'un centre de santé communautaire", 1, 0),
    trust_vaccine_6 = 
      ifelse(trust_vaccine == "Conseil de l'Institut National de la Santé", 1, 0),
    trust_vaccine_7 = 
      ifelse(trust_vaccine == "Guérisseurs traditionnels", 1, 0),
    trust_vaccine_8 = 
      ifelse(trust_vaccine == "Médias traditionnels (journaux, radio)", 1, 0),
    trust_vaccine_9 = 
      ifelse(trust_vaccine == "Groupes de discussion médicale en ligne", 1, 0),
    trust_vaccine_dk = 
      ifelse(trust_vaccine == "Ne sait pas", 1, 0),
    trust_vaccine_refuse = 
      ifelse(trust_vaccine == "Refuse de répondre", 1, 0),
    trust_vaccine_nr = 
      ifelse(is.na(trust_vaccine), 1, 0),
    trust_vaccine_666 = 
      ifelse(trust_vaccine == "Aucun de ces éléments/une autre personne- préciser", 1, 0),
    vaccine_safe = 
      ifelse(vacc_safe == "Tout à fait d'accord" | vacc_safe == "Plutôt d'accord" ,1,0),
    vaccine_effective = 
      ifelse(vacc_effective == "Tout à fait d'accord" | vacc_effective == "Plutôt d'accord" , 1, 0),
    educ = droplevels(educ),
    educ = recode(educ, 
                  "None" = "Less than secondary", 
                  "Less than primary" = "Less than secondary",
                  "Completed primary" = "Less than secondary",
                  "No Formal schooling" = "Less than secondary",
                  "Completed secondary"  =  "Secondary",
                  "Completed tertiary" = "More than secondary",
                  "MBA" = "More than secondary",
                  "Master" = "More than secondary",
                  "more than PhD" = "More than secondary")) %>%
  dplyr::select(
    study, age, gender, educ, take_vaccine, 
    starts_with(c("yes_vaccine", "no_vaccine", "trust_vaccine_")),
    vaccine_safe, vaccine_effective, weight = ps_weight) %>%
  dplyr::mutate(across(c(yes_vaccine_1:trust_vaccine_666), ~replace(., is.na(.), 0))) 

col <-  
  read.dta13("1_country_data/col/COL_vaccine_data.dta") %>% 
  dplyr::as_tibble() %>% 
  dplyr::select(-gov_reaction, -market, -wash_hands, -wear_mask, -hh_size, -caseid) %>%
  dplyr::mutate(
    yes_vaccine_1 = ifelse(yes_vaccine_2 == 1, 1, yes_vaccine_1),
    yes_vaccine_2 = yes_vaccine_3,
    yes_vaccine_3 = yes_vaccine_4,
    yes_vaccine_4 = yes_vaccine_5,
    yes_vaccine_5 = yes_vaccine_6,
    no_vaccine__666 = ifelse(no_vaccine_10 == 1 | no_vaccine_11 == 1, 1, no_vaccine__666)) %>%
  dplyr::rename(
    yes_vaccine_666 = yes_vaccine__666,
    no_vaccine_666 = no_vaccine__666) %>%
  dplyr::mutate(
    study = "col_recovr",
    take_vaccine = droplevels(take_vaccine),
    educ = droplevels(educ),
    educ = recode(educ, 
                  "No formal schooling" = "Less than secondary", 
                  "Less than primary" = "Less than secondary",
                  "Primary" = "Less than secondary",
                  "Tertiary vocational I" = "More than secondary",
                  "Tertiary vocational II" = "More than secondary",
                  "Tertiary university" = "More than secondary",
                  "Master or higher" = "More than secondary"),
    yes_vaccine_1 = ifelse(is.na(yes_vaccine_1), 0, yes_vaccine_1),
    yes_vaccine_2 = ifelse(is.na(yes_vaccine_2), 0, yes_vaccine_2),
    yes_vaccine_3 = ifelse(is.na(yes_vaccine_3), 0, yes_vaccine_3),
    yes_vaccine_4 = ifelse(is.na(yes_vaccine_4), 0, yes_vaccine_4),
    yes_vaccine_666 = ifelse(is.na(yes_vaccine_666), 0, yes_vaccine_666),
    no_vaccine_1 = ifelse(is.na(no_vaccine_1), 0, no_vaccine_1),
    no_vaccine_2 = ifelse(is.na(no_vaccine_2), 0, no_vaccine_2),
    no_vaccine_3 = ifelse(is.na(no_vaccine_3), 0, no_vaccine_3),
    no_vaccine_4 = ifelse(is.na(no_vaccine_4), 0, no_vaccine_4),
    no_vaccine_5 = ifelse(is.na(no_vaccine_5), 0, no_vaccine_5),
    no_vaccine_6 = ifelse(is.na(no_vaccine_6), 0, no_vaccine_6),
    no_vaccine_7 = ifelse(is.na(no_vaccine_7), 0, no_vaccine_7),
    no_vaccine_8 = ifelse(is.na(no_vaccine_8), 0, no_vaccine_8),
    no_vaccine_9 = ifelse(is.na(no_vaccine_9), 0, no_vaccine_9),
    no_vaccine_666 = ifelse(is.na(no_vaccine_666), 0, no_vaccine_666),
    trust_vaccine_1 = ifelse(trust_vaccine == "Family members",1,0),
    trust_vaccine_2 = ifelse(trust_vaccine == "Friends you see and talk to", 1, 0),
    trust_vaccine_2 = ifelse(trust_vaccine == "Friends you’ve made online", 1, trust_vaccine_2),
    trust_vaccine_3 = ifelse(trust_vaccine == "Religious leaders", 1, 0),
    trust_vaccine_4 = ifelse(trust_vaccine == "Famous person", 1, 0),
    trust_vaccine_5 = ifelse(trust_vaccine == "Doctors or other staff at a community health clinic", 1,0),
    trust_vaccine_6 = ifelse(trust_vaccine == "Advice of the Instituto Nacional de Salud", 1, 0),
    trust_vaccine_7 = ifelse(trust_vaccine == "Traditional healers", 1, 0),
    trust_vaccine_8 = ifelse(trust_vaccine == "Traditional media (newspaper, radio)", 1, 0),
    trust_vaccine_9 = ifelse(trust_vaccine == "Online medical discussion groups", 1, 0),
    trust_vaccine_dk = ifelse(trust_vaccine == "Don't know", 1, 0),
    trust_vaccine_refuse = ifelse(trust_vaccine == "Refuses", 1, 0),
    trust_vaccine_nr = ifelse(is.na(trust_vaccine), 1, 0),
    trust_vaccine_666 = ifelse(trust_vaccine == "Other/ Someone else", 1, 0)
  )  %>%
  dplyr::select(
    study, age, gender, educ, take_vaccine, 
    starts_with(c("yes_vaccine", "no_vaccine", "trust_vaccine_")), 
    weight = ps_weight) %>%
  dplyr::mutate(across(c(yes_vaccine_1:trust_vaccine_nr), ~replace(., is.na(.), 0))) 


ind <-
  read.dta13("1_country_data/ind/IGC_vacstudy_blfu12.dta", 
             nonint.factors = TRUE) %>%
  dplyr::as_tibble() %>% 
  dplyr::filter(!is.na(vac1)) %>% 
  dplyr::rename(
    age = HH_age,
    gender = HH_sex,
    take_vaccine = vac1) %>%
  dplyr::mutate(
    w_id = paste0(slum_code, hhid),
    study = "ind_igc",
    educ = ifelse(hhb_hh_head_educs == 1, "More than secondary", NA_character_),
    educ = ifelse(hhb_hh_head_educs == 0, "Less than secondary", educ),
    gender = as.factor(gender),
    gender = recode(gender, 
                    "1" = "Male", 
                    "0" = "Female"), 
    take_vaccine = recode(take_vaccine, 
                          "Yes, only for free" = "Yes", 
                          "Yes, even if I have to pay" = "Yes")) %>%
  dplyr::select(w_id, survey_date,
                study, age, gender, educ, take_vaccine, 
                starts_with("trust_vaccine_"), weight = pweight, cluster = slum_code) %>%
  dplyr::mutate(across(c(starts_with("trust_vaccine"), weight), ~ replace(., is.na(.), 0)))

ind_avg <- ind %>%
  dplyr::mutate(
    take_vaccine_num = ifelse(take_vaccine == "Yes", 1, 0),
    take_vaccine_num = ifelse(is.na(take_vaccine), 0, take_vaccine_num)) %>%
  group_by(w_id) %>%
  summarise(mean = mean(take_vaccine_num))

ind <- ind %>%
  mutate(date = lubridate::mdy(survey_date)) %>%
  group_by(w_id) %>%
  slice(which.max(date)) %>%
  left_join(ind_avg)

moz <-  
  read.dta13("1_country_data/moz/Survey_Data_Maputo__6_Mar_2021_18_33_33.dta", 
             nonint.factors = TRUE) %>%
  dplyr::bind_rows(read.dta13("1_country_data/moz/Survey_Data_Pemba__6_Mar_2021_18_33_35.dta", 
                              nonint.factors = TRUE)) %>%
  dplyr::as_tibble() %>% 
  dplyr::mutate(
    cluster = market_sampling,
    cluster = ifelse(is.na(cluster), ae_id_sampling, cluster)) %>%
  dplyr::rename(
    age = age_Nov2020,
    gender = ls01,
    admin_level_1 = province_current,
    take_vaccine = vc01,
    yes_vaccine_1 = vc02_A,
    yes_vaccine_2 = vc02_B,
    yes_vaccine_3 = vc02_C,
    yes_vaccine_5 = vc02_E,
    yes_vaccine_666 = vc02_F,
    no_vaccine_3 = vc04_B,
    no_vaccine_4 = vc04_C,
    no_vaccine_5 = vc04_D,
    no_vaccine_6 = vc04_E,
    no_vaccine_8 = vc04_F,
    no_vaccine_666 = vc04_G,
    vacc_gov_support = vc06,
    vacc_comm_support = vc07) %>%
  dplyr::mutate(
    yes_vaccine_666 = ifelse(vc02_D == 1, 1, yes_vaccine_666),
    vacc_gov_support = recode(vacc_gov_support, 
                              "Nao" = "No", 
                              "Sim" = "Yes", 
                              "Recusa" = NA_character_),
    vacc_gov_support = forcats::fct_relevel(vacc_gov_support, "No", "Yes"),
    vacc_comm_support = recode(vacc_comm_support, 
                               "Nao" = "No", 
                               "Sim" = "Yes", 
                               "Recusa" = NA_character_),
    vacc_comm_support = forcats::fct_relevel(vacc_comm_support, "No", "Yes"),
    gender = recode(gender, 
                    "Masculino" = "Male", 
                    "Feminino" = "Female"),
    gender = forcats::fct_relevel(gender, "Male", "Female"),
    gender = droplevels(gender),
    take_vaccine = recode(take_vaccine, 
                          "Sim" = "Yes", 
                          "Nao" = "No", 
                          "Recusa" = NA_character_),
    take_vaccine = forcats::fct_relevel(take_vaccine, "No", "Yes"),
    study = "moz_igc",
    educ_r = education_NR,
    educ_r = ifelse(is.na(educ_r), education_MKTS, educ_r),
    educ = ifelse(educ_r < 8, "Less than secondary", NA_character_),
    educ = ifelse(educ_r >= 8 & educ_r <= 12, "Secondary", educ),
    educ = ifelse(educ_r > 12 & educ_r < 18, "More than secondary", educ),
    educ = ifelse(educ_r == 18, "Less than secondary", educ)) %>%
  dplyr::select(
    study, age, gender, educ, take_vaccine, 
    starts_with(c("yes_vaccine", "no_vaccine", "trust_vaccine_")), 
    cluster, -no_vaccine_5, -no_vaccine_6, -no_vaccine_8) %>%
  mutate(across(c(yes_vaccine_1:no_vaccine_666), ~replace(., is.na(.), 0)))

nga <- 
  readr::read_csv("1_country_data/nga/Covidonlysheet.csv") %>%
  dplyr::rename(
    take_vaccine = covid_takevaccine,
    yes_vaccine_1 = "I want to protect myself from having COVID-19 in the future",
    yes_vaccine_2 = "I want to protect my family/members of my household against having COVID-19 in the future",
    yes_vaccine_3 = "I want to protect my community against having COVID-19 in the future",
    yes_vaccine_5 = "The NCDC / government recommends getting vaccines",
    yes_vaccine_666 = "Other_covidyes",
    yes_vaccine_other1 = "Life won't go back to normal until most people are vaccinated",
    yes_vaccine_other2 = "It is free",
    yes_vaccine_nr = covid_takeyes_why__99,
    no_vaccine_1 = "I would be concerned about the side effects from the vaccine",
    no_vaccine_2 = "I would be concerned about getting infected with coronavirus from the vaccine",
    no_vaccine_3 = "I'm not concerned about getting seriously ill from the virus",
    no_vaccine_4 = "I don't think vaccines work very well",
    no_vaccine_5 = "The coronavirus outbreak is not as serious as people say it is",
    no_vaccine_6 = "I don't like needles",
    no_vaccine_7 = "I'm allergic to vaccines",
    no_vaccine_8 = "I won't have time to go get vaccinated",  
    no_vaccine_9 = "The virus is a hoax / does not exist",
    no_vaccine_666 = "Other_covidno",
    no_vaccine_other1 = "Religious / community leaders advising me not to take it",
    no_vaccine_other2 = "The vaccine has microchips/ tracking devices") %>%
  dplyr::mutate(
    trust_vaccine = vaccine_trust,
    take_vaccine = recode(take_vaccine, "Yes/Agree" = "Yes", "No/Disagree" = "No"),
    yes_vaccine_1 = as_numeric(yes_vaccine_1, start.at = 0),
    yes_vaccine_2 = as_numeric(yes_vaccine_2, start.at = 0),
    yes_vaccine_3 = as_numeric(yes_vaccine_3, start.at = 0),
    yes_vaccine_5 = as_numeric(yes_vaccine_5, start.at = 0),
    yes_vaccine_666 = as_numeric(yes_vaccine_666, start.at = 0),
    yes_vaccine_other1 = as_numeric(yes_vaccine_other1, start.at = 0),
    yes_vaccine_other2 = as_numeric(yes_vaccine_other2, start.at = 0),
    yes_vaccine_666 = ifelse(yes_vaccine_other1 == 1, 1, yes_vaccine_666),
    yes_vaccine_666 = ifelse(yes_vaccine_other2 == 1, 1, yes_vaccine_666),
    vaccine_safe = ifelse(vaccines_safe == "Yes/Agree",1,0),
    vaccine_effective = ifelse(vaccines_effective == "Yes/Agree",1,0),
    no_vaccine_1 = as_numeric(no_vaccine_1, start.at = 0),
    no_vaccine_2 = as_numeric(no_vaccine_2, start.at = 0),
    no_vaccine_3 = as_numeric(no_vaccine_3, start.at = 0),
    no_vaccine_4 = as_numeric(no_vaccine_4, start.at = 0),
    no_vaccine_5 = as_numeric(no_vaccine_5, start.at = 0),
    no_vaccine_6 = as_numeric(no_vaccine_6, start.at = 0),
    no_vaccine_7 = as_numeric(no_vaccine_7, start.at = 0),
    no_vaccine_8 = as_numeric(no_vaccine_8, start.at = 0),
    no_vaccine_9 = as_numeric(no_vaccine_9, start.at = 0),
    no_vaccine_666 = as_numeric(no_vaccine_666, start.at = 0),
    no_vaccine_other1 = as_numeric(no_vaccine_other1, start.at = 0),
    no_vaccine_other2 = as_numeric(no_vaccine_other2, start.at = 0),
    no_vaccine_666 = ifelse(no_vaccine_other1, 1, no_vaccine_666),
    no_vaccine_9 = ifelse(no_vaccine_other2, 1, no_vaccine_9),
    trust_vaccine_1 = ifelse(vaccine_trust == "Family members and friends", 1, 0),
    trust_vaccine_3 = ifelse(vaccine_trust == "Religious leaders", 1, 0),
    trust_vaccine_5 = ifelse(vaccine_trust == "Medical professionals like doctors", 1, 0),
    trust_vaccine_6 = ifelse(vaccine_trust == "NCDC", 1, 0),
    trust_vaccine_6 = ifelse(vaccine_trust == "Government officials", 1, trust_vaccine_6),
    trust_vaccine_nr = ifelse(is.na(vaccine_trust), 1, 0),
    trust_vaccine_other = ifelse(vaccine_trust == "Other community leaders", 1, 0),
    trust_vaccine_666 = ifelse(vaccine_trust == "Some other source", 1, 0),
    study = "nga_scacco",
    gender = ifelse(`E_manip_small$female` == 1, "Female", "Male"),
    age = `E_manip_small$age`
  ) %>%
  dplyr::select(
    study, age, gender, take_vaccine, 
    starts_with(c("yes_vaccine", "no_vaccine", "trust_vaccine")), 
    vaccine_safe, vaccine_effective, -no_vaccine_5, -no_vaccine_6, -no_vaccine_8) %>%
  dplyr::mutate(across(c(yes_vaccine_1:trust_vaccine_666), ~replace(., is.na(.), 0))) 

npl <- 
  read.dta13("1_country_data/npl/vaccines_nepal.dta") %>%
  dplyr::as_tibble() %>% 
  dplyr::rename(
    take_vaccine = if_vaccine_take,
    weight = wt_s) %>%
  dplyr::mutate(
    gender = ifelse(resp_female == 1, "Female", "Male"),
    educ = recode(educ, 
                  "0" = "Less than secondary",
                  "1" = "Less than secondary",
                  "2" = "Less than secondary",
                  "3" = "Secondary",
                  "4" = "Secondary",
                  "5" = "Secondary",
                  "6" = "More than secondary",
                  "7" = "More than secondary"),
    across(matches(c("reason_take", "reason_not_take")), 
           ~ recode(., "Selected" = 1, "Not selected" = 0, .default = NaN))) %>%
  dplyr::rename(
    yes_vaccine_1 = reason_take_vaccine1,
    yes_vaccine_2 = reason_take_vaccine3,
    yes_vaccine_3 = reason_take_vaccine4,
    yes_vaccine_4 = reason_take_vaccine5,
    yes_vaccine_5 = reason_take_vaccine6,
    yes_vaccine_666 = reason_take_vaccine9,
    no_vaccine_1 = reason_not_take1,
    no_vaccine_2 = reason_not_take2,
    no_vaccine_3 = reason_not_take3,
    no_vaccine_4 = reason_not_take4,
    no_vaccine_5 = reason_not_take5,
    no_vaccine_6 = reason_not_take6,
    no_vaccine_7 = reason_not_take7,
    no_vaccine_8 = reason_not_take8,
    no_vaccine_9 = reason_not_take9,
    no_vaccine_666 = reason_not_take10) %>%
  dplyr::mutate(
    yes_vaccine_1 = 
      ifelse(reason_take_vaccine2 == 1, 1, yes_vaccine_1),
    yes_vaccine_666 = 
      ifelse(reason_take_vaccine7 == 1, 1, yes_vaccine_666),
    yes_vaccine_666 = 
      ifelse(reason_take_vaccine8 == 1, 1, yes_vaccine_666),
    vaccine_safe = 
      ifelse(vacc_safe == "Somewhat agree" | vacc_safe == "Strongly agree",1,0),
    vaccine_effective = 
      ifelse(vacc_effective == "Somewhat agree" | vacc_effective == "Strongly agree",1,0),
    trust_vaccine_1 = 
      ifelse(trust_vaccine == "Family members", 1, 0),
    trust_vaccine_2 = 
      ifelse(trust_vaccine == "Friends you see and talk to", 1, 0),
    trust_vaccine_3 = 
      ifelse(trust_vaccine == "Religious leaders", 1, 0),
    trust_vaccine_4 = 
      ifelse(trust_vaccine == "Famous person", 1, 0),
    trust_vaccine_5 = 
      ifelse(trust_vaccine == "Doctors or other staff at a community health clinic", 1, 0),
    trust_vaccine_6 = 
      ifelse(trust_vaccine == "Advice of the national health service", 1, 0),
    trust_vaccine_7 = 
      ifelse(trust_vaccine == "Traditional healers", 1, 0),
    trust_vaccine_8 = 
      ifelse(trust_vaccine == "Traditional media (newspaper, radio)", 1, 0),
    trust_vaccine_9 = 
      ifelse(trust_vaccine == "Online medical discussion groups", 1, 0),
    trust_vaccine_dk = 
      ifelse(trust_vaccine == "DNK", 1, 0),
    trust_vaccine_refuse = 
      ifelse(trust_vaccine == "Refused", 1, 0),
    trust_vaccine_nr = 
      ifelse(is.na(trust_vaccine), 1, 0),
    trust_vaccine_other = 
      ifelse(trust_vaccine == "Advice of the WHO", 1, 0),
    trust_vaccine_666 = 
      ifelse(trust_vaccine == "None of these/Someone else- specify", 1, 0),
    study = "npl_yale") %>%
  dplyr::select(
    study, age, gender, -educ, take_vaccine, trust_vaccine, 
    starts_with(c("yes_vaccine", "no_vaccine", "trust_vaccine_")), 
    vaccine_safe, vaccine_effective, weight, cluster = Village_Code, -no_vaccine_8) %>%
  dplyr::mutate(across(c(yes_vaccine_1:trust_vaccine_666), ~replace(., is.na(.), 0)))

pak_1 <- 
  read.dta13("1_country_data/pak_1/Sheikhupura Policing Project_EVA Tracker_Vaccine Subdata_210223.dta") %>%
  dplyr::as_tibble() %>% 
  dplyr::rename(
    educ = i_0,
    gender = gender_o,
    take_vaccine = vaccine_use,
    age = age_s,
    no_vaccine_1 = vaccine_no_q3,
    no_vaccine_2 = vaccine_no_q4,
    no_vaccine_3 = vaccine_no_q2,
    no_vaccine_4 = vaccine_no_q5,
    no_vaccine_5 = vaccine_no_q1,
    no_vaccine_6 = vaccine_no_q6,
    no_vaccine_9 = vaccine_no_q8,
    no_vaccine_666 = vaccine_no_q9) %>%
  dplyr::mutate(
    no_vaccine_1 = as_numeric(no_vaccine_1, start.at = 0),
    no_vaccine_2 = as_numeric(no_vaccine_2, start.at = 0),
    no_vaccine_3 = as_numeric(no_vaccine_3, start.at = 0),
    no_vaccine_4 = as_numeric(no_vaccine_4, start.at = 0),
    no_vaccine_5 = as_numeric(no_vaccine_5, start.at = 0),
    no_vaccine_6 = as_numeric(no_vaccine_6, start.at = 0),
    no_vaccine_9 = as_numeric(no_vaccine_9, start.at = 0),
    no_vaccine_666 = as_numeric(no_vaccine_666, start.at = 0),
    study = "pak_igc1",
    take_vaccine = recode(take_vaccine, 
                          "Don't know" = "DK",
                          "Refused to answer" = NA_character_),
    educ = recode(educ, 
                  "Refused to answer" = NA_character_,
                  "Did not go to school" = "Less than secondary",
                  "Dropped out before Matric" = "Less than secondary",
                  "Matric" = "Secondary",
                  "Intermediate/ F.A/ FS.c" = "More than secondary",
                  "Bachelors" = "More than secondary",
                  "Masters/Mphil/ M.Ed" = "More than secondary",
                  "Phd" = "More than secondary",
                  "Religious Studies" = "Less than secondary",
                  "Vocational training/Diploma/D.A.E" = "More than secondary")) %>%
  dplyr::select(
    study, educ, age, gender, take_vaccine, 
    starts_with(c("yes_vaccine", "no_vaccine", "trust_vaccine_")), 
    weight = weights, cluster = Cluster_ID) %>%
  dplyr::mutate(across(starts_with("no_vaccine_"), ~replace(., is.na(.), 0)))

pak_2 <- 
  read.dta13("1_country_data/pak_2/CERP_EVA_Data.dta") %>%
  filter(resp_province == "Punjab") %>%
  dplyr::as_tibble() %>% 
  dplyr::mutate(
    educ = recode(i_0, 
                  "Refused to answer" = NA_character_,
                  "Did not go to school" = "Less than secondary",
                  "Dropped out before Matric" = "Less than secondary",
                  "Matric" = "Secondary",
                  "Intermediate/ F.A/ FS.c" = "More than secondary",
                  "Bachelors" = "More than secondary",
                  "Masters/Mphil/ M.Ed" = "More than secondary",
                  "Phd" = "More than secondary",
                  "Religious Studies" = "Less than secondary",
                  "Vocational training/Diploma/D.A.E" = "More than secondary"),
    take_vaccine = recode(s7_10, 
                          "Absolutely yes" = "Yes",
                          "Absolutely no" = "No",
                          "Neutral" = "DK"),
    study = "pak_igc2") %>%
  dplyr::select(
    study, educ, take_vaccine, 
    starts_with(c("yes_vaccine", "no_vaccine", "trust_vaccine_")))

load("1_country_data/rus/russia_vaccine_full.rdata")
rus <- 
  russia_vaccine %>%
  dplyr::rename(
    yes_vaccine_1 = D17vaccineyes_1,
    yes_vaccine_2 = D17vaccineyes_2,
    yes_vaccine_3 = D17vaccineyes_3,
    yes_vaccine_4 = D17vaccineyes_5,
    yes_vaccine_5 = D17vaccineyes_4,
    yes_vaccine_666 = D17vaccineyes_7,
    no_vaccine_1 = D18vaccineno_1,
    no_vaccine_2 = D18vaccineno_2,
    no_vaccine_3 = D18vaccineno_3,
    no_vaccine_4 = D18vaccineno_4,
    no_vaccine_5 = D18vaccineno_8,
    no_vaccine_6 = D18vaccineno_6,
    no_vaccine_7 = D18vaccineno_7,
    no_vaccine_8 = D18vaccineno_5,
    no_vaccine_666 = D18vaccineno_16,
    weight = wgt) %>%
  dplyr::mutate(
    take_vaccine = ifelse(D16takevaccine_1 == 1 | D16takevaccine_2 == 1, 
                          "Yes",  NA_character_),
    take_vaccine = ifelse(D16takevaccine_3 == 1, "No",  take_vaccine),
    take_vaccine = ifelse(D16takevaccine_999 == 1, "DK", take_vaccine),
    yes_vaccine_666 = ifelse(D17vaccineyes_6 == 1, 1, yes_vaccine_666),
    no_vaccine_9 = 
      ifelse(D18vaccineno_11 == 1 | D18vaccineno_12 == 1 | D18vaccineno_13 == 1 | 
               D18vaccineno_14 == 1 | D18vaccineno_15 == 1, 1, 0),
    no_vaccine_666 = ifelse(D18vaccineno_10 == 1, 1, no_vaccine_666),
    trust_vaccine_1 = ifelse(D19vaccinetrust == 1, 1, 0),
    trust_vaccine_2 = ifelse(D19vaccinetrust == 2, 1, 0),
    trust_vaccine_3 = ifelse(D19vaccinetrust == 3, 1, 0),
    trust_vaccine_4 = ifelse(D19vaccinetrust == 5, 1, 0),
    trust_vaccine_5 = ifelse(D19vaccinetrust == 4, 1, 0),
    trust_vaccine_6 = ifelse(D19vaccinetrust == 6, 1, 0),
    trust_vaccine_6 = ifelse(D19vaccinetrust == 7, 1, trust_vaccine_6),
    trust_vaccine_8 = ifelse(D19vaccinetrust == 8, 1, 0),
    trust_vaccine_9 = ifelse(D19vaccinetrust == 9, 1, 0),
    trust_vaccine_nr = ifelse(is.na(D19vaccinetrust), 1, 0),
    trust_vaccine_666 = ifelse(D19vaccinetrust == 10, 1, 0),
    gender = recode(gender, 
                    "1" = "Male", "2" = "Female"),
    educ = recode(educ, 
                  "1" = "Less than secondary", 
                  "2" = "Secondary", 
                  "3" = "More than secondary", 
                  "4" = "More than secondary", 
                  "5" = "More than secondary"),
    study = "rus_syunyaev") %>%
  dplyr::select(
    study, age, gender, educ, take_vaccine, 
    starts_with(c("yes_vaccine", "no_vaccine", "trust_vaccine_")), 
    weight) %>%
  dplyr::mutate(across(c(yes_vaccine_1:trust_vaccine_666), ~replace(., is.na(.), 0)))

rwa <- 
  read.dta13("1_country_data/rwa/RWA_vaccine_data.dta",
             nonint.factors = TRUE, generate.factors = TRUE) %>%
  dplyr::as_tibble() %>% 
  dplyr::mutate(
    yes_vaccine_1 = as_numeric(yes_vaccine_1, start.at = 0),
    yes_vaccine_2 = as_numeric(yes_vaccine_2, start.at = 0),
    yes_vaccine_3 = as_numeric(yes_vaccine_3, start.at = 0),
    yes_vaccine_4 = as_numeric(yes_vaccine_4, start.at = 0),
    yes_vaccine_5 = as_numeric(yes_vaccine_5, start.at = 0),
    yes_vaccine_6 = as_numeric(yes_vaccine_6, start.at = 0),
    yes_vaccine__666 = as_numeric(yes_vaccine__666, start.at = 0),
    no_vaccine_1 = as_numeric(no_vaccine_1, start.at = 0),
    no_vaccine_2 = as_numeric(no_vaccine_2, start.at = 0),
    no_vaccine_3 = as_numeric(no_vaccine_3, start.at = 0),
    no_vaccine_4 = as_numeric(no_vaccine_4, start.at = 0),
    no_vaccine_5 = as_numeric(no_vaccine_5, start.at = 0),
    no_vaccine_6 = as_numeric(no_vaccine_6, start.at = 0),
    no_vaccine_7 = as_numeric(no_vaccine_7, start.at = 0),
    no_vaccine_8 = as_numeric(no_vaccine_8, start.at = 0),
    no_vaccine_9 = as_numeric(no_vaccine_9, start.at = 0),
    no_vaccine_10 = as_numeric(no_vaccine_10, start.at = 0),
    no_vaccine_11 = as_numeric(no_vaccine_11, start.at = 0),
    no_vaccine__666 = as_numeric(no_vaccine__666, start.at = 0),
    yes_vaccine_1 = ifelse(yes_vaccine_2 == 1, 1, yes_vaccine_1),
    no_vaccine__666 = ifelse(no_vaccine_10 == 1, 1, no_vaccine__666),
    no_vaccine__666 = ifelse(no_vaccine_11 == 1, 1, no_vaccine__666)) %>%
  dplyr::select(
    -yes_vaccine_2, -no_vaccine_10, -no_vaccine_11) %>%
  dplyr::rename(
    yes_vaccine_2 = yes_vaccine_3,
    yes_vaccine_3 = yes_vaccine_4,
    yes_vaccine_4 = yes_vaccine_5,
    yes_vaccine_5 = yes_vaccine_6,
    yes_vaccine_666 = yes_vaccine__666,
    no_vaccine_666 = no_vaccine__666) %>%
  dplyr::mutate(
    study = "rwa_recovr",
    yes_vaccine_1 = ifelse(is.na(yes_vaccine_1), 0, yes_vaccine_1),
    yes_vaccine_2 = ifelse(is.na(yes_vaccine_2), 0, yes_vaccine_2),
    yes_vaccine_3 = ifelse(is.na(yes_vaccine_3), 0, yes_vaccine_3),
    yes_vaccine_4 = ifelse(is.na(yes_vaccine_4), 0, yes_vaccine_4),
    yes_vaccine_5 = ifelse(is.na(yes_vaccine_5), 0, yes_vaccine_5),
    yes_vaccine_666 = ifelse(is.na(yes_vaccine_666), 0, yes_vaccine_666),
    no_vaccine_1 = ifelse(is.na(no_vaccine_1), 0, no_vaccine_1),
    no_vaccine_2 = ifelse(is.na(no_vaccine_2), 0, no_vaccine_2),
    no_vaccine_3 = ifelse(is.na(no_vaccine_3), 0, no_vaccine_3),
    no_vaccine_4 = ifelse(is.na(no_vaccine_4), 0, no_vaccine_4),
    no_vaccine_5 = ifelse(is.na(no_vaccine_5), 0, no_vaccine_5),
    no_vaccine_6 = ifelse(is.na(no_vaccine_6), 0, no_vaccine_6),
    no_vaccine_7 = ifelse(is.na(no_vaccine_7), 0, no_vaccine_7),
    no_vaccine_8 = ifelse(is.na(no_vaccine_8), 0, no_vaccine_8),
    no_vaccine_9 = ifelse(is.na(no_vaccine_9), 0, no_vaccine_9),
    no_vaccine_666 = ifelse(is.na(no_vaccine_666), 0, no_vaccine_666),
    trust_vaccine_1 = 
      ifelse(trust_vaccine == "Family members", 1, 0),
    trust_vaccine_2 = 
      ifelse(trust_vaccine == "Friends you see and talk to" | 
               trust_vaccine == "Friends you’ve made online", 1, 0),
    trust_vaccine_3 = 
      ifelse(trust_vaccine == "Religious leaders", 1, 0),
    trust_vaccine_4 = 
      ifelse(trust_vaccine == "Famous person", 1, 0),
    trust_vaccine_5 = 
      ifelse(trust_vaccine == "Doctors or other staff at a community health clinic", 1, 0),
    trust_vaccine_6 = 
      ifelse(trust_vaccine == "Advice of the Ministry of Health", 1, 0),
    trust_vaccine_7 = 
      ifelse(trust_vaccine == "Traditional healers", 1, 0),
    trust_vaccine_8 = 
      ifelse(trust_vaccine == "Traditional media (newspaper, radio)", 1, 0),
    trust_vaccine_9 = 
      ifelse(trust_vaccine == "Online medical discussion groups", 1, 0),
    trust_vaccine_nr = 
      ifelse(is.na(trust_vaccine), 1, 0),
    trust_vaccine_dk = 
      ifelse(trust_vaccine == "DNK", 1, 0),
    trust_vaccine_refuse = 
      ifelse(trust_vaccine == "Refused", 1, 0),
    trust_vaccine_other = 
      ifelse(trust_vaccine == "Myself", 1, 0),
    trust_vaccine_666 = 
      ifelse(trust_vaccine == "None of these/Someone else- specify", 1, 0),
    educ = 
      recode(educ,
             "1" = "Less than secondary", 
             "2" = "Secondary", 
             "3" = "More than secondary")
  ) %>%
  dplyr::select(study, age, gender, educ, take_vaccine, 
                starts_with(c("yes_vaccine", "no_vaccine", "trust_vaccine_")), 
                weight = ps_weight, -no_vaccine_8) %>%
  dplyr::mutate(across(c(yes_vaccine_1:trust_vaccine_666), ~replace(., is.na(.), 0)) )

sle_1 <- 
  readstata13::read.dta13(
    file = "1_country_data/sle_1/SLE_vaccine_data.dta",
    nonint.factors = TRUE) %>%
  dplyr::as_tibble() %>% 
  dplyr::mutate(
    yes_vaccine_1 = as_numeric(yes_vaccine_1, start.at = 0),
    yes_vaccine_2 = as_numeric(yes_vaccine_2, start.at = 0),
    yes_vaccine_3 = as_numeric(yes_vaccine_3, start.at = 0),
    yes_vaccine_4 = as_numeric(yes_vaccine_4, start.at = 0),
    yes_vaccine_5 = as_numeric(yes_vaccine_5, start.at = 0),
    yes_vaccine_6 = as_numeric(yes_vaccine_6, start.at = 0),
    yes_vaccine_7 = as_numeric(yes_vaccine_7, start.at = 0),
    yes_vaccine__666 = as_numeric(yes_vaccine__666, start.at = 0),
    no_vaccine_1 = as_numeric(no_vaccine_1, start.at = 0),
    no_vaccine_2 = as_numeric(no_vaccine_2, start.at = 0),
    no_vaccine_3 = as_numeric(no_vaccine_3, start.at = 0),
    no_vaccine_4 = as_numeric(no_vaccine_4, start.at = 0),
    no_vaccine_5 = as_numeric(no_vaccine_5, start.at = 0),
    no_vaccine_6 = as_numeric(no_vaccine_6, start.at = 0),
    no_vaccine_7 = as_numeric(no_vaccine_7, start.at = 0),
    no_vaccine_8 = as_numeric(no_vaccine_8, start.at = 0),
    no_vaccine_9 = as_numeric(no_vaccine_9, start.at = 0),
    no_vaccine__666 = as_numeric(no_vaccine__666, start.at = 0),
    yes_vaccine__666 = ifelse(yes_vaccine_7 == 1, 1, yes_vaccine__666),
    yes_vaccine_1 = ifelse(yes_vaccine_2 == 1, 1, yes_vaccine_1)) %>%
  dplyr::select(-yes_vaccine_2) %>%
  dplyr::rename(
    yes_vaccine_2 = yes_vaccine_3,
    yes_vaccine_3 = yes_vaccine_4,
    yes_vaccine_4 = yes_vaccine_5,
    yes_vaccine_5 = yes_vaccine_6,
    no_vaccine_666 = no_vaccine__666,
    yes_vaccine_666 = yes_vaccine__666) %>%
  dplyr::mutate(
    study = "sle_recovr",
    yes_vaccine_1 = ifelse(is.na(yes_vaccine_1), 0, yes_vaccine_1),
    yes_vaccine_2 = ifelse(is.na(yes_vaccine_2), 0, yes_vaccine_2),
    yes_vaccine_3 = ifelse(is.na(yes_vaccine_3), 0, yes_vaccine_3),
    yes_vaccine_4 = ifelse(is.na(yes_vaccine_4), 0, yes_vaccine_4),
    yes_vaccine_5 = ifelse(is.na(yes_vaccine_5), 0, yes_vaccine_5),
    yes_vaccine_666 = ifelse(is.na(yes_vaccine_666), 0, yes_vaccine_666),
    no_vaccine_1 = ifelse(is.na(no_vaccine_1), 0, no_vaccine_1),
    no_vaccine_2 = ifelse(is.na(no_vaccine_2), 0, no_vaccine_2),
    no_vaccine_3 = ifelse(is.na(no_vaccine_3), 0, no_vaccine_3),
    no_vaccine_4 = ifelse(is.na(no_vaccine_4), 0, no_vaccine_4),
    no_vaccine_5 = ifelse(is.na(no_vaccine_5), 0, no_vaccine_5),
    no_vaccine_6 = ifelse(is.na(no_vaccine_6), 0, no_vaccine_6),
    no_vaccine_7 = ifelse(is.na(no_vaccine_7), 0, no_vaccine_7),
    no_vaccine_8 = ifelse(is.na(no_vaccine_8), 0, no_vaccine_8),
    no_vaccine_9 = ifelse(is.na(no_vaccine_9), 0, no_vaccine_9),
    no_vaccine_666 = ifelse(is.na(no_vaccine_666), 0, no_vaccine_666),
    trust_vaccine_1 = 
      ifelse(trust_vaccine == "Family members", 1, 0),
    trust_vaccine_2 = 
      ifelse(trust_vaccine == "Friends you see and talk to" |
               trust_vaccine == "Friends you’ve made online", 1, 0),
    trust_vaccine_3 = 
      ifelse(trust_vaccine == "Religious leaders", 1, 0),
    trust_vaccine_4 = 
      ifelse(trust_vaccine == "Famous person", 1, 0),
    trust_vaccine_5 = 
      ifelse(trust_vaccine == "Doctors or other staff at a community health clinic", 1, 0),
    trust_vaccine_6 = 
      ifelse(trust_vaccine == "Advice of the Ministry of Health and Sanitation", 1, 0),
    trust_vaccine_7 = 
      ifelse(trust_vaccine == "Traditional healers", 1, 0),
    trust_vaccine_8 = 
      ifelse(trust_vaccine == "Traditional media (newspaper, radio)", 1,0),
    trust_vaccine_9 = 
      ifelse(trust_vaccine == "Online medical discussion groups", 1, 0),
    trust_vaccine_other = 
      ifelse(trust_vaccine == "I do trust NOBODY", 1, 0),
    trust_vaccine_nr = 
      ifelse(is.na(trust_vaccine), 1, 0),
    trust_vaccine_dk = 
      ifelse(trust_vaccine == "DNK", 1, 0),
    trust_vaccine_refuse = 
      ifelse(trust_vaccine == "Refused", 1, 0),
    trust_vaccine_666 = 
      ifelse(trust_vaccine == "None of these/Someone else- specify", 1, 0),
    educ = recode(educ, 
                  "1" = "Less than secondary", 
                  "2" = "Secondary", 
                  "3" = "More than secondary"),
    gender = recode(gender, 
                    "Men" = "Male", 
                    "Women" = "Female"))  %>%
  dplyr::select(study, age, gender, educ, take_vaccine, 
                starts_with(c("yes_vaccine", "no_vaccine", "trust_vaccine_")), 
                weight = ps_weight) %>%
  dplyr::mutate(across(c(yes_vaccine_1:trust_vaccine_666), ~replace(., is.na(.), 0)))


sle_2 <-  
  readstata13::read.dta13(
    file = "1_country_data/sle_2/vaccine_subset.dta",
    nonint.factors = TRUE, generate.factors = TRUE) %>% 
  dplyr::as_tibble() %>% 
  dplyr::rename(
    trust_vaccine = vaccine_trust,
    yes_vaccine_1 = vacc_prtoectme,
    yes_vaccine_2 = vacc_protectfam,
    yes_vaccine_3 = vacc_protectcomm,
    yes_vaccine_4 = vacc_chc,
    yes_vaccine_666 = vacc_oth,
    no_vaccine_1 = vaccno_sideef,
    no_vaccine_3 = vaccno_risk,
    no_vaccine_8 = vaccno_notime,
    no_vaccine_4 = vaccno_effective,
    no_vaccine_6 = vaccno_needle,
    no_vaccine_9 = vaccno_noexist,
    no_vaccine_666 = vaccno_other) %>%
  dplyr::mutate(
    yes_vaccine_666 = ifelse(vacc_normal == 1, 1, yes_vaccine_666),
    no_vaccine_666 = 
      ifelse(vaccno_relig_obj == 1  | vaccno_nosympt == 1 | 
               vaccno_immune == 1 | vaccno_foreign == 1 | 
               vaccno_dontknow == 1, 1, no_vaccine_666),
    educ = recode(educ, 
                  "Pre-primary or nursery" = "Less than secondary",
                  "Primary school" = "Less than secondary",
                  "Bachelors" = "More than secondary",
                  "Secondary School" = "Secondary",
                  "Further Education" = "More than secondary",
                  "No formal education (arabic)" = "Less than secondary",
                  "Non-other" = "Less than secondary"),
    study = "sle_igc",
    age = ifelse(age == -99 | age == -98, NA, age),
    yes_vaccine_1 = ifelse(take_vaccine == "No", NA, yes_vaccine_1),
    yes_vaccine_2 = ifelse(take_vaccine == "No", NA, yes_vaccine_2),
    yes_vaccine_3 = ifelse(take_vaccine == "No", NA, yes_vaccine_3),
    yes_vaccine_4 = ifelse(take_vaccine == "No", NA, yes_vaccine_4),
    yes_vaccine_666 = ifelse(take_vaccine == "No", NA, yes_vaccine_666),
    no_vaccine_1 = ifelse(take_vaccine == "Yes", NA, no_vaccine_1),
    no_vaccine_3 = ifelse(take_vaccine == "Yes", NA, no_vaccine_3),
    no_vaccine_4 = ifelse(take_vaccine == "Yes", NA, no_vaccine_4),
    no_vaccine_6 = ifelse(take_vaccine == "Yes", NA, no_vaccine_6),
    no_vaccine_8 = ifelse(take_vaccine == "Yes", NA, no_vaccine_8),
    no_vaccine_9 = ifelse(take_vaccine == "Yes", NA, no_vaccine_9),
    no_vaccine_666 = ifelse(take_vaccine == "Yes", NA, no_vaccine_666),
    yes_vaccine_1 = ifelse(is.na(yes_vaccine_1), 0, yes_vaccine_1),
    yes_vaccine_2 = ifelse(is.na(yes_vaccine_2), 0, yes_vaccine_2),
    yes_vaccine_3 = ifelse(is.na(yes_vaccine_3), 0, yes_vaccine_3),
    yes_vaccine_4 = ifelse(is.na(yes_vaccine_4), 0, yes_vaccine_4),
    yes_vaccine_666 = ifelse(is.na(yes_vaccine_666), 0, yes_vaccine_666),
    no_vaccine_1 = ifelse(is.na(no_vaccine_1), 0, no_vaccine_1),
    no_vaccine_3 = ifelse(is.na(no_vaccine_3), 0, no_vaccine_3),
    no_vaccine_4 = ifelse(is.na(no_vaccine_4), 0, no_vaccine_4),
    no_vaccine_6 = ifelse(is.na(no_vaccine_6), 0, no_vaccine_6),
    no_vaccine_8 = ifelse(is.na(no_vaccine_8), 0, no_vaccine_8),
    no_vaccine_9 = ifelse(is.na(no_vaccine_9), 0, no_vaccine_9),
    no_vaccine_666 = ifelse(is.na(no_vaccine_666), 0, no_vaccine_666), 
    vaccine_safe = 
      ifelse(vaccine_safe == "Somewhat agree" | 
               vaccine_safe == "Strongly agree", 1, 0),
    vaccine_effective = 
      ifelse(vaccine_effective == "Somewhat agree" | 
               vaccine_effective == "Strongly agree", 1, 0),
    trust_vaccine_1 = 
      ifelse(trust_vaccine == "Family", 1, 0),
    trust_vaccine_2 = 
      ifelse(trust_vaccine == "Friends you see and talk to " |
               trust_vaccine == "Friends you've made online", 1, 0),
    trust_vaccine_3 = 
      ifelse(trust_vaccine == "A religious leader", 1, 0),
    trust_vaccine_4 = 
      ifelse(trust_vaccine == "A famous person (specify)", 1, 0),
    trust_vaccine_5 = ifelse(
      trust_vaccine == "A doctor, nurse or other staff at a community health clinic" |
        trust_vaccine == "A country medical staff", 1, 0),
    trust_vaccine_7 = 
      ifelse(trust_vaccine == "A traditional healer", 1, 0),
    trust_vaccine_9 = 
      ifelse(trust_vaccine == "Online medical discussion groups", 1, 0),
    trust_vaccine_nr = 
      ifelse(is.na(trust_vaccine), 1, 0),
    trust_vaccine_666 = 
      ifelse(trust_vaccine == "None of these - someone else, specify", 1, 0)) %>%
  dplyr::select(
    study, age, gender, educ, take_vaccine, 
    starts_with(c("yes_vaccine", "no_vaccine", "trust_vaccine_")), 
    vaccine_safe, vaccine_effective, cluster = communityid, -no_vaccine_3, no_vaccine_8) %>%
  dplyr::mutate(across(c(yes_vaccine_1:trust_vaccine_666), ~replace(., is.na(.), 0)))

uga_1 <-  
  readstata13::read.dta13(
    file = "1_country_data/uga/IGC_Covid19vaccine_questions_final_clusterID.dta",
    nonint.factors = FALSE, generate.factors = TRUE) %>%
  dplyr::as_tibble() %>% 
  dplyr::rename(
    take_vaccine = vacc_yn,
    age = current_age) %>%
  dplyr::mutate(
    study = "uga_igc",
    no_vaccine_1 = ifelse(vacc_why_no_1 == 1, 1, 0),
    no_vaccine_3 = ifelse(vacc_why_no_2 == 1, 1, 0),
    no_vaccine_4 = ifelse(vacc_why_no_3 == 1, 1, 0),
    no_vaccine_5 = ifelse(vacc_why_no_4 == 1, 1, 0),
    no_vaccine_6 = ifelse(vacc_why_no_5 == 1, 1, 0),
    no_vaccine_8 = ifelse(vacc_why_no_6 == 1, 1, 0),
    no_vaccine_666 = ifelse(vacc_why_no_7 == 1 | vacc_why_no__66 == 1, 1, 0),
    yes_vaccine_1 = ifelse(vacc_why_yes_1 == 1, 1, 0),
    yes_vaccine_2 = ifelse(vacc_why_yes_2 == 1, 1, 0),
    yes_vaccine_3 = ifelse(vacc_why_yes_3 == 1, 1, 0),
    yes_vaccine_5 = ifelse(vacc_why_yes_5 == 1, 1, 0),
    yes_vaccine_666 = ifelse(vacc_why_yes_4 == 1, 1, 0),
    yes_vaccine_666 = ifelse(vacc_why_yes__66 == 1, 1, yes_vaccine_666),
    educ = ifelse(respondent_educ <= 8, "Less than secondary", NA_character_),
    educ = ifelse(respondent_educ <= 14 & respondent_educ > 8, "Secondary", educ),
    educ = ifelse(respondent_educ > 14, "More than secondary", educ),
    gender = as.factor("Female")) %>%
  dplyr::select(
    study, age, gender, educ, take_vaccine, 
    starts_with(c("yes_vaccine", "no_vaccine", "trust_vaccine_")), 
    cluster = clusterID) %>%
  dplyr::mutate(across(c(yes_vaccine_1:no_vaccine_666), ~replace(., is.na(.), 0)))

uga <- 
  read.dta13("1_country_data/uga_1/20210125_baseline_covid.1.2.3_merged_final.dta") %>% 
  dplyr::as_tibble() %>% 
  tidyr::separate(resp_id, c("first","second"), sep = "-|\\s", remove = FALSE) %>%
  # Manual correction for age, and manually plugging in missing values
  # Manual plugging in missing values for education
  dplyr::mutate(
    cluster = as.numeric(substring(first, 1, 3)),
    bsln_age = bsln_age + 1,
    bsln_age = ifelse(is.na(bsln_age), w1_age, bsln_age),
    bsln_education = 
      ifelse(is.na(bsln_education), w1_education, bsln_education)) %>%
  dplyr::filter(is.na(second) & cluster < 960) %>%
  dplyr::rename(
    take_vaccine = w3_take_vaccine,
    gender = bsln_gender,
    age = bsln_age,
    educ = bsln_education,
    yes_vaccine_1 = w3_whytake_vaccine_1,
    yes_vaccine_2 = w3_whytake_vaccine_3,
    yes_vaccine_3 = w3_whytake_vaccine_4,
    yes_vaccine_4 = w3_whytake_vaccine_5,
    yes_vaccine_5 = w3_whytake_vaccine_6,
    yes_vaccine_666 = w3_whytake_vaccine_888,
    no_vaccine_1 = w3_nottake_vaccine_1,
    no_vaccine_2 = w3_nottake_vaccine_2,
    no_vaccine_3 = w3_nottake_vaccine_3,
    no_vaccine_4 = w3_nottake_vaccine_4,
    no_vaccine_5 = w3_nottake_vaccine_5,
    no_vaccine_6 = w3_nottake_vaccine_6,
    no_vaccine_7 = w3_nottake_vaccine_7,
    no_vaccine_8 = w3_nottake_vaccine_8,
    no_vaccine_9 = w3_nottake_vaccine_9,
    no_vaccine_666 = w3_nottake_vaccine_888,
    trust_vaccine = w3_person_trust) %>%
  dplyr::filter(!is.na(take_vaccine)) %>%
  dplyr::mutate(
    study = "uga_wzb",
    take_vaccine = recode(take_vaccine, 
                          "Refused" = NA_character_, 
                          "Don't know" = "DK"),
    gender = recode(gender, 
                    "1" = "Male", 
                    "2" = "Female"),
    educ = case_when(educ %in% c(1,2,3,4,5,6,7,15,18) ~ "Less than secondary",
                     educ %in% c(8,9,10,11,12,13) ~ "Secondary",
                     educ %in% c(14,16,17) ~ "More than secondary"),
    yes_vaccine_1 = ifelse(w3_whytake_vaccine_2 == 1, 1, yes_vaccine_1),
    yes_vaccine_666 = 
      ifelse(w3_whytake_vaccine_7 == 1, 1, yes_vaccine_666),
    trust_vaccine_1 = 
      ifelse(trust_vaccine == "Family members", 1, 0),
    trust_vaccine_2 = 
      ifelse(trust_vaccine == "Friends you see and talk to" |
               trust_vaccine == "Friends youâ€™ve made online", 1, 0),
    trust_vaccine_3 = 
      ifelse(trust_vaccine == "Religious leaders", 1, 0),
    trust_vaccine_4 = 
      ifelse(trust_vaccine == "Famous person", 1, 0),
    trust_vaccine_5 = 
      ifelse(trust_vaccine == "Doctors or other staff at a community health clinic", 1, 0),
    trust_vaccine_6 = 
      ifelse(trust_vaccine == "Advice of the national health service", 1, 0),
    trust_vaccine_7 = 
      ifelse(trust_vaccine == "Traditional healers", 1, 0),
    trust_vaccine_8 = 
      ifelse(trust_vaccine == "Traditional media (newspaper, radio)", 1, 0),
    trust_vaccine_9 = 
      ifelse(trust_vaccine == "Online medical discussion groups", 1, 0),
    trust_vaccine_refuse = 
      ifelse(trust_vaccine == "Refused", 1, 0),
    trust_vaccine_dk = 
      ifelse(trust_vaccine == "Don't know", 1, 0),
    trust_vaccine_other = 
      ifelse(trust_vaccine == "None of these", 1, 0),
    trust_vaccine_nr = 
      ifelse(is.na(trust_vaccine), 1, 0),
    trust_vaccine_666 = 
      ifelse(trust_vaccine == "Someone else:_______________", 1, 0)) %>%
  dplyr::select(study, age, gender, educ, take_vaccine, 
                starts_with(c("yes_vaccine", "no_vaccine", "trust_vaccine_")), 
                cluster, -no_vaccine_8) %>%
  dplyr::mutate(across(c(yes_vaccine_1:no_vaccine_666), ~replace(., is.na(.), 0)))

load("1_country_data/usa/usa_covid_survey.RData")
usa <- 
  d %>%
  dplyr::as_tibble() %>% 
  dplyr::select(
    gender, age, starts_with("WZB"), weight = weights, edu) %>%
  dplyr::mutate(
    yes_vaccine_1 = as.numeric(WZB_yes_vaccine_1),
    yes_vaccine_2 = as.numeric(WZB_yes_vaccine_2),
    yes_vaccine_3 = as.numeric(WZB_yes_vaccine_3),
    yes_vaccine_5 = as.numeric(WZB_yes_vaccine_4),
    take_vaccine = ifelse(WZB_take_vaccine == 2 | WZB_take_vaccine == 1, "Yes", WZB_take_vaccine),
    take_vaccine = ifelse(take_vaccine == -1 | take_vaccine == -2, "No", take_vaccine),
    take_vaccine = ifelse(take_vaccine == -99, NA_character_, take_vaccine),
    no_vaccine_1 = ifelse(WZB_no_vaccine_1 == 1, 1, 0),
    no_vaccine_3 = ifelse(WZB_no_vaccine_2 == 1, 1, 0),
    no_vaccine_4 = ifelse(WZB_no_vaccine_3 == 1, 1, 0),
    no_vaccine_9 = ifelse(WZB_no_vaccine_conspiracy == 1, 1, 0),
    no_vaccine_666 = 
      ifelse(WZB_no_vaccine_4 == 1 | 
               WZB_no_vaccine_text_code %in% c("2", "5", "11", "3, 11", "888"), 1, 0),
    trust_vaccine_1 = ifelse(WZB_trust_vaccine == 1, 1, 0),
    trust_vaccine_3 = ifelse(WZB_trust_vaccine == 8, 1, 0),
    trust_vaccine_5 = ifelse(WZB_trust_vaccine == 2 , 1, 0),
    trust_vaccine_6 = ifelse(WZB_trust_vaccine == 3 | WZB_trust_vaccine == 4 | 
                               WZB_trust_vaccine == 6 | WZB_trust_vaccine == 7, 1, 0),
    trust_vaccine_nr = ifelse(WZB_trust_vaccine == -99, 1, 0),
    trust_vaccine_other = ifelse(WZB_trust_vaccine == 5, 1, 0),
    trust_vaccine_666 = ifelse(WZB_trust_vaccine == 9, 1, 0),
    gender = recode(gender, 
                    "1" = "Male", 
                    "2" = "Female"),
    educ = ifelse(edu > 2 | edu == -99, "More than secondary", "Secondary"),
    age = as.integer(age),
    study = "usa_wzb") %>%
  dplyr::select(
    study, age, gender, educ, take_vaccine, 
    starts_with(c("yes_vaccine", "no_vaccine", "trust_vaccine_")), 
    weight) %>%
  dplyr::mutate(across(c(yes_vaccine_1:trust_vaccine_666), ~replace(., is.na(.), 0)))


dfs <- list(col, bfa, ind, moz, pak_1, pak_2, sle_2, uga, uga_1, rwa, sle_1, nga, usa, npl, rus)

```

```{r data, echo = FALSE, warning = FALSE}

## Puts cleaned dfs into a single df. All NAs are considered NA (recoded are explicit "Refuse" and "Don't know")

vars <- readxl::read_excel("2_input_data/dictionary.xlsx") %>% pull(outcome)

df <- 
  dfs %>% 
  plyr::llply(
    .data = .,
    .fun = function(df) {
      if (!is.null(df$cluster)) df %<>% mutate(cluster = as.character(cluster))
      return(df)
    }
  ) %>% 
  dplyr::bind_rows() %>% 
  dplyr::left_join(dplyr::select(studies, study, country), by = "study") %>%
  dplyr::mutate(
    take_vaccine_num = ifelse(take_vaccine == "Yes", 1, 0),
    take_vaccine_num = ifelse(is.na(take_vaccine), 0, take_vaccine_num),
    take_vaccine_num = ifelse(study == "ind_igc", mean, take_vaccine_num),
    cluster = 
      ifelse(is.na(cluster), NA, 
             paste0(gsub(pattern = " ", replacement = "_", x = tolower(country)), "_", cluster))) %>%
  dplyr::select(any_of(vars))

readr::write_csv(df, "3_rep_data/combined.csv")
```

```{r wgmdata, echo = FALSE, warning = FALSE}

## Generates data from WGM

wgmdata <- readr::read_csv("2_input_data/wgm_2018_publiccsv.csv")

# recode the name of countries
ctries <- list("1" = "USA", "9" = "Pakistan",
               "31" = "India", "35" = "Nigeria",
               "41" = "Uganda", "63" = "Mozambique",
               "65" = "Rwanda", "76" = "Russia",
               "78" = "Burkina Faso", "80" = "Sierra Leone",
               "105" = "Colombia", "157" = "Nepal")

wgmdata %<>%
  # keep only these countries we have data for
  dplyr::filter(WP5 %in% c(1,9,31,35,41,63,65,76,78,80,105,157)) %>% 
  dplyr::mutate(
    country = purrr::map_chr(WP5, ~ ctries[[as.character(.x)]]),
    # turn the variables into binary versions
    importance_binary = 2 - cut(Q24,  breaks = c(-Inf, 2, 6), labels = FALSE),
    safe_binary       = 2 - cut(Q25,  breaks = c(-Inf, 2, 6), labels = FALSE),
    effective_binary  = 2 - cut(Q26,  breaks = c(-Inf, 2, 6), labels = FALSE),
    cov_binary        = 2 - cut(Q28,  breaks = c(-Inf, 1, 3), labels = FALSE),
    t_neigh           = 2 - cut(Q11A, breaks = c(-Inf, 2, 4), labels = FALSE),
    t_gov             = 2 - cut(Q11B, breaks = c(-Inf, 2, 4), labels = FALSE),
    t_scient          = 2 - cut(Q11C, breaks = c(-Inf, 2, 4), labels = FALSE),
    t_jour            = 2 - cut(Q11D, breaks = c(-Inf, 2, 4), labels = FALSE),
    t_doc             = 2 - cut(Q11E, breaks = c(-Inf, 2, 4), labels = FALSE),
    t_ngo             = 2 - cut(Q11F, breaks = c(-Inf, 2, 4), labels = FALSE),
    t_theal           = 2 - cut(Q11G, breaks = c(-Inf, 2, 4), labels = FALSE),
    
  )

#compute rates
wgm_results <- 
  wgmdata %>%
  dplyr::nest_by(country) %>%
  dplyr::mutate(
    modelimp = list(lm_robust(importance_binary ~ 1, 
                              weight = wgt,
                              data = data, se_type = "stata")),
    modelsaf = list(lm_robust(safe_binary ~ 1, 
                              weight = wgt,
                              data = data, se_type = "stata")),
    modeleff = list(lm_robust(effective_binary ~ 1, 
                              weight = wgt,
                              data = data, se_type = "stata")),
    modelcov = list(lm_robust(cov_binary ~ 1, 
                              weight = wgt,
                              data = data, se_type = "stata")),
    modelneigh = list(lm_robust(t_neigh ~ 1, 
                                weight = wgt,
                                data = data, se_type = "stata")),
    modelgov = list(lm_robust(t_gov ~ 1, 
                              weight = wgt,
                              data = data, se_type = "stata")),
    modelscient = list(lm_robust(t_scient ~ 1, 
                                 weight = wgt,
                                 data = data, se_type = "stata")),
    modeljour = list(lm_robust(t_jour ~ 1, 
                               weight = wgt,
                               data = data, se_type = "stata")),
    modeldoc = list(lm_robust(t_doc ~ 1, 
                              weight = wgt,
                              data = data, se_type = "stata")),
    modelngo = list(lm_robust(t_ngo ~ 1, 
                              weight = wgt,
                              data = data, se_type = "stata")),
    modeltheal = list(lm_robust(t_theal ~ 1, 
                                weight = wgt,
                                data = data, se_type = "stata")))

#complete table
table_wgm <- 
  wgm_results %>%
  summarize(tidy(modelimp)) %>%
  mutate(estimate = 
           paste0(round(estimate*100, 0))
  ) %>%
  select(estimate) %>% 
  rename(Important = estimate)

table_wgm <- 
  wgm_results %>%
  summarize(tidy(modelsaf)) %>%
  mutate(estimate = 
           paste0(round(estimate*100, 0))
  ) %>%
  select(estimate) %>% 
  rename(Safety = estimate) %>%
  left_join(table_wgm)

table_wgm <- 
  wgm_results %>%
  summarize(tidy(modeleff)) %>%
  mutate(estimate = 
           paste0(round(estimate*100, 0))
  ) %>%
  select(estimate) %>% 
  rename(Effectiveness = estimate) %>%
  left_join(table_wgm)

table_wgm <- 
  wgm_results %>%
  summarize(tidy(modelcov)) %>%
  mutate(estimate = 
           paste0(round(estimate*100, 0))
  ) %>%
  select(estimate) %>% 
  rename(Coverage = estimate) %>%
  left_join(table_wgm)


readr::write_csv(table_wgm, "3_rep_data/table_wgm.csv")

```


```{r covdata, echo = FALSE, warning = FALSE}

##Generates table of WHO coverage estimates

BCG  <- 
  readxl::read_xls("2_input_data/WHO_coverage_estimates_series.xls",sheet = "BCG") %>%
  select(Cname,`2019`) %>% rename(BCG = "2019")

MCV1 <- 
  readxl::read_xls("2_input_data/WHO_coverage_estimates_series.xls",sheet = "MCV1") %>%
  select(Cname,`2019`) %>% rename(MCV1 = "2019")

DTP1 <- 
  readxl::read_xls("2_input_data/WHO_coverage_estimates_series.xls",sheet = "DTP1") %>%
  select(Cname,`2019`) %>% rename(DTP1 = "2019")

vacc_cov <- 
  full_join(BCG, DTP1, by = 'Cname') %>%
  full_join(., MCV1, by = 'Cname') 


nmctry <- 
  c("United States of America",
    "Pakistan",
    "India",
    "Nigeria",
    "Uganda",
    "Mozambique",
    "Rwanda",
    "Russian Federation",
    "Burkina Faso",
    "Sierra Leone",
    "Colombia",
    "Nepal")

#keep only these countries
vacc_cov <- 
  vacc_cov %>% 
  dplyr::filter(Cname %in% nmctry) %>% 
  dplyr::rename(country = Cname)

vacc_cov$country <- 
  replace(vacc_cov$country, vacc_cov$country == "United States of America", "USA")
vacc_cov$country <- 
  replace(vacc_cov$country, vacc_cov$country == "Russian Federation", "Russia")

readr::write_csv(vacc_cov, "3_rep_data/vacc_cov.csv")

```
