==> 2_tables.Rmd <==
---
title: 'COVID-19 Vaccine Acceptance and Hesitancy in Low and Middle Income Countries: Replication Code'
date: "1 July 2021"
output:
  html_document:
    number_sections: yes
    theme: cosmo
    code_folding: hide
    toc: yes
    toc_float: yes
    self_contained: yes
  bookdown::html_document2: default
keywords: vaccine hesitancy
---

```{r setup3, include=FALSE}

knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)

if (!require(pacman)) install.packages("pacman")

pacman::p_load(tidyverse, magrittr, knitr, kableExtra, DT, lazyeval, 
               labelled, forcats, readxl, googledrive, estimatr, ggforce, fastDummies,
               stringr, RColorBrewer, readstata13, metaplus, sjlabelled, ggrepel, tikzDevice)

options(knitr.kable.NA = '.')

base_font_size <- 12

```

This replication file and ancillary data and code are hosted at https://github.com/wzb-ipi/covid_vaccines_nmed

# Set up

This code takes cleaned and aggregated data as input. Cleaning and aggregation is done in `1_cleaning.Rmd`.

Aggregated data can be downloaded directly from: https://wzb-ipi.github.io/covid_vaccines_nmed/combined.csv

## Helper functions

Code used to prepare data (prepare survey weights), run analyses and robustness checks.

```{r}

# The helper renormalizes weights so that each study gets the 
# same total weight even if they are missing data

study_weighting <- function(data)
  data %>% 
    dplyr::group_by(country) %>% 
    dplyr::mutate(weight = weight/sum(weight)) %>% 
    dplyr::ungroup() 

lm_helper <- function(data,
                      ...) {
  data %>% 
    study_weighting() %>% 
    estimatr::lm_robust(data = .,...) %>% 
    {bind_cols( tidy(.), n = nobs(.) )}
}

# Leave-X-out helper that takes data and sample_var, 
# nests by sample_var performs LOO with loo_n observations out
# applying loo_fun to each sample

loo_helper <- function(data, 
                       sample_var, 
                       loo_n = 1,
                       loo_fun = 
                         function(dat) 
                           lm_helper(data = dat, 
                                     formula = take_vaccine_num ~ 1, 
                                     cluster = cluster,
                                     weight = weight, 
                                     se_type = "stata")) {
  
  .var <- data[[sample_var]]
  
  # Data is recombined and loo_fun applied to each combination 
  data %>% 
    {
      plyr::adply(.data = combn(unique(.var), loo_n), 
                  .margins = 2, 
                  .fun = function(x) loo_fun(.[!(.var %in% x), ]) )
    }
}

# # Illustration of loo_helper
# loo_helper(
#   data = data.frame(country = 1:3, Y = 1:3),
#   loo_n = 2,
#   sample_var = "country",
#   loo_fun = function(dat) mean(dat$Y)
# )

# Subgroup analysis : Function to apply analysis function over groups

grp_analysis <- function(df, 
                         y, 
                         x) {
  df %>%
    dplyr::filter(if_all(c(all_of(x), all_of(y), cluster, weight), ~ !is.na(.))) %>%
    dplyr::nest_by(group, get(x)) %>%
    dplyr::summarize(
      lm_helper(data = data, 
                formula = as.formula(paste0(y, "~ 1")), cluster = cluster,
                weight = weight, se_type = "stata"), .groups = "drop") %>% 
    dplyr::rename(!!x := "get(x)")
}

# Reasons analysis: Function to apply analysis function over groups

reasons_together <- function(df, 
                             reason, 
                             num = "Yes") {
  df %>%
    dplyr::filter(take_vaccine %in% num, 
                  if_all(c(all_of(reason), cluster, weight), ~ !is.na(.))) %>%
    dplyr::nest_by(group) %>%
    dplyr::summarize(
      lm_helper(data = data, 
                formula = as.formula(paste0(reason, "~ 1")), 
                cluster = cluster,
                weight = weight, se_type = "stata"), .groups = "drop")
}


reasons_together_subgroup <- function(df, 
                                      reason, num = "Yes", 
                                      dem_group = NA, 
                                      dem_subgroup = NA) {
  
  if (dem_group == "gender")
    df <- filter(df, gender %in% dem_subgroup)
  
  df %>%
    dplyr::filter(take_vaccine %in% num,
                  !is.na(get(reason))) %>%
    dplyr::nest_by(group) %>%
    dplyr::summarize(
      lm_helper(data = data, 
                formula = as.formula(paste0(reason, "~ 1")), cluster = cluster,
                weight = weight, se_type = "stata"), .groups = "drop")
}

# Age analysis for reasons

age_analysis <- function(df, 
                         reason, 
                         num = "Yes", 
                         filter_by = NA) {
  df %>%
    dplyr::filter({{filter_by}} == 1)  %>%
    dplyr::filter(take_vaccine %in% num, 
                  if_all(c(all_of(reason), cluster, weight), ~ !is.na(.))) %>%
    dplyr::nest_by(group) %>%
    dplyr::summarize(
      lm_helper(data = data, 
                formula = as.formula(paste0(reason, "~ 1")), 
                cluster = cluster,
                weight = weight, se_type = "stata"), .groups = "drop")
}

```

## Data preparation

Load data:

```{r df2}

# Call data created in 1_cleaning.Rmd
df <- readr::read_csv("3_rep_data/combined.csv", guess_max = 30000)

```

Groups variables into discrete categories and prepares survey weights.

```{r}
# If no cluster information given for a study then individuals are clusters 
# Ensure cluster ids are distinct across studies
df <- 
  df %>% 
  dplyr::group_by(study) %>% 
  dplyr::mutate(
    cluster = ifelse(is.na(cluster), paste(1:n()), cluster),
    cluster = paste0(gsub(" ", "_", tolower(country)), "_", cluster))


# Weights sum to 1 in each study and recode age and education into bins
df <- 
  df %>% 
  dplyr::group_by(study) %>% 
  dplyr::mutate(
         weight_replace = mean(weight, rm.na = TRUE),
         weight = if_else(is.na(weight), 
                          if_else(is.na(weight_replace), 1, weight_replace), 
                          weight),
         weight = weight/sum(weight)) %>% 
  dplyr::ungroup() %>%
  dplyr::mutate(
    age_groups = 
      as.character(cut(x = age, breaks = c(-Inf, 18, 30, 45, 60, +Inf), right = F)),
    age_groups_binary = ifelse(age >= 55, "55+", NA),
    age_groups_binary = ifelse(age < 55, "<55", age_groups_binary),
    age_less24 = ifelse(age <= 24, 1, 0),
    age_25_54 = ifelse(age >= 25 & age <= 54, 1, 0),
    age_55_more = ifelse(age >= 55, 1, 0),
    age_groups_three = ifelse(age <= 24, "<25", NA),
    age_groups_three = ifelse(age >= 25 & age <= 54, "25-54", age_groups_three),
    age_groups_three = ifelse(age >= 55, "55+", age_groups_three),
    educ_binary = if_else(educ == "More than secondary", "> Secondary", "Up to Secondary")) 


# We create a new dataframe with countries and with "All" (only LMICs). Countries are clusters in "All" analysis
# USA and Russia excluded from "All" set

df2 <- 
  dplyr::bind_rows(
    mutate(df, group = country),
    mutate(filter(df, country != "USA" & country != "Russia"), group = "All")) %>% 
  mutate(
    cluster = if_else(group == "All", 
                      gsub(pattern = " ", replacement = "_", x = tolower(country)), 
                      cluster)) 
```

## Data checks

Checks on data structure. Note `n`, missingness, presence of data on weights or clusters. 

### Data structure

```{r}

df %>% group_by(country) %>%
  summarize(n = n(), 
            sd_wt = sd(weight, na.rm = TRUE)/mean(weight, na.rm = TRUE), 
            cl_size = n()/length(unique(cluster)), 
            take_1 = mean((take_vaccine == "Yes")[take_vaccine == "Yes" | take_vaccine == "No" ], na.rm = TRUE),
            take_2 = mean((take_vaccine == "Yes"), na.rm = TRUE),
            take_3 = mean(take_vaccine_num, na.rm = TRUE),
            take_dk = mean(take_vaccine == "DK"),
            .take = mean(is.na(take_vaccine_num)), .age = mean(is.na(age)),
            .gender = mean(is.na(gender)), .educ = mean(is.na(educ))) %>%
  kable(digits = 2, 
        caption = "Observations, missingness patterns, data structure. Column .var is the share missing for variable var.", booktabs = TRUE, linesep = "", format.args = list(big.mark = ",", 
  scientific = FALSE))

```

```{r, include=FALSE}

nas <- df %>% group_by(country) %>%
  summarize(n = n(),
            take_vaccine_na = mean(!is.na(take_vaccine))*100,
            gender_na = mean(!is.na(gender))*100,
            educ_na = mean(!is.na(educ))*100,
            age_na = mean(!is.na(age))*100) %>%
  kable(digits = 2, 
        col.names = c("Country", "N obs", "Take vaccine", "Gender", "Education", "Age"),
        caption = "Observations and missingness patterns", 
        booktabs = TRUE, linesep = "", 
        format.args = list(big.mark = ",", scientific = FALSE),
  format = "latex") %>% 
  kableExtra::kable_styling(font_size = base_font_size - 2, full_width = FALSE) %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table S8 show the proportion of observations that are not missing values for each variable included in Figure 1.",
    threeparttable = T) 

df %>% group_by(country) %>%
  summarize(n = n(),
            take_vaccine_na = mean(!is.na(take_vaccine)) * 100,
            gender_na = mean(is.na(gender)) * 100,
            educ_na = mean(is.na(educ)) * 100,
            age_na = mean(is.na(age)) * 100) %>%
  kable(digits = 2, 
        col.names = c("Country", "N obs", "Take vaccine", "Gender", "Education", "Age"),
        caption = "Observations and missingness patterns.", 
        linesep = "", 
        format.args = list(big.mark = ",", scientific = FALSE),
  format = "html") %>% 
  kableExtra::kable_styling(full_width = FALSE) %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table S8 show the percentage of observations that are not missing values for each variable included in Figure 1.",
    threeparttable = T)

```


### Age distribution

```{r, fig.height = 6, fig.width = 6}
df %>% 
  select(country, age) %>%
  gather(category, value, -country) %>%
  mutate(value = as.numeric(value)) %>%
  ggplot(aes(value)) + geom_density() + facet_wrap(~country, ncol = 3)
```

### Gender, education distribution

```{r, fig.height = 13, fig.width = 15}
df %>% 
  select(country, gender, educ) %>%
  gather(category, value, -country) %>%
    ggplot(aes(value)) +  geom_bar() + facet_grid(country ~ category, scales = "free")
 
```

# Tables and Figures

The following code produces all figures and tables from the paper as well as cited statistics and supplementary material.

## Table 2: Vaccine data from WGM, WHO

```{r tab1}


# Call data from WGM
dfwgm <- read.csv("3_rep_data/table_wgm.csv")

# Call data from WHO
df_vacc_coveragebis <- read.csv("3_rep_data/vacc_cov.csv")

# Put together and order labels

table_1b <- dfwgm %>%
  left_join(df_vacc_coveragebis) %>%
  mutate(country = as.factor(country),
         country = forcats::fct_relevel(country, "Russia", "USA", after = Inf)) %>% 
  arrange(country) %>%
  select(country, Effectiveness, Safety, Important, BCG, DTP1, MCV1, Coverage)

# To Latex

tab_1b <- 
knitr::kable(
  table_1b,
  caption =  "Vaccination beliefs and coverage for the countries in our sample",
  col.names = c("",
                "Effective","Safe","Important for children to have",
                "Tuberculosis (BCG)", "Diphtheria, Tetanus and Pertussis (DTP1)",
                "Measles (MCV1)",
                "% of parents with any child that was ever vaccinated"),
  format = "latex", booktabs = T, linesep = "", align = c("l", rep("c", 7)), label = "otherv") %>% 
  kableExtra::kable_styling(latex_options = c("scale_down", "hold_position"), 
                            full_width = FALSE, font_size = base_font_size - 2)  %>%
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::add_header_above(c(" " = 1, 
                                 "% Respondents agreeing Vaccines are..." = 3,
                                 "Vaccine coverage in 2019 (% of infants)" = 3,
                                 " " = 1), bold = TRUE) %>%
  kableExtra::column_spec(1:5, width = "9em") %>% 
  kableExtra::column_spec(6:7, width = "5em") %>%
  kableExtra::column_spec(8, width = "9em") %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table 2 presents an overview of vaccination beliefs and incidence across countries in our sample. Columns 2-4 and 8 use data from the Wellcome Global Monitor 2018. Column 8 shows the percentage of respondents who are parents and report having had any of their children ever vaccinated. Columns 2-4 show the percentage of all respondents that either strongly agree or somewhat agree with the statement above each column. All percentages are obtained using national weights. Columns 5-7 use data from the World Health Organization on vaccine incidence. Columns 5-7 report the percentage of infants per country receiving the vaccine indicated in each column.", 
    threeparttable = T) %>%
  kableExtra::landscape()

knitr::kable(
  table_1b,
  caption =  "Vaccination beliefs and coverage for the countries in our sample",
  col.names = c("",
                "Effective","Safe","Important for children to have",
                "Tuberculosis (BCG)", "Diphtheria, Tetanus and Pertussis (DTP1)",
                "Measles (MCV1)",
                "% of parents with any child that was ever vaccinated"),
  format = "html", booktabs = T, linesep = "", align = c("l", rep("c", 7)), label = "otherv") %>% 
  kableExtra::kable_styling(full_width = FALSE)  %>%
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::add_header_above(c(" " = 1, 
                                 "% Respondents agreeing Vaccines are..." = 3,
                                 "Vaccine coverage in 2019 (% of infants)" = 3,
                                 " " = 1), bold = TRUE) %>%
  kableExtra::column_spec(1:5, width = "9em") %>% 
  kableExtra::column_spec(6:7, width = "5em") %>%
  kableExtra::column_spec(8, width = "9em") %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table 2 presents an overview of vaccination beliefs and incidence across countries in our sample. Columns 2-5 use data from the Wellcome Global Monitor 2018. Column 2 shows the percentage of respondents who are parents and report having had any of their children ever vaccinated. Columns 3-5 show the percentage of all respondents that either strongly agree or somewhat agree with the statement above each column. All percentages are obtained using national weights. Columns 6-8 use data from the World Health Organization on vaccine incidence. Columns 6-8 report the percentage of infants per country receiving the vaccine indicated in each column.", 
    threeparttable = T)

```

## Table with summary of samples

```{r sampling, fig.width = 6, fig.height = 8}

tab_sampling <- 
  readxl::read_excel("2_input_data/studies_info.xlsx", sheet = "sample") %>%
  dplyr::select("Study" = "country", "Date" = "date",
                "Geographic scope", "Sampling methodology", "Survey modality", "Weights") %>%
  knitr::kable(
    caption =  "Summary of studies sampling",
    format = "latex", booktabs = T, linesep = "", label = "sampling") %>% 
  kableExtra::kable_styling(latex_options = c("scale_down", "hold_position"),
                            font_size = base_font_size - 2) %>%
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::column_spec(1:2, width = "8em") %>%
  kableExtra::column_spec(3, width = "12em")  %>%
  kableExtra::column_spec(4, width = "30em") %>% 
  kableExtra::landscape()

readxl::read_excel("2_input_data/studies_info.xlsx", sheet = "sample") %>%
  dplyr::select("Study" = "country", "Date" = "date",
                "Geographic scope", "Sampling methodology", "Survey modality", "Weights") %>%
  knitr::kable(caption =  "Summary of studies' sampling", linesep = "", format = "html") %>% 
  kableExtra::kable_styling(full_width = FALSE) %>%
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::column_spec(1:2, width = "8em") %>%
  kableExtra::column_spec(3, width = "12em")  %>%
  kableExtra::column_spec(4, width = "30em") %>% 
  kableExtra::landscape()
```

## Main Results: Acceptance Rates (disaggregated by group)

### Figure 1

```{r fig1, fig.width = 7, fig.height = 7}


# Prep levels

main_results <- 
  df2 %>% 
  dplyr::filter(dplyr::if_all(c(take_vaccine_num, cluster, weight), ~ !is.na(.))) %>% 
  dplyr::nest_by(group) %>%
  dplyr::summarize(
    lm_helper(data = data, 
              formula = take_vaccine_num ~ 1, cluster = cluster,
              weight = weight, se_type = "stata"),
    .groups = "drop")

# Gender
acc_by_gender <- grp_analysis(df2, y = "take_vaccine_num", x = "gender")

# Education (all original categories and binary recoding)
acc_by_educ_binary <- grp_analysis(df2, y = "take_vaccine_num", x = "educ_binary")

# Age (all original categories and binary recoding)
acc_by_age <- 
  grp_analysis(df2, y = "take_vaccine_num", x = "age_groups_three") %>%
  dplyr::filter(statistic != Inf) %>%
  dplyr::filter(conf.low > 0)

acc_by_age_binary <- grp_analysis(df2, y = "take_vaccine_num", x = "age_groups_binary")


# Put them together in a single df. Make estimates "percentages" and round
ans <- 
  dplyr::bind_rows(
    main_results %>% mutate(cat = "All", var = "All"),
    acc_by_gender %>% rename(cat = gender) %>% mutate(var = "By gender"),
    acc_by_educ_binary %>% rename(cat = educ_binary) %>% mutate(var = "By education"),
    acc_by_age %>% rename(cat = age_groups_three) %>% mutate(var = "By age")) %>%
  dplyr::mutate(across(c(conf.low, conf.high, estimate), ~ round(. * 100, digits = 1)))

# Join with a tags df, which includes details on the study (national or subnational)
tags <- 
  readxl::read_excel("2_input_data/studies_info.xlsx", sheet = "sample") %>%
  dplyr::select(group = country, tag = "Geographic scope") %>%
  dplyr::left_join(filter(ans, cat == "All"), by = "group") %>%
  dplyr::mutate(tag = paste0(group, " (", tag, ", ", n, ")")) %>%
  dplyr::select(group, tag)

# Prepare df to plot. Important but ugly relevel of factors, happening all over the code

ans %<>% 
  dplyr::left_join(tags) %>%
  dplyr::mutate(tag = ifelse(group == "All", "All LMICs", tag)) %>%
  # group_by(var) %>% 
  # arrange(cat) %>%
  dplyr::mutate(
    var = factor(var, levels = c("All", "By gender", "By education", "By age")),
    cat = factor(
      cat, ordered = TRUE,
      levels = rev(c("Female", "Male", "Up to Secondary", 
                       "> Secondary", "<25", "25-54", "55+", "All")),
      labels = rev(c("Female", "Male", "Up to Secondary", 
                     "More than Secondary", "$< 25$", "$25-54$", "$55 +$", "All"))),
    tag = gsub(pattern = " \\(", "\\\n\\(", tag))


special_cases <- 
  sort(unique(ans$tag)[grep(unique(ans$tag), pattern = "All LMICs|Russia|USA")])

ans %<>% 
  dplyr::mutate(
    tag = 
      factor(x = tag, ordered = TRUE,
             levels = rev(c(sort(unique(tag)[!(unique(tag) %in% special_cases)]), special_cases))))

ans_loo <- ans

#Add a row averaging over only national samples

nationals <- 
  df2 %>%
  dplyr::filter(country == "Burkina Faso" |
           country == "Colombia" |
           country == "Rwanda" |
           country == "Sierra Leone 1" |
           country == "Sierra Leone 2")

main_results_n <- 
  nationals %>% 
  dplyr::filter(dplyr::if_all(c(take_vaccine_num, cluster, weight), ~ !is.na(.))) %>% 
  dplyr::nest_by(group) %>%
  dplyr::summarize(
    lm_helper(data = data, 
              formula = take_vaccine_num ~ 1, cluster = cluster,
              weight = weight, se_type = "stata"),
    .groups = "drop")

# Gender
acc_by_gender_n <- grp_analysis(nationals, y = "take_vaccine_num", x = "gender")

# Education (all original categories and binary recoding)
acc_by_educ_binary_n <- grp_analysis(nationals, y = "take_vaccine_num", x = "educ_binary")

# Age (all original categories and binary recoding)
acc_by_age_n <- 
  grp_analysis(nationals, y = "take_vaccine_num", x = "age_groups_three") %>%
  dplyr::filter(statistic != Inf) %>%
  dplyr::filter(conf.low > 0)

ans_n <- 
  dplyr::bind_rows(
    main_results_n %>% mutate(cat = "All", var = "All"),
    acc_by_gender_n %>% rename(cat = gender) %>% mutate(var = "By gender"),
    acc_by_educ_binary_n %>% rename(cat = educ_binary) %>% mutate(var = "By education"),
    acc_by_age_n %>% rename(cat = age_groups_three) %>% mutate(var = "By age")) %>%
  dplyr::mutate(across(c(conf.low, conf.high, estimate), ~ round(. * 100, digits = 1))) %>%
  filter(group == "All") %>%
  mutate(group = "All LMICs (National samples)",
         tag = "All LMICs (National samples)")  


# Prepare df to plot. Important but ugly relevel of factors, happening all over the code

ans_n %<>% 
  # group_by(var) %>% 
  # arrange(cat) %>%
  dplyr::mutate(
    var = factor(var, levels = c("All", "By gender", "By education", "By age")),
    cat = factor(
      cat, ordered = TRUE,
      levels = rev(c(c("Female", "Male", "Up to Secondary", 
                       "> Secondary", "<25", "25-54", "55+", "All"))),
      labels = rev(c("Female", "Male", "Up to Secondary", 
                     "More than Secondary", "$< 25$", "$25-54$", "$55 +$", "All"))),
    tag = gsub(pattern = " \\(", "\\\n\\(", tag)) %>%
  bind_rows(ans)

special_cases <- 
  sort(unique(ans_n$tag)[grep(unique(ans_n$tag), pattern = "All LMICs (National samples)|All LMICs|Russia|USA")])

ans_n %<>% 
  dplyr::mutate(
    tag = 
      factor(x = tag, ordered = TRUE,
             levels = rev(c(sort(unique(tag)[!(unique(tag) %in% special_cases)]), special_cases))))

ans_national <- 
  ans_n %>% 
  dplyr::filter(group == "All LMICs (National samples)" & cat == "All") 


#Colour blind palette (available palletes are smaller so we expand it)
safe_colorblind_palette <- c("#CC6677", "#DDCC77", "#117733", "#332288", "#AA4499", 
                             "#44AA99", "#999933", "#882255", "#661100", "#6699CC", "#888888", 
                             "#88CCEE")


fig_1_ages <- 
  ans_n %>% 
  ggplot(data = ., aes(x = tag, y = estimate, color = cat)) + 
  geom_errorbar(aes(ymin = conf.low, ymax = conf.high), 
                size = .5, width = .2, position = position_dodge(0.6)) + 
  geom_point(position = position_dodge(0.6)) +
  facet_grid(. ~ var, scales = "free_x", space = "free") + 
  coord_flip() +
  guides(color = guide_legend(reverse = TRUE, nrow = 2)) +
  geom_vline(xintercept = 4.5, color = "darkgrey") +
  geom_vline(xintercept = 3.5, color = "darkgrey") +
  geom_vline(xintercept = 2.5, color = "darkgrey") +
  scale_colour_manual(values = safe_colorblind_palette) +
  labs(title = "If a COVID-19 vaccine becomes available in [country], would you take it?", 
       color = "Subgroups", x = "") +
  theme_bw(base_size = (base_font_size - 2)) + ylim(c(0,100)) +
  theme(legend.position = "bottom",
        plot.caption = element_text(hjust = 0), #Default is hjust=1
        plot.title.position = "plot", #NEW parameter. Apply for subtitle too.
        plot.caption.position =  "plot",
        axis.text.y = element_text(hjust = 0))

fig_1_ages

```


### Invariance to aggregation at country level

Weighting so that each country (rather than each study) has equal weight has only a marginal impact on estimated LMIC average acceptance.


```{r}
df_country <- 
  df %>% 
  filter(!is.na(take_vaccine_num)) %>%
  filter(!(country %in% c("USA", "Russia"))) %>%
  mutate(cluster_1 = country,
         cluster_2 = recode(country, 
                          "Pakistan 1" = "Pakistan",
                          "Pakistan 2" = "Pakistan",
                          "Sierra Leone 1" = "Sierra Leone",
                          "Sierra Leone 2" = "Sierra Leone",
                          "Uganda 1" = "Uganda",
                          "Uganda 2" = "Uganda",
                          )) %>%
  group_by(country) %>% mutate(weight_1 = weight/sum(weight)) %>% ungroup() %>%
  group_by(cluster_2) %>% mutate(weight_2 = weight_1/sum(weight_1)) %>% ungroup() 
# Inspect weights 
df_country %>% group_by(country) %>% summarize(w1 = sum(weight_1), w2 = sum(weight_2)) %>% kable
unit_check <- 
  list(
    study = lm_robust(data = df_country, 
              formula = take_vaccine_num ~ 1, cluster = cluster_1,
              weight = weight_1, se_type = "stata"),
    
    country = lm_robust(data = df_country, 
              formula = take_vaccine_num ~ 1, cluster = cluster_2,
              weight = weight_2, se_type = "stata")) %>%
  
  lapply(tidy) %>% bind_rows(.id = "Unit")
unit_check %>% kable(digits = 2)
```




### Robustness check using leave-m-out approach

The main estimates for average LMIC acceptance do not vary strongly if we exclude one or two samples at a time and re-estimate average each time

```{r loo_main, fig.width=10, fig.height=7}

# Loo estimates for all LMIC respondents
loo_estimates <- 
  plyr::ldply(
    .data = list(1,2), 
    .fun = function(x) {
      df %>% 
        dplyr::filter(dplyr::if_all(c(take_vaccine_num, cluster, weight), ~ !is.na(.))) %>% 
        dplyr::filter(!(country %in% c("USA", "Russia"))) %>% 
        loo_helper(., "country", loo_n = x) %>% 
        dplyr::mutate(m = paste0("Leaving ", x, " out"), tag = "All", var = "All")
    })

# Loo estimates broken down by category (appended to above)

loo_estimates <- 
  plyr::mdply(
    .data = expand_grid(x = c(1,2), var = c("gender", "educ_binary", "age_groups_three")), 
    .fun = function(x, var) {
      df2 %>% 
        dplyr::filter(dplyr::if_all(c(take_vaccine_num, cluster, weight), ~ !is.na(.))) %>% 
        dplyr::filter(group == "All") %>% 
        loo_helper(., "country", 
                   loo_n = x, 
                   loo_fun = function(dat) grp_analysis(dat, y = "take_vaccine_num", x = var)) %>% 
        dplyr::mutate(m = paste0("Leaving ", x, " out"), var = var)
    }) %>% 
  dplyr::mutate(tag = dplyr::coalesce(gender, educ_binary, age_groups_three),
                var = plyr::mapvalues(var, 
                                      from = c("gender", "educ_binary", "age_groups_three"),
                                      to = c("By gender", "By education", "By age"))) %>%
  dplyr::bind_rows(loo_estimates, .)

# Clean up labels
loo_estimates %<>% 
  dplyr::mutate(
    var = factor(var, levels = c("All", "By gender", "By education", "By age")),
    tag = factor(
      tag, ordered = TRUE,
      levels = rev(c("Female", "Male", "Up to Secondary", 
                       "> Secondary", "<25", "25-54", "55+", "All")),
      labels = rev(c("Female", "Male", "Up to Secondary", 
                     "More than Secondary", "<25 yr", "25-54 yr", "55+ yr", "All")))
  ) %>% 
  dplyr::select(var, m, tag, estimate)

safe_colorblind_palette <- c("#CC6677", "#DDCC77", "#117733", "#332288", "#AA4499", 
                             "#44AA99", "#999933", "#882255", "#661100", "#6699CC",
                             "#888888", "#88CCEE")

ans_loo %<>% 
  dplyr::filter(group %in% c("All", "USA", "Russia")) %>% 
  dplyr::mutate(
    group = plyr::mapvalues(group, from = "All", to = "All LMIC"),
    cat = plyr::mapvalues(cat, from = c("$< 25$", "$25-54$", "$55 +$"),
                          to = c("<25 yr", "25-54 yr", "55+ yr")),
    tag = cat)

fig_hist_loo <- 
  loo_estimates %>%
  dplyr::mutate(estimate = estimate*100) %>% 
  ggplot(aes(estimate, color = tag, fill = tag)) + 
  geom_histogram(aes(y = ..density..), bins = 200, 
                 position = "dodge", alpha = .3, size = .3) +
  geom_vline(data = ans_loo, 
             aes(xintercept = estimate, color = cat, linetype = group), size = .9) +
  facet_grid_paginate(var + tag ~ m)  +
  scale_color_manual(
    name = "Subgroups", 
    values = safe_colorblind_palette) + 
  scale_fill_manual(
    name = "Subgroups", 
    values = safe_colorblind_palette) + 
  scale_linetype_manual(
    name = "Sample Average", 
    values = c("solid", "11", "dashed")) + 
  scale_x_continuous(n.breaks = 8, name = "estimate") +
  scale_y_continuous(
    labels = scales::percent_format(scale = 1, suffix = "", accuracy = 0.1)) +
  labs(title = "",
       x = "") +
  guides(color = guide_legend(reverse = TRUE, nrow = 2, keyheight = 1),
         fill = guide_legend(reverse = TRUE, nrow = 2, keyheight = 1),
         linetype = guide_legend(keyheight = 2)) +
  theme_bw(base_size = (base_font_size - 2)) +
  theme(legend.position = "bottom")

fig_hist_loo

```

### Differences in means

```{r analysis}


# Analysis of differences in means only LMICs
# Notice that Uganda 1 is dropped, because it does not have reference categories for gender or age
# Notice that we are using df (and not df2) as data, since it does not include "All"

# Population estimate (clustering on country)

differences_means_gen_age <- 
  lapply(c("gender", "age_groups_binary", "age_groups_three"), function(i) {
    df %>% 
      dplyr::filter(country != "USA" & country != "Russia" & country != "Uganda 1") %>%
      dplyr::filter(if_all(c(all_of(i), all_of("take_vaccine_num")), ~ !is.na(.))) %>%
      study_weighting() %>%
      estimatr::lm_robust(as.formula(paste("take_vaccine_num ~", i)),
                          fixed_effects = ~country,
                          weight = weight,
                          cluster = country,
                          se_type = "stata",
                          data = .) %>% 
      tidy %>% 
      dplyr::select(estimate, std.error, p.value, df, term)
  }) %>% 
  dplyr::bind_rows(.)

differences_means_educ <- 
    df %>% 
      dplyr::filter(country != "USA" & country != "Russia") %>% 
      dplyr::filter(if_all(c(all_of("educ_binary"), all_of("take_vaccine_num")), ~ !is.na(.))) %>%
      study_weighting() %>%
      estimatr::lm_robust(
        take_vaccine_num ~educ_binary, 
        fixed_effects = ~country, 
        weight = weight, 
        cluster = country,
        se_type = "stata",
        data = .) %>% 
      tidy %>% 
      dplyr::select(estimate, std.error, p.value, df, term)


dif_gender <- 
  differences_means_gen_age %>% 
  dplyr::filter(term == "genderMale") %>% 
  pull(estimate) %>% 
  {. * 100} %>% 
  round(., 1)

dif_age <- 
  differences_means_gen_age %>% 
  dplyr::filter(term == "age_groups_binary55+") %>% 
  pull(estimate) %>% 
  {. * 100} %>% 
  round(., 1)

dif_age_three <- 
  differences_means_gen_age %>%
  dplyr::filter(term == "age_groups_three25-54" | term == "age_groups_three55+") %>%
  pull(estimate) %>% 
  {. * 100} %>% 
  round(., 1)
 
dif_educ <- 
  differences_means_educ %>% 
  dplyr::filter(term == "educ_binaryUp to Secondary") %>% 
  pull(estimate) %>% 
  {. * 100} %>% 
  round(., 1)

diffmeans <- 
  rbind(differences_means_gen_age, differences_means_educ) %>%
  dplyr::rename(Estimate = estimate,
         Std.error = std.error,
         `P-value` = p.value,
         "Degrees of freedom" = df,
         "Baseline category" = term) %>%
  dplyr::mutate(
    Variable = ifelse(`Baseline category` == "genderMale", 
                      "Gender (Female)", ""),
    Variable = ifelse(`Baseline category` == "age_groups_binary55+", 
                      "Age", Variable),
    Variable = ifelse(`Baseline category` == "educ_binaryUp to Secondary", 
                      "Education (Secondary +)", Variable),
    `Baseline category` = ifelse(`Baseline category` == "genderMale", 
                                 "Male", `Baseline category`),
    `Baseline category` = ifelse(`Baseline category` == "age_groups_binary55+", 
                                 "55+", `Baseline category`),
    `Baseline category` = ifelse(`Baseline category` == "educ_binaryUp to Secondary", 
                                 "Up to secondary", `Baseline category`)) 


diffmeans <- 
  diffmeans %>%
  filter(Variable != "Age") %>%
  dplyr::mutate(
    Variable = ifelse(`Baseline category` == "age_groups_three25-54", 
                      "Age (25-54)", Variable),
    Variable = ifelse(`Baseline category` == "age_groups_three55+", 
                      "Age (55+)", Variable),
    `Baseline category` = ifelse(`Baseline category` == "age_groups_three25-54", 
                                 "<25", `Baseline category`),
    `Baseline category` = ifelse(`Baseline category` == "age_groups_three55+", 
                                 "<25", `Baseline category`))

dmeans <- diffmeans %>%
  filter(Variable != "") %>%
  knitr::kable(
    digits = 2,
    caption =  "Differences in means",
    format = "latex", booktabs = T, linesep = "", label = "dmeans", align = "c") %>% 
  kableExtra::kable_styling(latex_options = c("hold_position"),
    font_size = base_font_size - 2, full_width = FALSE) %>%
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table S7 shows the results of subgroup mean differences. Subgroup differences were generated considering only LMICs. p-values come from a two-sided t-test from a linear regression.",
    threeparttable = T) 

diffmeans %>%
  filter(Variable != "") %>%
  knitr::kable(
  caption =  "Differences in means",
  booktabs = T, linesep = "", label = "dmeans", digits = 2) %>% 
  kableExtra::kable_styling(full_width = FALSE) %>%
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table S7 shows the results of subgroup mean differences. Subgroup differences were generated considering only LMICs. The differences in means for gender and age do not include the Uganda 1 study, which only included female respondents under the age of 55.",
    threeparttable = T) 

```




### Differences in means: by study 

```{r}
country_differences <-
  unique(df$country) %>%
  lapply(function(j){{
    dff <- filter(df, country == j)
    
    lapply(c("gender", "age_groups_three", "educ_binary"), function(i){
      if (length(table(dff[[i]])) < 2)  {
        return(NULL)
      } else {
        m <- estimatr::lm_robust(as.formula(paste("take_vaccine_num ~", i)),
                            weight = weight,
                            cluster = cluster,
                            se_type = "stata",
                            data = dff) 
          m %>% 
            tidy %>%
            dplyr::select(estimate, std.error, p.value, df, term) %>%
            dplyr::mutate(n = m$nobs)
        }}
    ) } %>%
      dplyr::bind_rows() %>% 
      dplyr::mutate(country = j)}) %>% 
  dplyr::bind_rows() %>% 
  dplyr::arrange(term, country) %>% 
  dplyr::relocate(country, term) %>%
  dplyr::filter(term != "(Intercept)") %>% 
  dplyr::mutate(significant = p.value <= .05) %>%
  dplyr::mutate(
    term = ifelse(term == "age_groups_three25-54", "25-54", term),
    term = ifelse(term == "age_groups_three55+", "55+", term),
    term = ifelse(term == "educ_binaryUp to Secondary", "Up to secondary", term),
    term = ifelse(term == "genderMale", "Male", term))

country_differences_summary <- 
  country_differences %>% 
  dplyr::filter(!(country %in% c("Russia", "USA"))) %>% 
  dplyr::group_by(term) %>% summarize(
    "positive " = sum(estimate > 0),
    "positive and significant" = sum(estimate > 0 & significant),
    "negative and significant" = sum(estimate < 0 & significant),
    "not significant" = sum(!significant),
    n = n()) 

t_country_differences <- country_differences %>%
  dplyr::mutate(
                baseline = ifelse(term == "Male", "Female", NA),
                baseline = ifelse(term == "Up to secondary", "Secondary +", baseline),
                baseline = ifelse(is.na(baseline), "<25", baseline),
                group = ifelse(baseline == "<25", "Age", NA),
                group = ifelse(baseline == "Secondary +", "Education", group),
                group = ifelse(baseline == "Female", "Gender", group),
                estimate = round(estimate, 2),
                std.error = round(std.error, 2),
                p.value = round(p.value, 2)) %>%
  dplyr::select(country, group, baseline, term, everything(), -significant)

t_country <- t_country_differences %>%
  kable(digits = 2, 
        col.names = c("Country", "Variable", "Baseline category", "Group", "Estimate", "Std. Error", "P-value", "Degrees of freedom", "N Obs"),
        caption = "Differences between groups within studies", 
        booktabs = TRUE, linesep = "", 
        format.args = list(big.mark = ",", scientific = FALSE),
  format = "latex", label = "countrydiff") %>% 
  kableExtra::kable_styling(latex_options = c("scale_down"), font_size = base_font_size - 2, full_width = FALSE) %>% 
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::column_spec(1, width = "6em") %>% 
  kableExtra::column_spec(2:3, width = "6em") %>% 
  kableExtra::column_spec(4, width = "9em")  %>% 
  kableExtra::column_spec(5:8, width = "6em") %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table S9 shows differences of means between groups within single studies. Estimates are calculated through OLS and represent the difference in the average acceptance rate between the subgroup in column Group and that in column Baseline category. p-values come from a two-sided t-test from a linear regression.",
    threeparttable = T)


knitr::kable(t_country_differences, 
             digits = 2, caption = "Differences between groups within studies ")


knitr::kable(country_differences_summary, 
             digits = 2, caption = "Differences between groups within studies (Summary)")

```

### Metaplus analysis

```{r, comment = ""}
# Metaplus (variance between countries, included in text)
mp <- metaplus::metaplus(
  yi  = main_results %>% filter(!(group %in% c("USA", "Russia", "All"))) %>% pull(estimate),
  sei = main_results %>% filter(!(group %in% c("USA", "Russia", "All"))) %>% pull(std.error))
mp_tau <- mp$results[2,1]
mp_ratio <- sqrt(mp$results[2,1])/mp$results[1,1]

mp$results %>% kable(caption = "Metaplus  results")

```

### Generate quantities used in text

```{r}

# Mean of LMICs
ans_mean <- 
  ans %>% 
  dplyr::filter(cat == "All", group == "All") %>% 
  pull(estimate)

# Lowest and highest bounds for All mean estimate
ans_mean_low <- 
  ans %>% 
  dplyr::filter(cat == "All", group == "All") %>% 
  pull(conf.low)

ans_mean_high <-  
  ans %>% 
  dplyr::filter(cat == "All", group == "All") %>% 
  pull(conf.high)

# df of LMICs (no USA, Russia or All)
ans_stats <- 
  ans %>% 
  dplyr::filter(cat == "All", !(group %in% c("USA", "Russia", "All")))

# Median of LMICs
ans_median <- median(ans_stats$estimate)

# Interquartile range of LMIcs
ans_iqr <- IQR(ans_stats$estimate)

# Smallest and biggest acceptability of LMICs
ans_min <- 
  ans %>% 
  dplyr::filter(cat == "All" & group != "Russia" & group != "USA") %>% 
  pull(estimate) %>% min

ans_max <- 
  ans %>% 
  dplyr::filter(cat == "All" & group != "Russia" & group != "USA") %>% 
  pull(estimate) %>% max

# Rank by estimate
top_all <- 
  ans %>%
  dplyr::filter(cat == "All") %>% 
  dplyr::arrange(desc(estimate))

bottom_all <- 
  ans %>%
  dplyr::filter(cat == "All" & group != "Russia" & group != "USA") %>% 
  dplyr::arrange(estimate)

# USA and Russia estimate acceptability
usa_ans <- 
  ans %>%
  dplyr::filter(cat == "All" & group == "USA") 

rus_ans <- 
  ans %>%
  dplyr::filter(cat == "All" & group == "Russia") 

```

### Table version of Figure 1

```{r tabfig1}

# Here we are making "percentages" from the estimates and putting them together with confidence intervals
# Also we are going from long to wide

main_table <- 
  main_results %>%
  dplyr::mutate(
    across(c(estimate, conf.low, conf.high), list("_main" = ~ format(round(100 * ., 1), nsmall = 1))),
    conf_int__main = paste0("(", conf.low__main, ", ", conf.high__main, ")")) %>%
  dplyr::select(group, estimate__main, conf_int__main)


gender_table <- 
  acc_by_gender %>%
  dplyr::mutate(
    across(c(estimate, conf.low, conf.high), ~ format(round(100 * ., 1), nsmall = 1)),
    conf_int = paste0("(", conf.low, ", ", conf.high, ")")) %>%
  dplyr::select(group, gender, estimate, conf_int) %>% 
  tidyr::pivot_wider(names_from = gender, 
                     values_from = c(estimate, conf_int), 
                     names_sep = "__")

educ_table <- 
  acc_by_educ_binary %>%
  dplyr::mutate(
    across(c(estimate, conf.low, conf.high), ~ format(round(100 * ., 1), nsmall = 1)),
    conf_int = paste0("(", conf.low, ", ", conf.high, ")")) %>%
  dplyr::select(group, educ_binary, estimate, conf_int) %>% 
  tidyr::pivot_wider(names_from = educ_binary, 
                     values_from = c(estimate, conf_int), 
                     names_sep = "__")


age_table <- 
  acc_by_age %>%
  dplyr::mutate(
    across(c(estimate, conf.low, conf.high), ~ format(round(100 * ., 1), nsmall = 1)),
    conf_int = paste0("(", conf.low, ", ", conf.high, ")")) %>%
  dplyr::select(group, age_groups_three, estimate, conf_int) %>% 
  tidyr::pivot_wider(names_from = age_groups_three, 
                     values_from = c(estimate, conf_int), 
                     names_sep = "__")

main_table_n <- 
  main_results_n %>%
  dplyr::mutate(
    across(c(estimate, conf.low, conf.high), list("_main" = ~ format(round(100 * ., 1), nsmall = 1))),
    conf_int__main = paste0("(", conf.low__main, ", ", conf.high__main, ")")) %>%
  dplyr::select(group, estimate__main, conf_int__main) 


gender_table_n <- 
  acc_by_gender_n %>%
  dplyr::mutate(
    across(c(estimate, conf.low, conf.high), ~ format(round(100 * ., 1), nsmall = 1)),
    conf_int = paste0("(", conf.low, ", ", conf.high, ")")) %>%
  dplyr::select(group, gender, estimate, conf_int) %>% 
  tidyr::pivot_wider(names_from = gender, 
                     values_from = c(estimate, conf_int), 
                     names_sep = "__")

educ_table_n <- 
  acc_by_educ_binary_n %>%
  dplyr::mutate(
    across(c(estimate, conf.low, conf.high), ~ format(round(100 * ., 1), nsmall = 1)),
    conf_int = paste0("(", conf.low, ", ", conf.high, ")")) %>%
  dplyr::select(group, educ_binary, estimate, conf_int) %>% 
  tidyr::pivot_wider(names_from = educ_binary, 
                     values_from = c(estimate, conf_int), 
                     names_sep = "__")


age_table_n <- 
  acc_by_age_n %>%
  dplyr::mutate(
    across(c(estimate, conf.low, conf.high), ~ format(round(100 * ., 1), nsmall = 1)),
    conf_int = paste0("(", conf.low, ", ", conf.high, ")")) %>%
  dplyr::select(group, age_groups_three, estimate, conf_int) %>% 
  tidyr::pivot_wider(names_from = age_groups_three, 
                     values_from = c(estimate, conf_int), 
                     names_sep = "__")

all_tables_n <- 
  main_table_n %>% 
  dplyr::left_join(gender_table_n) %>%
  dplyr::left_join(educ_table_n) %>% 
  dplyr::left_join(age_table_n) %>%
  dplyr::filter(group == "All") %>%
  dplyr::mutate(group = "All LMICs (National)")

all_tables <- 
  main_table %>% 
  dplyr::left_join(gender_table) %>%
  dplyr::left_join(educ_table) %>% 
  dplyr::left_join(age_table) %>%
  dplyr::bind_rows(all_tables_n) %>%
  dplyr::select(group, everything()) %>%
  dplyr::mutate(group = as.factor(group),
                group = forcats::fct_relevel(group, "All", "All LMICs (National)", "Russia", "USA", after = Inf)) %>% 
  dplyr::arrange(group) %>% 
  tidyr::pivot_longer(cols = c(starts_with("estimate__"), starts_with("conf_int__")),
                      names_to = c("type", ".value"),
                      names_pattern = "(.*)__(.*)") %>% 
  dplyr::mutate(group = ifelse(type == "conf_int", "", as.character(group)),
                group = ifelse(group == "All", "All LMICs", group)) %>% 
  dplyr::select(-type) %>% 
  dplyr::rename("Country" = "group", 
                "Average acceptability" = "main")



tab_fig1 <- 
  all_tables %>%
  knitr::kable(
    all_tables,
    caption = "If a COVID-19 vaccine becomes available in [country], would you take it? Disaggregated by subgroups", 
    format = "latex",  booktabs = T, linesep = "", 
    format.args = list(big.mark = ",", scientific = FALSE), 
    escape = F, align = "lcccccccc", label = "maintabledis")  %>% 
  kableExtra::kable_styling(latex_options = c("scale_down", "hold_position"),
                            font_size = base_font_size - 2, full_width = FALSE) %>%
  kableExtra::add_header_above(c("", "", "Gender" = 2, "Education" = 2, "Age" = 3), bold = TRUE) %>%
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table S1 shows percentage of respondents willing to take the COVID-19 vaccine as plotted in Figure 1. A 95% confidence interval is shown between parentheses.",
    threeparttable = T) 


all_tables %>%
  knitr::kable(
    caption = "If a COVID-19 vaccine becomes available in [country], would you take it? Disaggregated by subgroups", 
    align = "lcccccccc", booktabs = T) %>% 
  kableExtra::kable_styling(full_width = FALSE) %>%
  kableExtra::add_header_above(c("", "", "Gender" = 2, "Education" = 2, "Age" = 3), bold = TRUE) %>%
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table S1 shows percentage of respondents willing to take the COVID-19 vaccine as plotted in Figure 1. A 95% confidence interval is shown between parentheses",
    threeparttable = T) 

```

## Reasons

### Table: Reasons to take


```{r tab2, fig.width = 12}


#There are idiosyncratic reasons why people would take the vaccine. I recoded them. But we keep only the core, which is common almost in all studies.
yes_vars <- 
  df2 %>% 
  dplyr::select(yes_vaccine_1, yes_vaccine_2, yes_vaccine_3) %>% 
  names

## Generate data for analysis of yes reasons
yes_vacc1 <- 
  lapply(yes_vars, reasons_together, df = df2, num = "Yes") %>%
  dplyr::bind_rows() %>%
  dplyr::mutate(across(c(conf.low, conf.high, estimate), ~ round(. * 100, digits = 0)))

#Get percentage per yes reason category and make wide table
yes_vacc2 <- 
  yes_vacc1 %>%
  dplyr::mutate(estimate = format(estimate, nsmall = 0),
                conf_int = paste0("(", conf.low, 
                                  ", ", conf.high, ")")) %>%
  dplyr::select(group, estimate, conf_int, outcome, n) %>%
  tidyr::pivot_wider(names_from = outcome, 
                     values_from = c(estimate, conf_int, n), 
                     names_sep = "__") %>%
  tidyr::pivot_longer(cols = c(starts_with("estimate__"), starts_with("conf_int__")),
                      names_to = c("type", ".value"),
                      names_pattern = "(.*)__(.*)") %>%
  dplyr::rowwise() %>% 
  dplyr::mutate(
    n = ifelse(group == "All", NA, unique(na.omit(c_across(starts_with("n__")))))) %>% 
  dplyr::ungroup() %>% 
  dplyr::mutate(
    group = forcats::fct_relevel(as.factor(group), "All", "Russia", "USA", after = Inf)) %>% 
  dplyr::arrange(group) %>% 
  dplyr::mutate(across(c(group, n), ~ifelse(type == "conf_int", "", as.character(.))),
                group = ifelse(group == "All", "All LMICs", group)) %>% 
  dplyr::select(group, n, type, starts_with("yes_vaccine_"), -starts_with("n_yes_vaccine"), -type)

cnames <- c("Study", "N", "Self", "Family", 
          "Community")


#Table to Latex
tab_reasons_y <- 
  yes_vacc2 %>%
  knitr::kable(
    col.names = cnames,
    caption = "Reasons to take the vaccine", format = "latex", booktabs = T, 
    linesep = "", label = "yes", 
    format.args = list(big.mark = ",", scientific = FALSE), 
    escape = F, align = "lcccc") %>%
  kableExtra::kable_styling(full_width = FALSE, 
                            font_size = base_font_size - 2) %>%
  kableExtra::add_header_above(c(" " = 2, "Protection" = 3), bold = TRUE) %>% 
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::column_spec(1, width = "8em") %>% 
  kableExtra::column_spec(2:4, width = "4em") %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table S2 shows percentage of respondents mentioning reasons why they would take the Covid-19 vaccine. The number of observations and percentage corresponds only to people who would take the vaccine. Respondents in all countries could give more than one reason. A 95% confidence interval is shown between parentheses. Studies India, Pakistan 1 and Pakistan 2 are not included because they either did not include the question or were not properly harmonized with the other studies.",
    threeparttable = T) 

yes_vacc2 %>%
  knitr::kable(
    col.names = cnames,
    caption = "\\label{yes}Reasons to take the vaccine", 
    booktabs = T, linesep = "", 
    format.args = list(big.mark = ",", scientific = FALSE),
    align = "lcccc") %>% 
  kableExtra::kable_styling(full_width = FALSE) %>%
  kableExtra::add_header_above(c(" " = 2, "Protection" = 3), bold = TRUE) %>% 
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::column_spec(1, width = "8em") %>% 
  kableExtra::column_spec(2:5, width = "8em") %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table S2 shows percentage of respondents mentioning reasons why they would take the Covid-19 vaccine. The number of observations and percentage correponds only to people who would take the vaccine. Respondents in all countries could give more than one reason. A 95% confidence interval is shown between parentheses. Studies India, Pakistan 1 and Pakistan 2 are not included because they either did not include the question or were not properly harmonized with the other studies.",
    threeparttable = T) 

```


### Calculate numbers used in text for acceptance

```{r analysis_yes, fig.width = 12, fig.height=10}

# All LMIcs estimate of self protection
yes_all <- 
  yes_vacc1 %>% 
  dplyr::filter(group == "All" & outcome == "yes_vaccine_1")

# Top and bottom LMICs countries for category self protection
yes_ <- 
  yes_vacc1 %>% 
  dplyr::filter(group != "All" & group != "USA" & group != "Russia" & outcome == "yes_vaccine_1") 

yes_top <- yes_ %>% dplyr::arrange(desc(estimate))
yes_bottom <- yes_ %>% dplyr::arrange(estimate)

# All LMICs estimate of protection of family
yes_all_2 <- 
  yes_vacc1 %>% 
  dplyr::filter(group == "All" & outcome == "yes_vaccine_2")

# Top and bottom LMICs countries for category protection of family
yes_2 <- 
  yes_vacc1 %>% 
  dplyr::filter(group != "All" & group != "USA" & group != "Russia" & outcome == "yes_vaccine_2") 

yes2_top <- yes_2 %>% dplyr::arrange(desc(estimate))
yes2_bottom <- yes_2 %>% dplyr::arrange(estimate)

# Estimate of self protection for Russia and the US
yes_usa <- 
  yes_vacc1 %>% 
  dplyr::filter(group != "All" & group == "USA" & outcome == "yes_vaccine_1") 

yes_rus <- 
  yes_vacc1 %>% 
  dplyr::filter(group != "All" & group == "Russia" & outcome == "yes_vaccine_1") 

# Estimate of protection of family for Russia and the US
yes2_usa <- 
  yes_vacc1 %>% 
  dplyr::filter(group != "All" & group == "USA" & outcome == "yes_vaccine_2")

yes2_rus <- 
  yes_vacc1 %>% 
  dplyr::filter(group != "All" & group == "Russia" & outcome == "yes_vaccine_2") 

```

```{r tab2all, fig.width = 12}


yes_vars <- 
  df2 %>% 
  dplyr::select(yes_vaccine_1, yes_vaccine_2, 
                yes_vaccine_3, yes_vaccine_4, 
                yes_vaccine_5, yes_vaccine_666) %>% 
  names

## Generate data for analysis of yes reasons
yes_vacc1 <- 
  lapply(yes_vars, reasons_together, df = df2, num = "Yes") %>%
  dplyr::bind_rows() %>%
  dplyr::mutate(across(c(conf.low, conf.high, estimate), ~ round(. * 100, digits = 0)))


#Get percentage per yes reason category and make wide table
yes_vacc2 <- 
  yes_vacc1 %>%
  dplyr::mutate(estimate = format(estimate, nsmall = 0),
                conf_int = paste0("(", conf.low, 
                                  ", ", conf.high, ")")) %>%
  dplyr::select(group, estimate, conf_int, outcome, n) %>%
  tidyr::pivot_wider(names_from = outcome, 
                     values_from = c(estimate, conf_int, n), 
                     names_sep = "__") %>%
  tidyr::pivot_longer(cols = c(starts_with("estimate__"), starts_with("conf_int__")),
                      names_to = c("type", ".value"),
                      names_pattern = "(.*)__(.*)") %>%
  dplyr::rowwise() %>% 
  dplyr::mutate(
    n = ifelse(group == "All", NA, unique(na.omit(c_across(starts_with("n__")))))) %>% 
  dplyr::ungroup() %>% 
  dplyr::mutate(
    group = forcats::fct_relevel(as.factor(group), "All", "Russia", "USA", after = Inf)) %>% 
  dplyr::arrange(group) %>% 
  dplyr::mutate(across(c(group, n), ~ifelse(type == "conf_int", "", as.character(.))),
                group = ifelse(group == "All", "All LMICs", group)) %>% 
  dplyr::select(group, n, type, starts_with("yes_vaccine_"), -starts_with("n_yes_vaccine"), -type)

cnames <- c("Study", "N", "Self", "Family", 
          "Community", "Health workers", "Government", "Other")

#Table to Latex
tab_reasons_y_all <- 
  yes_vacc2  %>% 
  knitr::kable(col.names = cnames,
        caption = "Reasons to take the vaccine- all categories", 
        format = "latex", booktabs = T, linesep = "", 
        label = "yesall", 
        format.args = list(big.mark = ",", scientific = FALSE), 
        escape = F, align = c("l", rep("c", 7))) %>%
  kableExtra::kable_styling(latex_options = c("scale_down", "hold_position"),
                            font_size = base_font_size - 2, full_width = FALSE) %>%
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::add_header_above(c(" " = 2, "Protection" = 3, "If recommended by" = 2, " " = 1), 
                               bold = TRUE) %>% 
  kableExtra::column_spec(1:8, width = "7em") %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table S3 shows percentage of respondents mentioning reasons why they would take the Covid-19 vaccine. The number of observations and percentage correponds only to people who would take the vaccine. Respondents in all countries could give more than one reason. A 95% confidence interval is shown between parentheses.",
    threeparttable = T) 

yes_vacc2 %>%
  knitr::kable(col.names = cnames,
               caption = "Reasons to take the vaccine. All categories", 
               booktabs = T, linesep = "", align = c("l", rep("c", 7)),
               format.args = list(big.mark = ",", scientific = FALSE)) %>% 
  kableExtra::kable_styling(full_width = FALSE) %>%
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::add_header_above(c(" " = 2, "Protection" = 3, "If recommended by" = 2, " " = 1), 
                               bold = TRUE) %>% 
  kableExtra::column_spec(1:8, width = "7em") %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table S3 shows percentage of respondents mentioning reasons why they would take the Covid-19 vaccine. The number of observations and percentage correponds only to people who would take the vaccine. Respondents in all countries could give more than one reason. A 95% confidence interval is shown between parentheses.",
    threeparttable = T) 

  

```

### Reasons to take (by age)

```{r, fig.width = 10, fig.height = 7}

yes_vars <- 
  df2 %>% 
  dplyr::select(yes_vaccine_1, yes_vaccine_2, yes_vaccine_3) %>% 
  names

## Generate data for analysis of yes reasons for different age groups
yes_vacc_age_1 <- 
  lapply(yes_vars, age_analysis, df = df2, num = "Yes", filter_by = `age_less24`) %>%
  dplyr::bind_rows() %>%
  mutate(age = "<25")

yes_vacc_age_2 <- 
  lapply(yes_vars, age_analysis, df = df2, num = "Yes", filter_by = `age_25_54`) %>%
  dplyr::bind_rows() %>%
  mutate(age = "25-54")

yes_vacc_age_3 <- 
  lapply(yes_vars, age_analysis, df = df2, num = "Yes", filter_by = `age_55_more`) %>%
  dplyr::bind_rows() %>%
  mutate(age = "55+")

#Get percentage per yes reason category and make wide table
yes_vacc_age <- 
  rbind(yes_vacc_age_1, yes_vacc_age_2, yes_vacc_age_3) %>%
  dplyr::mutate(across(c(conf.low, conf.high, estimate), ~ round(. * 100, digits = 0))) %>%
  dplyr::mutate(estimate = format(estimate, nsmall = 0),
                conf_int = paste0("(", conf.low, 
                                  ", ", conf.high, ")"),
                n = as.character(n)) %>%
  dplyr::select(group, estimate, conf_int, outcome, age, n) %>%
  tidyr::pivot_wider(names_from = c(outcome, age), values_from = c(estimate, conf_int, n), names_sep = "__") %>%
  tidyr::pivot_longer(cols = c(starts_with("estimate__"), starts_with("conf_int__"), starts_with("n__")),
                      names_to = c("type", ".value"),
                      names_pattern = "(.*)__(.*)")

y1 <- yes_vacc_age %>%
  dplyr::filter(grepl("yes_vaccine_1", type)) %>%
  dplyr::mutate(type = ifelse(grepl("estimate", type), "estimate", type),
         type = ifelse(grepl("conf_int", type), "conf_int", type),
         type = ifelse(grepl("n__", type), "n", type))


y2 <- yes_vacc_age %>%
  dplyr::filter(grepl("yes_vaccine_2", type)) %>%
  dplyr::mutate(type = ifelse(grepl("estimate", type), "estimate", type),
         type = ifelse(grepl("conf_int", type), "conf_int", type),
         type = ifelse(grepl("n__", type), "n", type))

y3 <- yes_vacc_age %>%
  dplyr::filter(grepl("yes_vaccine_3", type)) %>%
  dplyr::mutate(type = ifelse(grepl("estimate", type), "estimate", type),
         type = ifelse(grepl("conf_int", type), "conf_int", type),
         type = ifelse(grepl("n__", type), "n", type))

yes_vacc_age <- 
  dplyr::left_join(y1, y2, by = c("group", "type")) %>%
  dplyr::left_join(y3, by = c("group", "type")) %>% 
  dplyr::mutate(
    group = forcats::fct_relevel(as.factor(group), "All", "Russia", "USA", after = Inf)) %>% 
  dplyr::arrange(group) %>% 
  dplyr::mutate(
    across(c(group), 
           ~ifelse(type == "conf_int", "Conf. interval", as.character(.))),
    across(c(group), ~ifelse(type == "n", "n", as.character(.))),
    group = ifelse(group == "All", "All LMICs", group)) %>% 
  dplyr::select(-type)

cnames <- c("Study", "<25", "25-54", "55+", "<25", "25-54", "55+", "<25", "25-54", "55+")


#Table to Latex
tab_reasons_y_age <- 
  yes_vacc_age %>%
  knitr::kable(
    col.names = cnames,
    caption = "Reasons to take the vaccine- by age groups", 
    format = "latex", booktabs = T, 
    linesep = "", label = "yes1", 
    format.args = list(big.mark = ",", scientific = FALSE), 
    escape = F, align = c("l", rep("c", 9))) %>%
  kableExtra::kable_styling(latex_options = c("scale_down", "hold_position"),
                            font_size = base_font_size - 2, full_width = FALSE) %>%
  kableExtra::add_header_above(c(" " = 1, "Self" = 3, "Family" = 3, "Community" = 3), 
                               bold = TRUE) %>% 
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::row_spec(seq(from = 3, to = 36, by = 3), hline_after = TRUE) %>%
  kableExtra::column_spec(1, width = "7em") %>% 
  kableExtra::column_spec(2:10, width = "5em") %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table S4 shows percentage of respondents mentioning reasons why they would take the Covid-19 vaccine by age groups. The number of observations and percentage correponds only to people who would take the vaccine. Respondents in all countries could give more than one reason. A 95% confidence interval is shown between parentheses.",
    threeparttable = T) 


yes_vacc_age %>%
  knitr::kable(
    col.names = cnames,
    caption = "\\label{yes1}Reasons to take the vaccine", 
    booktabs = T, linesep = "", 
    format.args = list(big.mark = ",", scientific = FALSE),
    align = c("l", rep("c", 9))) %>% 
  kableExtra::kable_styling(full_width = FALSE) %>%
  kableExtra::add_header_above(c(" " = 1, "Self" = 3, "Family" = 3, "Community" = 3), 
                               bold = TRUE) %>% 
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::column_spec(1, width = "8em") %>% 
  kableExtra::column_spec(2:10, width = "4em") %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table S4 shows percentage of respondents mentioning reasons why they would take the Covid-19 vaccine by age groups. The number of observations and percentage correponds only to people who would take the vaccine. Respondents in all countries could give more than one reason. A 95% confidence interval is shown between parentheses.",
    threeparttable = T) 

```

### Figure 2: reasons not to take

```{r, fig.width = 10, fig.height = 7}

#Import df with tags (names of categories)
dictionary <- read_excel("2_input_data/dictionary.xlsx")

#Select core variables to be included in figure
no_vars <- 
  df2 %>% 
  dplyr::select(starts_with("no_vaccine_")) %>% 
  names

#Create estimates of % of each reason, re-level factors and get and categorize number of observations by subgroup
no_vacc <- 
  lapply(no_vars, reasons_together, df = df2, num = c("No", "DK")) %>%
  dplyr::bind_rows() %>% 
  dplyr::arrange(outcome) %>% 
  dplyr::mutate(
    across(c(conf.low, conf.high, estimate), ~ round(. * 100, digits = 1)),
    n_sub = round(n * estimate, 0),
    n_sub = ifelse(n_sub == 0, NA_integer_, n_sub),
    group = 
      factor(group, 
             levels = rev(c("Burkina Faso", "Colombia", "Mozambique", 
                            "Nepal", "Nigeria", "Pakistan 1", "Rwanda", 
                            "Sierra Leone 1", "Sierra Leone 2", "Uganda 1", 
                            "Uganda 2", "All", "Russia", "USA" )))
  ) %>%
  dplyr::left_join(dictionary, by = "outcome") %>%
  dplyr::mutate(
    name = ifelse(group != "All", paste0(group, " (n=", n, ")"), "All"),
    name = gsub(pattern = " \\(", "\\\n\\(", name),
    tag = as.factor(tag),
    tag = forcats::fct_relevel(
      tag,  
      "Concerned about side effects", 
      "Concerned about getting coronavirus from the vaccine", 
      "Not concerned about getting seriously ill", 
      "Doesn't think vaccines are effective", 
      "Doesn't think Coronavirus outbreak is as serious as people say", 
      "Doesn't like needles", 
      "Allergic to vaccines", 
      "Won't have time to get vaccinated", 
      "Mentions a conspiracy theory", 
      "Other reasons"))

special_cases <- 
  sort(unique(no_vacc$name)[grep(unique(no_vacc$name), pattern = "All|Russia|USA")])


no_vacc %<>% 
  dplyr::mutate(
    name = 
      factor(x = name, ordered = TRUE,
             levels = rev(c(sort(unique(name)[!(unique(name) %in% special_cases)]), special_cases))))

fig_2 <- 
  no_vacc %>% 
  dplyr::filter(!is.na(n_sub)) %>%
  dplyr::mutate(name = plyr::mapvalues(name, "All", "All LMICs")) %>% 
  ggplot(aes(name, estimate, color = tag)) + 
  geom_errorbar(aes(ymin = conf.low, ymax = conf.high), 
                size = .5, width = .2, position = position_dodge(0.6)) + 
  geom_point(shape = 16, position = position_dodge(0.6)) +
  facet_grid(.~tag,  space = "free", labeller = label_wrap_gen(width = 15)) +
  scale_size_discrete(range = c(1,3), name = "Number of observations" ) +
  geom_vline(xintercept = 3.5, color = "darkgrey") +
  geom_vline(xintercept = 2.5, color = "darkgrey") + 
  guides(color = FALSE) +
  scale_colour_manual(values = safe_colorblind_palette) + 
  coord_flip() + theme_bw() +
  labs(title = "Why would you not take the COVID-19 vaccine?",
       x = "") +
  theme_bw() + ylim(c(-2,100)) +
  theme(legend.position = "bottom",
        plot.caption = element_text(hjust = 0), #Default is hjust=1
        plot.title.position = "plot", #NEW parameter. Apply for subtitle too.
        plot.caption.position =  "plot",
        axis.text.y = element_text(hjust = 0))


fig_2

```

### Calculate numbers used in text for refusal

```{r analysis_no}

## Generate data for analysis of no reasons

# Top LMIC countries concerned about side effects
side_top <- 
  no_vacc %>% 
  dplyr::filter(tag == "Concerned about side effects" & group != "All" & 
                  group != "USA" & group != "Russia") %>%
  dplyr::arrange(desc(estimate))

# Concern about side effects in Russia and the US
side_rus <- 
  no_vacc %>% 
  dplyr::filter(tag == "Concerned about side effects" &  group == "Russia") %>% 
  dplyr::arrange(desc(estimate))

side_usa <- 
  no_vacc %>% 
  dplyr::filter(tag == "Concerned about side effects" & group == "USA") %>% 
  dplyr::arrange(desc(estimate))

# Top LMIC counties that are allergic to vaccines
allergies_top <- 
  no_vacc %>% 
  dplyr::filter(tag == "Allergic to vaccines" & group != "All") %>% 
  dplyr::arrange(desc(estimate))

# Top LMIC countries that do not like needles
needles_top <- 
  no_vacc %>% 
  dplyr::filter(tag == "Doesn't like needles" & group != "All" ) %>% 
  dplyr::arrange(desc(estimate))

# Top LMIC countries that wont have time to get vaccinated
time_top <- 
  no_vacc %>% 
  dplyr::filter(tag == "Won't have time to get vaccinated" & group != "All" ) %>% 
  dplyr::arrange(desc(estimate))

# Top LMIC countries that are concerned about getting covid from the vaccine
get_top <- 
  no_vacc %>% 
  dplyr::filter(tag == "Concerned about getting coronavirus from the vaccine" & group != "All" ) %>%
  dplyr::arrange(desc(estimate))

# Top LMIC countries that thinks vaccines are not effective
effective_top <- 
  no_vacc %>% 
  dplyr::filter(tag == "Doesn't think vaccines are effective" & group != "USA" & 
           group != "Russia") %>% 
  dplyr::arrange(desc(estimate))

# Estimate of Doesn't think vaccines are effective in Russia and the USA
effective_rus <- 
  no_vacc %>% 
  dplyr::filter(tag == "Doesn't think vaccines are effective" & group == "Russia") %>% 
  dplyr::arrange(desc(estimate))

effective_usa <- 
  no_vacc %>% 
  dplyr::filter(tag == "Doesn't think vaccines are effective" & group == "USA" ) %>% 
  dplyr::arrange(desc(estimate))

# Top LMIC countries that are not concerned of getting seriously ill
ill_top <- 
  no_vacc %>% 
  dplyr::filter(tag == "Not concerned about getting seriously ill" & group != "All" ) %>% 
  dplyr::arrange(desc(estimate))

# Top LMIC countries that do not think covid outbreak is as serious as people say
serious_top <- 
  no_vacc %>% 
  dplyr::filter(tag == "Doesn't think Coronavirus outbreak is as serious as people say" & 
                  group != "All") %>% 
  dplyr::arrange(desc(estimate))

```

### Table version

```{r tabfig2}

no_vacc2 <- 
  no_vacc %>%
  dplyr::mutate(estimate = format(estimate, nsmall = 0),
                conf_int = paste0("(", format(conf.low, nsmall = 0), 
                                  ", ", format(conf.high, nsmall = 0), ")")) %>%
  dplyr::select(group, estimate, conf_int, outcome, n) %>%
  tidyr::pivot_wider(names_from = outcome, 
                     values_from = c(estimate, conf_int, n), 
                     names_sep = "__") %>%
  tidyr::pivot_longer(cols = c(starts_with("estimate__"), starts_with("conf_int__")),
                      names_to = c("type", ".value"),
                      names_pattern = "(.*)__(.*)") %>% 
  dplyr::rowwise() %>% 
  dplyr::mutate(
    n = ifelse(group == "All", NA, unique(na.omit(c_across(starts_with("n__")))))) %>% 
  dplyr::ungroup() %>% 
  dplyr::mutate(
    group = forcats::fct_relevel(forcats::fct_rev(group), "All", "Russia", "USA", after = Inf)) %>% 
  dplyr::arrange(group) %>% 
  dplyr::mutate(across(c(group, n), ~ifelse(type == "conf_int", "", as.character(.))),
                group = ifelse(group == "All", "All LMICs", group)) %>% 
  dplyr::select(group, n, starts_with("no_vaccine_"), -starts_with("n__no_vaccine"), -type) %>% 
  dplyr::relocate("no_vaccine_666", .after = last_col())


tab_fig2 <- 
  no_vacc2 %>%
  knitr::kable(
    col.names = 
      c("Study", "N", 
        "Concerned about side effects", 
        "Concerned about getting coronavirus from the vaccine", 
        "Not concerned about getting seriously ill", 
        "Doesn't think vaccines are effective", 
        "Doesn't think Coronavirus outbreak is as serious as people say", 
        "Doesn't like needles", 
        "Allergic to vaccines", 
        "Won't have time to get vaccinated", 
        "Mentions a conspiracy theory", 
        "Other reasons"),
    caption = "Reasons not to take the vaccine",
    align = c("l", rep("c", 11)),
    format = "latex", booktabs = T, linesep = "", 
    format.args = list(big.mark = ",", scientific = FALSE), 
    label = "no") %>%
  kableExtra::kable_styling(latex_options = c("scale_down", "hold_position"),
                            font_size = base_font_size - 2, full_width = FALSE) %>%
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::column_spec(1:12, width = "7em") %>%
  # kableExtra::column_spec(1, width = "7em") %>%
  kableExtra::footnote(
    general_title = "",
    general = "Table S5 shows percentage of respondents mentioning reasons why they would not take the Covid-19 vaccine. The number of observations and percentage correponds only to people who would NOT take the vaccine. Respondents in all countries could give more than one reason. A 95% confidence interval is shown between parentheses.",
    threeparttable = T)  %>%
  kableExtra::landscape()

no_vacc2 %>%
  knitr::kable(
    col.names = c("Study", "N", 
                  "Concerned about side effects", 
                  "Concerned about getting coronavirus from the vaccine", 
                  "Not concerned about getting seriously ill", 
                  "Doesn't think vaccines are effective", 
                  "Doesn't think Coronavirus outbreak is as serious as people say", 
                  "Doesn't like needles", 
                  "Allergic to vaccines", 
                  "Won't have time to get vaccinated", 
                  "Mentions a conspiracy theory", 
                  "Other reasons"),
    caption = "\\label{no}Reasons not to take the vaccine", 
    booktabs = T, linesep = "", 
    format.args = list(big.mark = ",",scientific = FALSE),
    align = c("l", rep("c", 11))) %>%
  kableExtra::kable_styling(full_width = F) %>%
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::column_spec(1:12, width = "7em") %>%
  # kableExtra::column_spec(1, width = "7em") %>%
  kableExtra::footnote(
    general_title = "",
    general = "Table S5 shows percentage of respondents mentioning reasons why they would not take the Covid-19 vaccine. The number of observations and percentage correponds only to people who would NOT take the vaccine. Respondents in all countries could give more than one reason. A 95% confidence interval is shown between parentheses.",
    threeparttable = T)

```

## Trust

### Main figure (Fig 3)

```{r fig3, fig.width = 12,  fig.height = 10}

#Group together categories

df2 %<>% 
  dplyr::mutate(
    trust_recode_1 = ifelse(trust_vaccine_1 == 1 | trust_vaccine_2 == 1, 1, 0),
    trust_recode_1 = ifelse((country == "Nigeria" | country == "USA") &
                              is.na(trust_recode_1), 0, trust_recode_1),
    trust_recode_2 = ifelse(trust_vaccine_8 == 1 | trust_vaccine_9 == 1, 1, 0),
    trust_recode_2 = ifelse((country == "Sierra Leone 2") & 
                              is.na(trust_recode_2), 0, trust_recode_2),
    trust_recode_3 = ifelse(trust_vaccine_3 == 1 | 
                              trust_vaccine_7 == 1 | 
                              trust_vaccine_4 == 1, 1, 0),
    trust_recode_3 = 
      ifelse((country == "Nigeria" | country == "USA" | country == "Russia") & 
               is.na(trust_recode_3), 0, trust_recode_3),
    trust_recode_4 = ifelse(trust_vaccine_666 == 1 | trust_vaccine_other == 1, 1, 0),
    trust_recode_4 = 
      ifelse((country == "Burkina Faso" | 
                country == "Sierra Leone 2" | 
                country == "Russia") & 
               is.na(trust_recode_4), 0, trust_recode_4),
    trust_recode_5 = ifelse(trust_vaccine_dk == 1 | 
                              trust_vaccine_refuse == 1 | 
                              trust_vaccine_nr == 1, 1, 0),
    trust_recode_5 = 
      ifelse((country == "Nigeria" | country == "Sierra Leone 2" | country == "USA") &
               is.na(trust_recode_5), 0, trust_recode_5))

#Recoded groups
trust_names <- c("trust_recode_1", "trust_recode_2", "trust_recode_3", 
           "trust_recode_4", "trust_recode_5", "trust_vaccine_5", "trust_vaccine_6")

studies_levels <- 
  c("Burkina Faso", "Colombia", "India", "Mozambique",
    "Nepal", "Nigeria", "Pakistan 1", "Rwanda",
    "Sierra Leone 1", "Sierra Leone 2", "Uganda 2",
    "All", "Russia", "USA" )

#Get estimates
trust_vacc_together <-
  list(
    All = lapply(trust_names, reasons_together, 
                 df = df2, num = c("Yes", "No", "DK")) %>% 
      dplyr::bind_rows(),
    Yes = lapply(trust_names, reasons_together, 
                 df = df2, num = c("Yes")) %>% 
      dplyr::bind_rows(),
    No = lapply(trust_names, reasons_together, 
                df = df2, num = c("No", "DK")) %>% 
      dplyr::bind_rows()) %>% 
  dplyr::bind_rows(.id = "sub") %>%
  dplyr::filter(!is.nan(statistic)) %>%
  dplyr::mutate(
    across(c(conf.low, conf.high, estimate), ~ round(. * 100, digits = 1)),
    n_sub = round(n * estimate, 0),
    n_sub = ifelse(n_sub == 0, NA_integer_, n_sub),
    group = factor(group, levels = studies_levels)) %>%
  dplyr::left_join(dictionary, by = "outcome") %>%
  dplyr::mutate(
    size = cut(n_sub, c(0, 50, 500, Inf), include.lowest = TRUE),
    size = forcats::fct_recode(size, "500+" = "(500,Inf]"),
    tag = as.factor(tag),
    tag = forcats::fct_relevel(tag, 
                               "Health workers", 
                               "Government or MoH", 
                               "Family or Friends", 
                               "Famous person, religious leader or traditional healers", 
                               "Newspapers, radio or online groups", 
                               "Other", 
                               "Don't know or Refuse"),
    sub = forcats::fct_relevel(as.factor(sub), "No", "Yes", "All"),
    sub = plyr::mapvalues(sub, from = c("No", "Yes", "All"),
                          to = c("No, Don't know", "Yes", "Any")))

#Plot
fig_hist2 <- 
  trust_vacc_together %>%
  dplyr::mutate(group = plyr::mapvalues(group, "All", "All LMICs")) %>% 
  dplyr::filter(sub == "Any") %>% 
  ggplot(aes(estimate, tag)) + 
  geom_bar(stat = "identity", position = "dodge", fill = "#DDCC77") + 
  facet_wrap(~group, ncol = 2, strip.position = "left")  +
  coord_flip() +
  scale_fill_manual(
    name = "Answer", 
    values = safe_colorblind_palette[c(1,3,2)]) + 
  scale_y_discrete(labels = function(x) stringr::str_wrap(x, width = 16), 
                   guide = guide_axis(angle = 90)) +
  labs(title = "Which of the following people would you trust MOST to help you decide whether you would get a COVID-19 vaccine?",
       y = "") +
  theme_bw() + 
  theme(legend.position = "bottom",
        plot.title.position = "plot", #NEW parameter. Apply for subtitle too.
        axis.text.y = element_text(hjust = 0))

fig_hist2

```

```{r fig3quantities, fig.width = 12,  fig.height = 10}

trust <- filter(trust_vacc_together, sub == "Any")

# All LMIcs estimate of trust HW

trust_all <- 
  trust %>% 
  dplyr::filter(group == "All" & tag == "Health workers") %>%
  dplyr::arrange(desc(estimate))

# Top and bottom LMICs countries for trust in Health works
trust_ <- 
  trust %>% 
  dplyr::filter(group != "All" & group != "USA" & 
                  group != "Russia" & tag == "Health workers") 

trust_top <- trust_ %>% dplyr::arrange(desc(estimate))

trust_rwa <- 
  trust %>% 
  dplyr::filter(group == "Rwanda") %>%
  dplyr::arrange(desc(estimate))

trust_npl <- 
  trust %>% 
  dplyr::filter(
    group == "Nepal" & 
      tag == "Famous person, religious leader or traditional healers") %>%
  dplyr::arrange(desc(estimate))

# Top and bottom LMICs countries for trust in Family and friends
trust_fam <- 
  trust %>% 
  dplyr::filter(group != "All" & tag == "Family or Friends") %>%
  dplyr::arrange(desc(estimate))

# Top and bottom LMICs countries for trust in Gov
trust_gov <- 
  trust %>% 
  dplyr::filter(group != "All" & group != "Rwanda" & tag == "Government or MoH") %>%
  dplyr::arrange(desc(estimate))

```

### ...by acceptance

```{r fig3categories, fig.width = 12,  fig.height = 10}

#Plot
fig_hist_categories <- 
  trust_vacc_together %>%
  dplyr::mutate(group = plyr::mapvalues(group, "All", "All LMICs")) %>% 
  ggplot(aes(estimate, tag, fill = sub)) + 
  geom_bar(stat = "identity", position = "dodge") + 
  facet_wrap(~group, ncol = 2, strip.position = "left")  +
  coord_flip() +
  scale_fill_manual(
    name = "Answer", 
    values = safe_colorblind_palette[c(1,3,2)]) + 
  scale_y_discrete(labels = function(x) stringr::str_wrap(x, width = 16), 
                   guide = guide_axis(angle = 90)) +
  labs(title = "Which of the following people would you trust MOST to help you decide whether you would get a COVID-19 vaccine?",
       y = "") +
  theme_bw() + 
  theme(legend.position = "bottom",
        plot.title.position = "plot", #NEW parameter. Apply for subtitle too.
        axis.text.y = element_text(hjust = 0))

fig_hist_categories

```


### ...by gender

```{r fig3gender, fig.width = 12,  fig.height = 10}

trust_vacc_gender <-  
  list(
    All = lapply(trust_names, reasons_together_subgroup, 
                 df = df2, 
                 num = c("Yes", "No", "DK"), 
                 dem_group = "gender", 
                 dem_subgroup = c("Female", "Male")) %>% dplyr::bind_rows(),
    Male = lapply(trust_names, reasons_together_subgroup, 
                  df = df2, 
                  num = c("Yes", "No", "DK"), 
                  dem_group = "gender", 
                  dem_subgroup = "Male") %>%
      dplyr::bind_rows(),
  Female = lapply(trust_names, reasons_together_subgroup, 
                  df = df2, 
                  num = c("Yes", "No", "DK"), 
                  dem_group = "gender", 
                  dem_subgroup = "Female") %>% 
  dplyr::bind_rows()) %>% 
  dplyr::bind_rows(.id = "sub") %>%
  dplyr::filter(!is.nan(statistic)) %>%
  dplyr::mutate(
    across(c(conf.low, conf.high, estimate), ~ round(. * 100, digits = 1)),
    n_sub = round(n * estimate, 0),
    n_sub = ifelse(n_sub == 0, NA_integer_, n_sub),
    group = factor(group, levels = studies_levels)) %>%
  dplyr::left_join(dictionary, by = "outcome") %>%
  dplyr::mutate(
    size = cut(n_sub, c(0, 50, 500, Inf), include.lowest = TRUE),
    size = forcats::fct_recode(size, "500+" = "(500,Inf]"),
    tag = as.factor(tag),
    tag = forcats::fct_relevel(tag, 
                               "Health workers", 
                               "Government or MoH", 
                               "Family or Friends", 
                               "Famous person, religious leader or traditional healers", 
                               "Newspapers, radio or online groups", 
                               "Other", 
                               "Don't know or Refuse"),
    sub = forcats::fct_relevel(as.factor(sub), "Female", "Male", "All"))

#Plot
hist_gender <- 
  trust_vacc_gender %>%
  dplyr::mutate(group = plyr::mapvalues(group, "All", "All LMICs")) %>% 
  ggplot(aes(estimate, tag, fill = sub)) + 
  geom_bar(stat = "identity", position = "dodge") + 
  facet_wrap(~group, ncol = 2, strip.position = "left")  +
  coord_flip() +
  scale_fill_manual(
    name = "Answer", 
    values = safe_colorblind_palette[c(1,3,2)]) + 
  scale_y_discrete(labels = function(x) stringr::str_wrap(x, width = 16), 
                   guide = guide_axis(angle = 90)) +
  labs(title = "Which of the following people would you trust MOST to help you decide whether you would get a COVID-19 vaccine?",
       y = "") +
  theme_bw() + 
  theme(legend.position = "bottom",
        plot.title.position = "plot", #NEW parameter. Apply for subtitle too.
        axis.text.y = element_text(hjust = 0))

hist_gender

```

### Gender difference means

```{r gendif}

differences_means_gen <- 
  lapply(trust_names, function(i) {
    df2 %>% 
      dplyr::filter(group != "All" & group != "USA" & 
                      group != "Russia" & group != "Uganda 1") %>% 
      estimatr::lm_robust(as.formula(paste(i, "~gender")),
                          fixed_effects = ~country,
                          weight = weight,
                          cluster = country,
                          se_type = "stata",
                          data = .) %>% 
      tidy %>% 
      dplyr::select(estimate, std.error, p.value, df, term) %>%
      dplyr::mutate(outcome = paste(i))
  }) %>% 
  dplyr::bind_rows(.) %>%
  dplyr::left_join(dictionary) %>%
  dplyr::select(-outcome)

differences_means_gen %>% 
  dplyr::mutate(adjusted_p = p.adjust(p.value, method = "BH")) %>%
  knitr::kable(digits = 3, caption = "Differences in means trust actors (BH adjustment)")


```

### Trusted Sources Table 

```{r tabfig3}

trust_vacc <- 
  plyr::ldply(
    .data = list("Yes", "No", "All"), 
    .fun = function(take_vac) {
      list(Yes = "Yes", 
           No = c("No", "DK"), 
           All = c("Yes", "No", "DK")) %>% 
        .[[take_vac]] %>% 
        plyr::ldply(trust_names, reasons_together, df = df2, num = .) %>%
        dplyr::mutate(
          across(c(conf.low, conf.high, estimate), 
                 ~ format(round(. * 100, digits = 1), nsmall = 1)),        
          conf_int = paste0("(", conf.low, ", ", conf.high, ")")) %>%
        dplyr::select(group, estimate, conf_int, outcome, n) %>%
        tidyr::pivot_wider(names_from = outcome, values_from = c(estimate, conf_int, n), 
                           names_sep = "__") %>%
        tidyr::pivot_longer(cols = c(starts_with("estimate__"), starts_with("conf_int__")),
                            names_to = c("type", ".value"),
                            names_pattern = "(.*)__(.*)") %>%
        dplyr::rowwise() %>% 
        dplyr::mutate(
          n = ifelse(group == "All", NA, unique(na.omit(c_across(starts_with("n__")))))) %>% 
        dplyr::ungroup() %>% 
        dplyr::mutate("Take vaccine?" = take_vac) %>%
        dplyr::select(group, n, type, "Take vaccine?", starts_with("trust_")) %>%
        dplyr::filter(!(group %in% c("Mozambique", "Pakistan 1", 
                                     "Pakistan 2", "Uganda 1", "India")))
    }) %>% 
  dplyr::mutate(
    group = as.factor(group),
    group = forcats::fct_relevel(group, "All", "Russia", "USA", after = Inf)) %>% 
  dplyr::arrange(group) %>% 
  dplyr::mutate(across(c(group, n, `Take vaccine?`), ~ifelse(type == "conf_int", "", as.character(.))),
                group = ifelse(group == "All", "All LMICs", group)) %>% 
  dplyr::select(-type)

tab_trust <- 
  trust_vacc %>%
  dplyr::select("group", "n", 
                "Take vaccine?", "trust_vaccine_5", 
                "trust_vaccine_6", "trust_recode_1", 
                "trust_recode_3", "trust_recode_2", 
                "trust_recode_4", "trust_recode_5") %>%
  knitr::kable(
    col.names = c("Study", "N", "Take vaccine?", "Health workers", 
                  "Government or Ministry of Health", 
                  "Family or friends", 
                  "Famous person, religious leader or traditional healers", 
                  "Newspapers, radio or online groups", "Other", 
                  "Don't know or Refuse"),
    caption = "COVID-19 Vaccination Decision-making: most trusted source",
    align = c("l", rep("c", 9)),
    format = "latex", booktabs = T, linesep = "", longtable = TRUE,
    format.args = list(big.mark = ",", scientific = FALSE), 
    label = "trust") %>% 
  kableExtra::kable_styling(latex_options = c("scale_down", "hold_position", "repeat_header"),
                            font_size = base_font_size - 2, full_width = FALSE) %>%
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::column_spec(1, width = "7em") %>% 
  kableExtra::column_spec(2:10, width = "4em") %>% 
  kableExtra::column_spec(4:10, width = "6em") %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table S6 shows percentage of respondents that mention actors who they would trust the most to help them decide whether to get a COVID-19 vaccine. For all countries the questions was asked regardless if respondent would take a vaccine, would not take it, does not know or does not respond. For India respondents were able to mention more than one actor, for the rest of countries only one actor was allowed. While rows should sum to 100%, rounding makes number slightly above or below. A 95% confidence interval is shown between parentheses.",
    threeparttable = T) %>%
  kableExtra::landscape()



trust_vacc %>%
  dplyr::select("group", "n", 
                "Take vaccine?", "trust_vaccine_5", 
                "trust_vaccine_6", "trust_recode_1", 
                "trust_recode_3", "trust_recode_2", 
                "trust_recode_4", "trust_recode_5") %>%
  knitr::kable(
    col.names = c("Study", "N", "Take vaccine?", "Health workers", 
                  "Government or \n Ministry of Health", 
                  "Family or friends", 
                  "Famous person, \n religious leader or \n traditional healers", 
                  "Newspapers, radio \n or online groups", "Other", 
                  "Don't know or Refuse"),
    caption = "COVID-19 Vaccination Decision-making: most trusted source", 
    format.args = list(big.mark = ",", scientific = FALSE),
    align = c("l", rep("c", 9))) %>% 
  kableExtra::kable_styling(full_width = FALSE) %>%
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::column_spec(1, width = "7em") %>% 
  kableExtra::column_spec(2:10, width = "4em") %>% 
  kableExtra::column_spec(4:10, width = "6em") %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table S6 shows percentage of respondents that mention actors who they would trust the most to help them decide whether to get a COVID-19 vaccine. For all countries the questions was asked regardless if respondent would take a vaccine, would not take it, does not know or does not respond. For India respondents were able to mention more than one actor, for the rest of countries only one actor was allowed. While rows should sum to 100%, rounding makes number slightly above or below. A 95% confidence interval is shown between parentheses.",
    threeparttable = T)

```


```{r question1, include=FALSE}

tab_question1 <- 
  readxl::read_excel("2_input_data/studies_info.xlsx", sheet = "sample") %>%
  dplyr::select(
    "Study" = "country", 
    "Question Fig. 1", 
    "Recoding Fig. 1") %>%
  knitr::kable(
    caption =  "Question wording and answer options: vaccine acceptance",
    format = "latex", booktabs = T, linesep = "", label = "q1") %>%
  kableExtra::kable_styling(latex_options = c("scale_down", "hold_position"),
                            font_size = base_font_size - 2, full_width = FALSE) %>%
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::column_spec(1, width = "8em") %>%
  kableExtra::column_spec(2, width = "30em")  %>%
  kableExtra::column_spec(3, width = "20em") %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table S10 presents question wording and answer options from answers used in Figure 1 to get estimated vaccine acceptance. Answer options are separated by a semicolon. In India options 'Yes, only for free' and 'Yes, even if I have to pay' are both recoded as 'Yes'. In Pakistan 2, 'Absolutely yes' is recoded as 'Yes', 'Neutral' is recoded as 'Don't know' and 'Absolutely no' is recoded as 'No'. In Russia, 'Yes, if a Russian vaccine will be available' and 'Yes, if an imported vaccine will be available' are both recoded as 'Yes'. In USA 'Definitely yes' and 'Probably yes' are recoded as 'Yes', and 'Probably not' and 'Definitely not' are recoded as 'No'." ,
    threeparttable = T)

readxl::read_excel("2_input_data/studies_info.xlsx", sheet = "sample") %>%
  dplyr::select("Study" = "country", "Question Fig. 1", "Recoding Fig. 1") %>%
  knitr::kable(
    caption =  "Question wording and answer options: vaccine acceptance", format = "html") %>% 
  kableExtra::kable_styling(full_width = FALSE) %>% 
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::column_spec(1, width = "8em") %>%
  kableExtra::column_spec(2, width = "30em")  %>%
  kableExtra::column_spec(3, width = "20em") %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table S10 presents question wording and answer options from answers used in Figure 1 to get estimated vaccine acceptance. Answer options are separated by a semicolon. In India options 'Yes, only for free' and 'Yes, even if I have to pay' are both recoded as 'Yes'. In Pakistan 2, 'Absolutely yes' is recoded as 'Yes', 'Neutral' is recoded as 'Don't know' and 'Absolutely no' is recoded as 'No'. In Russia, 'Yes, if a Russian vaccine will be available' and 'Yes, if an imported vaccine will be available' are both recoded as 'Yes'. In USA 'Definitely yes' and 'Probably yes' are recoded as 'Yes', and 'Probably not' and 'Definitely not' are recoded as 'No'," ,
    threeparttable = T)
```

```{r question2, include=FALSE}

tab_question2 <- 
  readxl::read_excel("2_input_data/studies_info.xlsx", sheet = "sample") %>%
  dplyr::select(
    "Study" = "country",
    "Question Tab. 2",
    "Protection: self", 
    "Protection: family",
    "Protection: community") %>%
  dplyr::filter(!is.na(`Question Tab. 2`)) %>%
  knitr::kable(
    caption =  "Question wording and answer options: reasons to take vaccine",
    format = "latex", booktabs = T, linesep = "", label = "q2") %>% 
  kableExtra::kable_styling(latex_options = c("scale_down", "hold_position"),
                            font_size = base_font_size - 2, full_width = FALSE) %>%
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::column_spec(1, width = "8em") %>%
  kableExtra::column_spec(2, width = "8em")  %>%
  kableExtra::column_spec(3:5, width = "10em") %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table S11 presents question wording and answer options used in Table S2 to get an estimated percentage of reasons to take the COVID-19 vaccine. Columns 'Protection: self', 'Protection: family' and  'Protection: community' show the answer options that were recoded in each category. Answer options are separated by a semicolon." ,
    threeparttable = T)

readxl::read_excel("2_input_data/studies_info.xlsx", sheet = "sample") %>%
  dplyr::select(
    "Study" = "country", 
    "Question Tab. 2", 
    "Protection: self", 
    "Protection: family", 
    "Protection: community") %>%
  knitr::kable(
    caption =  "Question wording and answer options: reasons to take vaccine", format = "html") %>% 
  kableExtra::kable_styling(full_width = FALSE) %>% 
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::column_spec(1, width = "8em") %>%
  kableExtra::column_spec(2, width = "8em")  %>%
  kableExtra::column_spec(3:5, width = "10em") %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table S11 presents question wording and answer options used in Table S2 to get an estimated percentage of reasons to take the COVID-19 vaccine. Columns 'Protection: self', 'Protection: family' and  'Protection: community' show the answer options that were recoded in each category. Answer options are separated by a semicolon." ,
    threeparttable = T)

```

```{r question3, include=FALSE}

tab_question3 <- 
  readxl::read_excel("2_input_data/studies_info.xlsx", sheet = "sample") %>%
  dplyr::select(
    "Study" = "country", 
    "Question Fig. 2", 
    "Concerned about side effects", 
    "Concerned about getting COVID-19 from the vaccine", 
    "Not concerned about getting seriously ill", 
    "Doesn't think vaccines are effective", 
    "Doesn't think COVID-19 outbreak is as serious as people say", 
    "Doesn't like needles", 
    "Allergic to vaccines", 
    "Won't have time to get vaccinated", 
    "Mentions a conspiracy theory",
    "Other reasons") %>%
  dplyr::filter(!is.na(`Question Fig. 2`)) %>%
  knitr::kable(
    caption =  "Question wording and answer options: reasons not to take the vaccine",
    format = "latex", booktabs = T, linesep = "", label = "q3") %>%
  kableExtra::kable_styling(latex_options = c("scale_down", "hold_position"),
                            font_size = base_font_size - 2, full_width = FALSE) %>%
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::column_spec(1:2, width = "5em") %>%
  kableExtra::column_spec(3:12, width = "10em")  %>%
  kableExtra::footnote(
    general_title = "",
    general = "Table S12 presents question wording and answer options used in Figure 2 to get an estimated percentage of reasons not to take the COVID-19 vaccine. Columns 3-10 show the answer options that were recoded in each category. Answer options are separated by a semicolon." ,
    threeparttable = T) %>% 
  kableExtra::landscape()

readxl::read_excel("2_input_data/studies_info.xlsx", sheet = "sample") %>%
  dplyr::select(
    "Study" = "country", 
    "Question Fig. 2", 
    "Concerned about side effects", 
    "Concerned about getting COVID-19 from the vaccine", 
    "Not concerned about getting seriously ill", 
    "Doesn't think vaccines are effective", 
    "Doesn't think COVID-19 outbreak is as serious as people say", 
    "Doesn't like needles", 
    "Allergic to vaccines", 
    "Won't have time to get vaccinated", 
    "Mentions a conspiracy theory",
    "Other reasons") %>%
  dplyr::filter(!is.na(`Question Fig. 2`)) %>%
  knitr::kable(
    caption =  "Question wording and answer options: reasons not to take the vaccine",
    format = "html", linesep = "") %>%
  kableExtra::kable_styling(full_width = FALSE) %>%
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::column_spec(1:2, width = "5em") %>%
  kableExtra::column_spec(3:12, width = "10em")  %>%
  kableExtra::footnote(
    general_title = "",
    general = "Table S12 presents question wording and answer options used in Figure 2 to get an estimated percentage of reasons not to take the COVID-19 vaccine. Columns 3-10 show the answer options that were recoded in each category. Answer options are separated by a semicolon." ,
    threeparttable = T)
```

```{r question4, include=FALSE}

tab_question4 <- 
  readxl::read_excel("2_input_data/studies_info.xlsx", sheet = "sample") %>%
  dplyr::select(
    "Study" = "country", 
    "Question Fig. 3", 
    "Health workers", 
    "Government or Ministry of Health", 
    "Family or friends", 
    "Famous person, religious leader or traditional healers", 
    "Newspapers, radio or online groups", 
    "Other") %>%
  dplyr::filter(!is.na(`Question Fig. 3`)) %>%
  knitr::kable(
    caption =  "Question wording and answer options: trusted actors and institutions",
    format = "latex", booktabs = T, linesep = "", label = "q4") %>% 
  kableExtra::kable_styling(latex_options = c("scale_down", "hold_position"),
                            font_size = base_font_size - 2, full_width = FALSE) %>%
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::column_spec(1, width = "8em") %>%
  kableExtra::column_spec(2, width = "20em")  %>%
  kableExtra::column_spec(3:8, width = "10em") %>% 
  kableExtra::landscape() %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table S13 presents question wording and answer options used in Figure 3 to get the percentage of respondents mentioning each actor or institution that they would trust to decide whether to get the COVID-19 vaccine. Columns 3-8 show the answer options that were recoded in each category. Answer options are separated by a semicolon." ,
    threeparttable = T)

readxl::read_excel("2_input_data/studies_info.xlsx", sheet = "sample") %>%
  dplyr::select(
    "Study" = "country",
    "Question Fig. 3",
    "Health workers",
    "Government or Ministry of Health",
    "Family or friends",
    "Famous person, religious leader or traditional healers",
    "Newspapers, radio or online groups",
    "Other") %>%
  knitr::kable(
    caption =  "Question wording and answer options: trusted actors and institutions", 
    format = "html") %>% 
  kableExtra::kable_styling(full_width = FALSE) %>% 
  kableExtra::row_spec(0, bold = TRUE) %>% 
  kableExtra::column_spec(1, width = "8em") %>%
  kableExtra::column_spec(2, width = "20em")  %>%
  kableExtra::column_spec(3:8, width = "10em") %>% 
  kableExtra::landscape() %>% 
  kableExtra::footnote(
    general_title = "",
    general = "Table S13 presents question wording and answer options used in Figure 3 to get the percentage of respondents mentioning each actor or institution that they would trust to decide whether to get the COVID-19 vaccine. Columns 3-8 show the answer options that were recoded in each category. Answer options are separated by a semicolon." ,
    threeparttable = T)

```


## Summary stats

```{r tab5, fig.width = 12}
# Summary statistics table
# transform the categorical variables into dummy variables
df2 %<>% 
  fastDummies::dummy_cols(select_columns = c("age_groups","educ_binary","gender"))

get_stat <- function(.var, .data, ...) {
  return(
    paste0("`", .var, "` ~ 1") %>% 
      as.formula() %>% 
      lm_robust(formula = ., data = .data, ...) %>% 
      coef()
  )
}

data_sumstat <- 
  df2 %>% 
  dplyr::nest_by(group) %>% 
  dplyr::mutate(
    Female = get_stat("gender_Female", .data = data),
    age_18_30 = get_stat("age_groups_[18,30)", data),
    age_30_45 = get_stat("age_groups_[30,45)", data),
    age_45_60 = get_stat("age_groups_[45,60)", data),
    age_60    = get_stat("age_groups_[60, Inf)", data),
    Less_than_secondary = get_stat("educ_binary_Up to Secondary", data),
    More_than_secondary = get_stat("educ_binary_> Secondary", data)
    ) %>% 
  dplyr::select(-data, country = group) %>% 
  dplyr::ungroup() %>% 
  dplyr::mutate(across(where(is.double), ~ . * 100)) 


#WGM data
wgmdata <- 
  readr::read_csv("2_input_data/wgm_2018_publiccsv.csv") %>% 
  dplyr::filter(WP5 %in% c(1,9,31,35,41,63,65,76,78,80,105,157),
                Age >= 18) %>% 
  dplyr::mutate(
    country = 
      plyr::mapvalues(WP5, 
                      from = c(1, 9, 31, 35, 41, 63, 65, 76, 78, 80, 105, 157),
                      to = c("USA","Pakistan","India","Nigeria","Uganda",
                             "Mozambique","Rwanda","Russia", 
                             "Burkina Faso", "Sierra Leone","Colombia","Nepal")),
    age_groups = cut(x = Age, breaks = c(-Inf, 18, 30, 45, 60, +Inf), right = F)) %>% 
  dplyr::select(country, wgt, gender = Gender, age = Age, educ = Education,
                age_groups) %>%
  fastDummies::dummy_cols(select_columns = "age_groups") %>% 
  dplyr::mutate(
    gender_Female = if_else(gender == 2, 1, 0), 
    `educ_binary_Up to Secondary` = if_else(educ == 1 | educ == 2, 1, 0),
    `educ_binary_> Secondary` = if_else(educ == 3, 1, 0))


wgmdata_sumstat <- 
  wgmdata %>% 
  dplyr::nest_by(country) %>% 
  dplyr::mutate(
    Female = get_stat("gender_Female", data, weight = wgt),
    age_18_30 = get_stat("age_groups_[18,30)", data, weight = wgt),
    age_30_45 = get_stat("age_groups_[30,45)", data, weight = wgt),
    age_45_60 = get_stat("age_groups_[45,60)", data, weight = wgt),
    age_60    = get_stat("age_groups_[60, Inf)", data, weight = wgt),
    Less_than_secondary = get_stat("educ_binary_Up to Secondary", data, weight = wgt),
    More_than_secondary = get_stat("educ_binary_> Secondary", data, weight = wgt)
    ) %>% 
  dplyr::select(-data) %>% 
  dplyr::ungroup() %>% 
  dplyr::mutate(across(where(is.double), ~ . * 100)) 

sum_stat_col <- 
  dplyr::left_join(data_sumstat, 
                   wgmdata_sumstat, 
                   by = "country", all = TRUE, suffix = c("", "_wgm")) %>% 
  dplyr::select(
    country,
    starts_with("Female"),
    starts_with("age_18_30"),
    starts_with("age_30_45"),
    starts_with("age_45_60"),
    starts_with("age_60"),
    starts_with("Less_than_secondary"),
    starts_with("More_than_secondary")
  )


tab_sum_col <-
  knitr::kable(
    sum_stat_col, 
    caption = "Summary statistics for gender, age, education",
    col.names = c("Study", rep(c("COVID-19 Study", "Population"), 7)),
    digits = 1,
    format = "latex", 
    align = c("l", rep("c", 14)), 
    booktabs = T, linesep = "") %>%
  kableExtra::kable_styling(full_width = FALSE) %>% 
  kableExtra::add_header_above(
    c(" " = 1, 
      "% Women" = 2, 
      "% Age in [18,30)" = 2, 
      "% Age in [30,45)" = 2, 
      "% Age in [45,60)" = 2,
      "% Age 60+" = 2,
      "% Up to Secondary" = 2, 
      "% More than Secondary" = 2)) %>%
  kableExtra::column_spec(2:16, width = "3em") %>% 
  kableExtra::column_spec(1, width = "5em") %>% 
  kableExtra::footnote(
    general = "This table presents summarys statistics for our data and compares it with estimates from other sources of data. For each category, the left column and the right column correspond respectively to the statistics computed from our sample and from previous surveys. Data for Russia comes from census data from the Statistical Agency. For the USA, we use data from the 2019 American Community Survey. For all other countries, the Wellcome Global Monitor 2018 was used. Statistics for our surveys are not weighted, while estimates from benchmark sources are obtained using sampling weights.",
    threeparttable = T) %>% 
  kableExtra::landscape() 

wgmdata_sumstat_label <-
  wgmdata_sumstat %>% 
  dplyr::mutate(country = paste0(country, " (WGM)"))

sum_stat_row <- 
  dplyr::bind_rows(data_sumstat, wgmdata_sumstat_label) %>%
  arrange(country) %>% 
  .[c(1:13,15,16,14,19,20,22,23,21,25,26,24,17,18,27,28),]

tab_sum_row <- 
  knitr::kable(
    sum_stat_row,caption = "Summary statistics for gender, age, education",
    col.names = c("Study", "% Women",
                  "% Age in [18,30)", "% Age in [30,45)","% Age in [45,60)","% Age 60+",
                  "% Up to Secondary", "% More than Secondary"),
    format = "latex", booktabs = T, linesep = "", align = c("l", rep("c", 7)), digits = 1) %>%
  kableExtra::kable_styling(full_width = FALSE) %>% 
  kableExtra::footnote(
    general = "This table presents summarys statistics for our data and compares it with estimates from other sources of data. Data for Russia comes from census data from the Statistical Agency. For the USA, we use data from the 2019 American Community Survey. For all other countries, the Wellcome Global Monitor 2018 was used. Statistics for our surveys are not weighted, while estimates from benchmark sources are obtained using sampling weights.",
    threeparttable = T) %>%
  kableExtra::column_spec(1:8, width = "5em") 

knitr::kable(
  sum_stat_row,
  caption = "Summary statistics for gender, age, education",
  col.names = c("Study", "% Women", 
                "% Age in [18,30)", "% Age in [30,45)", "% Age in [45,60)", "% Age 60+",
                "% Up to Secondary", "% More than Secondary"),  
  booktabs = T, linesep = "", align = c("l", rep("c", 7)), digits = 1) %>%
  kableExtra::kable_styling(full_width = FALSE) %>% 
  kableExtra::footnote(
    general = "This table presents summarys statistics for our data and compares it with estimates from other sources of data. Data for Russia comes from census data from the Statistical Agency. For the USA, we use data from the 2019 American Community Survey. For all other countries, the Wellcome Global Monitor 2018 was used. Statistics for our surveys are not weighted, while estimates from benchmark sources are obtained using sampling weights.") %>%
  kableExtra::column_spec(1:8, width = "5em") 

```
