==> 3_main_text.Rmd <==
<!-- COVID-19 Vaccine Acceptance and Hesitancy in Low and Middle Income Countries, and Implications for Messaging -->


Julio S. Solis Arce,^[WZB Berlin Social Science Center, $^2$Innovations for Poverty Action (IPA), $^3$International Growth Centre (IGC), $^4$Wageningen University & Research, $^5$International Center for the Study of Institutions and Development (HSE University, Moscow, Russia), $^6$Columbia University, $^7$Yale Institute for Global Health, $^8$Busara Center for Behavioral Economics, $^{10}$Department of Sociology, University of Lagos, $^{11}$Busara Nigeria, $^{12}$Agricultural and Rural Development Secretariat, Federal Capital Territory Administration (Abuja, Nigeria), $^{13}$Nova School of Business and Economics, $^{14}$The Institute for Fiscal Studies, $^{15}$Lahore University of Management Sciences, $^{16}$Innovations for Poverty Action (IPA) Uganda, $^{17}$Morsel Research & Development, $^{18}$University of St. Andrews,  $^{19}$Redes Peru, $^{20}$Stockholm School of Economics and Misum, $^{21}$Ghent University, Department of Economics, $^{22}$Innovations for Poverty Action (IPA) Colombia, $^{23}$Institute of Development and Economic Alternatives, $^{24}$Innovations for Poverty Action (IPA) Sierra Leone,$^{25}$NOVAFRICA, $^{26}$Trinity College Dublin, $^{27}$Institute of Public Administration and Management, University of Sierra Leone, $^{28}$Centre for the Study of Labour and Mobility (CESLAM), $^{29}$Cornell University, $^{30}$University of Illinois Chicago, $^{31}$Innovations for Poverty Action (IPA) Rwanda, $^{32}$Associação NOVAFRICA para o Desenvolvimento Empresarial e Económico de Moçambique, $^{33}$Innovations for Poverty Action (IPA) Burkina Faso, $^{34}$NYU Abu Dhabi, $^{35}$Centre for Economic Research in Pakistan (CERP), $^{36}$Yale Research Initiative on Innovation and Scale (Y-RISE), $^{37}$Princeton University, $^{38}$Institute for International Economic Studies (IIES), Stockholm University, $^{39}$Tufts University, $^{40}$University of Michigan, $^{41}$Kellogg School of Management at Northwestern University, $^{42}$London School of Economics, $^{43}$Yale University] Shana S. Warren,$^2$ Niccolò F. Meriggi,$^3$ Alexandra Scacco,$^1$ Nina McMurry,$^1$ Maarten Voors,$^4$ Georgiy Syunyaev,$^{1,5,6}$, Amyn Abdul Malik,$^7$ Samya Aboutajdine,$^3$ Opeyemi Adeojo,$^{8, 10}$ Deborah Anigo,$^{11, 12}$ Alex Armand,$^{13, 14}$ Saher Asad,$^{15}$ Martin Atyera,$^{16}$ Britta Augsburg,$^{14}$ Manisha Awasthi,$^{17}$ Gloria Eden Ayesiga,$^{16}$ Antonella Bancalari,$^{14, 18, 19}$ Martina Björkman Nyqvist,$^{20}$ Ekaterina Borisova,$^{5, 21}$ Constantin Manuel Bosancianu,$^1$ Magarita Rosa Cabra García,$^22$ Ali Cheema,$^{15, 23}$ Elliott Collins,$^{2}$ Filippo Cuccaro,$^{24}$ Ahsan Zia Farooqi,$^{23}$ Tatheer Fatima,$^{17}$ Mattia Fracchia,$^{13, 25}$ Mery Len Galindo Soria,$^{22}$ Andrea Guariso,$^{26}$ Ali Hasanain,$^{15}$ Sofía Jaramillo Arciniegas,$^{22}$ Sellu Kallon,$^{4, 27}$ Anthony Kamwesigye,$^{16}$ Arjun Kharel,$^{28}$ Sarah Kreps,$^{29}$ Madison Levine,$^{4}$ Rebecca Littman,$^{30}$ Shan-e-Mohammad Malik,$^{23}$ Gisele Manirabaruta,$^{31}$ Jean Léodomir Habarimana Mfura,$^{31}$ Fatoma Momoh,$^{24}$ Alberto Mucauque,$^{32}$ Imamo Mussa,$^{32}$ Jean Aime Nsabimana,$^{31}$ Isaac Obara,$^{8}$ María Juliana Otálora,$^{22}$ Béchir Wendemi Ouédraogo,$^{33}$ Touba Bakary Pare,$^{33}$ Melina R. Platas,$^{34}$ Laura Polanco,$^{22}$ Javaeria Ashraf Qureshi,$^{30}$ Mariam Raheem,$^{35}$ Vasudha Ramakrishna,$^{36}$ Ismail Rendrá,$^{32}$ Taimur Shah,$^{35}$ Sarene Eyla Shaked,$^{16}$ Jacob N. Shapiro, $^{37}$ Jakob Svensson,$^{38}$ Ahsan Tariq,$^{23}$ Achille Mignondo Tchibozo,$^{33}$ Hamid Ali Tiwana,$^{23}$ Bhartendu Trivedi,$^{17}$ Corey Vernot, $^{36}$ Pedro C. Vicente,$^{13, 25}$ Laurin B Weissinger,$^{39}$ Basit Zafar,$^{35, 40}$ Baobao Zhang,$^{29}$ Dean Karlan,$^{2, 41}$ Michael Callen,$^{42}$ Matthieu Teachout,$^{3}$ Macartan Humphreys,$^{1, 6}$ Ahmed Mushfiq Mobarak,$^{43}$ Saad B. Omer$^{7}$.


\newpage


```{r} 
empty = " " 
```
`r empty`


# Abstract {-}


We analyze COVID-19 vaccine acceptance across 15 survey samples covering ten low- and middle-income countries (LMICs) in Asia, Africa, and South America, Russia (an upper-middle-income country), and the United States, including  `r nice(nrow(df))` individuals. We find considerably higher willingness to take a COVID-19 vaccine in our LMIC samples (mean `r ans_mean`%; median `r ans_median`%; range of `r ans_max - ans_min`%) compared to the United States (mean `r usa_ans %>% pull(estimate)`%) and Russia (mean `r rus_ans %>% pull(estimate)`%). Vaccine acceptance is primarily explained by an interest in personal protection against COVID-19, while concern about side effects is the most common reason for hesitancy. Health workers are the most trusted sources of guidance about COVID-19 vaccines. Evidence from this sample of LMICs suggests that prioritizing vaccine distribution to the Global South should yield high returns in advancing global immunization coverage. Vaccination campaigns should focus on translating the high levels of stated acceptance into actual uptake. Messages highlighting vaccine efficacy and safety, delivered by healthcare workers, may be most effective for addressing remaining hesitancy.



\newpage


A safe and effective vaccine is a critical tool to control the COVID-19 pandemic. As of May 27, 2021, 23 vaccines had advanced to Stage 3 clinical trials [@who1] and more than a dozen had been approved in multiple countries[@track]. The BioNTech/Pfizer (Comirnaty) vaccine, for example, has been approved in more than 80 countries, while the AstraZeneca vaccine has the most country authorizations at 98 [@track]. At present, however, global vaccine distribution remains highly unequal, with much of the current supply directed toward high-income countries [@wouters2021challenges].


While effective and equitable distribution of COVID-19 vaccines is a key policy priority, ensuring acceptance is just as important. Trust in vaccines and the institutions that administer them are key determinants of the success of any vaccination campaign [@defigueiredo2020lancet]. Several studies have investigated willingness to take a potential COVID-19 vaccine in high-income countries [@boyon2020ipsos; @Malik2020; @kreps2020factors; @shekhar2021covid; @dror2020vaccine; @fisher2020attitudes], and some studies have included middle-income countries [@lazarus2020nature; @wouters2021challenges]. Less is known, however, about vaccine acceptance in low-income countries where large-scale vaccination has yet to begin. Understanding the drivers of COVID-19 vaccine acceptance is of global concern, since a lag in vaccination in any country may result in the emergence and spread of new variants that can overcome immunity conferred by vaccines and prior disease [@ong2021lack; @lancettask]. 


Our study complements the emerging global picture of COVID-19 vaccine acceptance by focusing primarily on lower income countries. We construct a sample of low- and middle-income countries (LMICs) with wide geographic coverage across Africa, Asia and Latin America. We move beyond documenting vaccine acceptance rates to collect and analyze data on the reasons for acceptance and hesitancy, which is critical for informing the design of effective vaccine distribution and messaging. A summary of the main findings, limitations and implications of the study is shown in Table \ref{tab:summary}.


Acceptance of childhood vaccination for common diseases---such as measles (MCV), Bacille Calmette-Guérin (BCG) and diphtheria, tetanus, and pertussis (DTP)---is generally high in LMICs, providing grounds for optimism about the prospects for COVID-19 vaccine uptake. Table \ref{tab:otherv} summarizes general vaccine acceptance and coverage rates of childhood vaccines in 2018, prior to the current pandemic, for the countries included in our study. Agreement on the importance of childhood vaccinations is markedly higher in the LMICs we study compared to Russia and the United States. Still, existing studies on COVID-19 vaccine acceptance document substantial variation, both across and within countries, including in settings with high acceptance of other vaccinations [@defigueiredo2020lancet; @lazarus2020nature; @wouters2021challenges]. 


Existing literature cites concern about COVID-19 vaccine safety, including the rapid pace of vaccine development, as a primary reason for hesitancy in higher-income settings [@boyon2020ipsos; @wouters2021challenges]. Other reasons may feature more prominently in LMICs. For instance, reported COVID-19 mortality rates have been consistently lower in most LMICs relative to higher income countries [@biccard2021patient; @rice2021variation; @maeda2021puzzle]. If individuals feel the risk of disease is less severe, they may be less willing to accept any perceived risks of vaccination [@brewer2007meta]. Previous studies of healthcare utilization in LMICs have also highlighted factors such as negative perceptions of healthcare quality [@christensen2020building], negative historical experiences involving foreign actors  [@Lowes2018; @martinez2021vaccines], weak support from traditional leaders [@Jegede2007], and mistrust in government [@BLAIR201789] as barriers to uptake, which could apply to COVID-19 vaccination as well. 


To promote vaccination against COVID-19, we need to know whether people are willing to take COVID-19 vaccines, the reasons why they are willing or unwilling to do so, and the most trusted sources of information in their decision-making. Our study investigates these questions using a common set of survey items deployed across 13 studies in Africa, South Asia, and Latin America: seven in low-income countries (Burkina Faso, Mozambique, Rwanda, Sierra Leone, Uganda), five in lower-middle-income countries (India, Nepal, Nigeria, Pakistan), and one in an upper-middle-income country (Colombia). We compare these findings to those from two countries at the forefront of vaccine research and development, Russia and the United States. 


To select studies to include in our sample, we conducted an internal search within Innovations for Poverty Action (IPA), the International Growth Center (IGC), and the Berlin Social Science Center (WZB) for projects with plans to collect survey data in the second half of 2020. Study PIs agreed to include a set of common questions about COVID-19 vaccine attitudes. This strategy was guided by the need to collect information quickly and cost-effectively using a survey modality (phone) that was both safe, given pandemic conditions, and appropriate for contexts with limited internet coverage. The final set of samples included in our study therefore reflects populations that fall under the current research priorities at IGC, IPA and WZB and, in the case of IPA and IGC, donors that prioritize working in the Global South. 




# Results {-}


Our main results are shown in Figure \@ref(fig:mainfigure1) and reproduced as \hyperref[tab:maintabledis]{Supplementary Table 1}. The first column provides overall acceptance rates in each study, while the remaining columns disaggregate acceptance by respondent characteristics. The “All LMICs” row reports averages for the LMIC samples included in our study and excludes Russia and the USA. The “LMICs National Samples” row reports averages for just the LMIC samples with national-level geographic coverage.


The average acceptance rate across the full set of LMIC studies is `r ans_mean`% (95% CI `r ans_mean_low`--`r ans_mean_high`%), with a median of `r ans_median`%, a range of `r ans_max - ans_min`% and an interquartile range of `r ans_iqr`%. Our estimate of the between-study standard deviation, $\tau$, using a random effects meta-analysis model is `r  round(sqrt(mp_tau), 3)` which represents only `r  round(100*mp_ratio, 1)`% of our estimate of the average acceptance across LMIC studies. 


The acceptance rate in every LMIC sample is higher than in the USA (`r usa_ans %>% pull(estimate)`%, CI `r usa_ans %>% pull(conf.low)`--`r usa_ans %>% pull(conf.high)`%) and Russia (`r rus_ans %>% pull(estimate)`%, CI `r rus_ans %>% pull(conf.low)`--`r rus_ans %>% pull(conf.high)`%). Reported acceptance is lowest in `r bottom_all$group[1]` (`r bottom_all$estimate[1]`%, CI `r bottom_all$conf.low[1]`--`r bottom_all$conf.high[1]`%) and `r bottom_all$group[2]`  (`r bottom_all$estimate[2]`%, CI `r bottom_all$conf.low[2]`--`r bottom_all$conf.high[2]`%). Pakistan’s relatively low acceptance rate may be linked to negative historical experiences with foreign-led vaccination campaigns [@ali2019polio; @robbins2012cia; @martinez2021vaccines]. This hesitancy may be particularly problematic given the magnitude of the second wave in neighboring India and acceleration of cases across South Asia that threaten to overwhelm health infrastructure. The relatively low acceptance rate in Burkina Faso may reflect general vaccine hesitancy. As shown in Table \ref{tab:otherv}, fewer people believe that vaccines in general are safe in Burkina Faso than in any other country included in our study except Russia. 


We find limited evidence of variation across demographic subgroups in our aggregate analysis of LMIC samples, as shown in \hyperref[tab:dmeans]{Supplementary Table 7}. Women are generally less willing to accept the vaccine than men (average difference about `r dif_gender` points, significant at $p<.01$). Respondents under age 25 and less educated respondents are marginally more willing to take the vaccine compared to older and more educated respondents, respectively, but these differences are not statistically significant. 


Table \hyperref[tab:countrydiff]{Supplementary Table 9} provides results disaggregated by demographic subgroups for individual studies. The average gender differences in the aggregate LMIC analysis are driven by the Burkina Faso, Mozambique, Pakistan 1, Rwanda and Sierra Leone 1 samples. However, these gender differences in acceptance are less than 10 percentage points in each of these samples, in contrast to larger gender gaps in acceptance we observe in the USA (17%) and Russia (16%). 


Less-educated respondents expressed significantly higher acceptance in the Burkina Faso, Rwanda, Sierra Leone 1, and Uganda 2 samples, which represent the majority of  studies from sub-Saharan Africa. Notably, we observe the opposite pattern in the India, Pakistan 1, and Pakistan 2 samples. In all three studies, acceptance is greater among more educated respondents, although this difference is not statistically significant in the India sample. Education is also a positive and significant predictor of acceptance in the USA.


We find mixed evidence across studies with respect to the relationship between age and COVID-19 vaccine acceptance. In India and Nigeria, respondents younger than 25 years old are significantly less willing to take the vaccine relative to adults 25-54 years old, while in Mozambique, Pakistan 1, and Rwanda, those under 25 are significantly more willing. In Mozambique and Rwanda, respondents under 25 are also significantly more accepting compared to those 55 and over; however, the difference between these age groups is not statistically significant in other LMIC samples. In the USA and Russia, older respondents are consistently more accepting than younger respondents. 




To better understand the reasoning behind vaccine acceptance, we asked those who were willing to take the vaccine why they would take it. We summarize these results in \hyperref[tab:yes]{Supplementary Table 2}, with additional details in \hyperref[tab:yesall]{Supplementary Table 3}. The reason most commonly given for vaccine acceptance across samples is personal protection against COVID-19 infection. The average across the LMIC samples is `r yes_all$estimate`% (CI `r yes_all$conf.low`--`r yes_all$conf.high`%) with a median of  `r yes_ %>% pull(estimate) %>% median()`% and a range of `r yes_ %>% pull(estimate) %>% max() - yes_ %>% pull(estimate) %>% min()`%. In every individual study, including the USA (`r yes_usa$estimate`%, CI `r yes_usa$conf.low`--`r yes_usa$conf.high`%) and Russia (`r yes_rus$estimate`%, CI `r yes_rus$conf.low`--`r yes_rus$conf.high`%), this ranks as the most cited reason. In distant second place in the aggregate LMIC analysis is family protection, with an average of `r yes_all_2$estimate`% (CI `r yes_all_2$conf.low`--`r yes_all_2$conf.high`%), a median of `r yes_2 %>% pull(estimate) %>% median()`% and a range of `r yes_2 %>% pull(estimate) %>% max() - yes_2 %>% pull(estimate) %>% min()`%. In comparison to protecting oneself and one’s family,  protecting one’s community does not feature prominently among stated reasons for acceptance. These reasons do not vary substantially by age group, as shown in \hyperref[tab:yes1]{Supplementary Table 4}.


Figure \@ref(fig:fig2paper) summarizes the reasons given by respondents who said they were not willing to take a COVID-19 vaccine. Concern about side effects is the most frequently expressed reason for reluctance in our LMIC samples. This concern is particularly evident among samples from sub-Saharan Africa. In studies  `r side_top$group[1]` (`r side_top$estimate[1]`%, CI `r side_top$conf.low[1]`--`r side_top$conf.high[1]`%), `r side_top$group[2]`  (`r side_top$estimate[2]`%, CI `r side_top$conf.low[2]`--`r side_top$conf.high[2]`%), `r side_top$group[3]` (`r side_top$estimate[3]`%, CI `r side_top$conf.low[3]`--`r side_top$conf.high[3]`%) and `r side_top$group[4]` (`r side_top$estimate[4]`%, CI `r side_top$conf.low[4]`--`r side_top$conf.high[4]`%), more than half of respondents unwilling to take the vaccine cited worries about side effects.  Respondents in Russia (`r side_rus$estimate`%, CI `r side_rus$conf.low`--`r side_rus$conf.high`%) and even more in the USA (`r side_usa$estimate`%, CI `r side_usa$conf.low`--`r side_usa$conf.high`%), frequently report this same concern. 


Study samples `r effective_top$group[1]` (`r effective_top$estimate[1]`%, CI `r effective_top$conf.low[1]`--`r effective_top$conf.high[1]`%), `r effective_top$group[2]`  (`r effective_top$estimate[2]`%, CI `r effective_top$conf.low[2]`--`r effective_top$conf.high[2]`%) and `r effective_top$group[3]` (`r effective_top$estimate[3]`%, CI `r effective_top$conf.low[3]`--`r effective_top$conf.high[3]`%) show relatively high levels of skepticism about vaccine effectiveness among hesitant respondents. This is also true in Russia (`r effective_rus$estimate`%, CI `r effective_rus$conf.low`--`r effective_rus$conf.high`%) and the USA (`r effective_usa$estimate`%, CI `r effective_usa$conf.low`--`r effective_usa$conf.high`%). In addition, some hesitant respondents cite lack of concern about COVID-19 infection as a reason not to be vaccinated. This answer is particularly common in the `r ill_top$group[1]` (`r ill_top$estimate[1]`%, CI `r ill_top$conf.low[1]`--`r ill_top$conf.high[1]`%), `r ill_top$group[2]`  (`r ill_top$estimate[2]`%, CI `r ill_top$conf.low[2]`--`r ill_top$conf.high[2]`%) and `r ill_top$group[3]` (`r ill_top$estimate[3]`%, CI `r ill_top$conf.low[3]`--`r ill_top$conf.high[3]`%) studies.


In Figure \@ref(fig:fig3paper) we report respondents’ most trusted source of guidance when deciding whether to take a COVID-19 vaccine. Results from Figure \@ref(fig:fig3paper) are reproduced in \hyperref[tab:trust]{Supplementary Table 6}, while \hyperref[tab:q4]{Supplementary Table 13} presents a complete description of response recoding from individual studies.


We find striking consistency across studies. In all samples except Rwanda, including those from Russia and the USA, respondents identify the health system as the most trustworthy source to help them decide whether to take the COVID-19 vaccine. The average across LMIC samples is  `r trust_all$estimate`% (CI `r trust_all$conf.low`--`r trust_all$conf.high`%), with a median of  `r trust_%>% pull(estimate) %>% median()`% and a range of `r trust_%>% pull(estimate) %>% max() - trust_%>% pull(estimate) %>% min()`%. Respondents in `r trust_top$group[1]` (`r trust_top$estimate[1]`%, CI `r trust_top$conf.low[1]`--`r trust_top$conf.high[1]`%), `r trust_top$group[2]`  (`r trust_top$estimate[2]`%, CI `r trust_top$conf.low[2]`--`r trust_top$conf.high[2]`%) and `r trust_top$group[3]` (`r trust_top$estimate[3]`%, CI `r trust_top$conf.low[3]`--`r trust_top$conf.high[3]`%) cited  health workers most often. Sierra Leone has the highest trust in health workers and the Ministry of Health, potentially reflecting investments in public health following the 2014-2015 Ebola epidemic [@deserrano].


In  `r trust_fam$group[1]` (`r trust_fam$estimate[1]`%, CI `r trust_fam$conf.low[1]`--`r trust_fam$conf.high[1]`%), `r trust_fam$group[2]`  (`r trust_fam$estimate[2]`%, CI `r trust_fam$conf.low[2]`--`r trust_fam$conf.high[2]`%), `r trust_fam$group[3]` (`r trust_fam$estimate[3]`%, CI `r trust_fam$conf.low[3]`--`r trust_fam$conf.high[3]`%) and `r trust_fam$group[4]` (`r trust_fam$estimate[4]`%, CI `r trust_fam$conf.low[4]`--`r trust_fam$conf.high[4]`%), the next most-cited sources are family and friends. Across the pooled samples, women are 3 percentage points more likely to rely on family and friends than male respondents, though this difference is not statistically significant (\hyperref[fig:genderhist]{Supplementary Figure 1}). 


By contrast, in `r trust_gov$group[1]` (`r trust_gov$estimate[1]`%, CI `r trust_gov$conf.low[1]`--`r trust_gov$conf.high[1]`%), `r trust_gov$group[2]`  (`r trust_gov$estimate[2]`%, CI `r trust_gov$conf.low[2]`--`r trust_gov$conf.high[2]`%), `r trust_gov$group[3]` (`r trust_gov$estimate[3]`%, CI `r trust_gov$conf.low[3]`--`r trust_gov$conf.high[3]`%) and `r trust_gov$group[4]` (`r trust_gov$estimate[4]`%, CI `r trust_gov$conf.low[4]`--`r trust_gov$conf.high[4]`%), government is the second most frequently cited. Religious leaders and celebrities are not seen as the top sources of guidance by many respondents in any sample other than Nepal, where many respondents say they most trust famous people (`r trust_npl$estimate`%, CI `r trust_npl$conf.low`--`r trust_npl$conf.high`%).


Finally, we highlight two idiosyncratic, yet frequently mentioned, trusted sources of information in deciding whether to take a COVID-19 vaccine. In Rwanda, 34% of respondents would most trust “themselves” for guidance, the most frequent response in this sample. In the USA, 14% of respondents cited Joe Biden, then president-elect and therefore excluded from the “government” category, as their most trusted source. 


# Discussion {.unnumbered #discussion}


<!-- Temp subheading: interpretation of results -->
The current study contributes to the emerging picture of global vaccine acceptance by focusing on COVID-19 vaccine attitudes in a set of low-income and lower-middle-income countries. Our findings show variable but broadly high levels of prospective COVID-19 vaccine acceptance across the LMICs we study, using data from `r nice(nrow(filter(df, country!="USA" | country!="Russia")))` respondents in `r length(unique(df$country))-2` studies in ten LMICs in Africa, South Asia, and Latin America. Acceptance across these LMIC samples  averaged `r ans_mean`%, ranging between `r ans_min`% and `r ans_max`% with a median of `r ans_median`%. The two benchmark countries, Russia and the USA, demonstrate lower COVID-19 vaccine acceptance, consistent with lower pre-pandemic vaccine confidence. 




Many metrics and indices measure vaccine acceptance and hesitancy globally [@betsch2018beyond; @domek2018measuring; @shapiro2018vaccine; @gilkey2014vaccination]. Our surveys use measures employed in other COVID-19 vaccine acceptance studies [@fisher2020attitudes;  @Malik2020; @wouters2021challenges; @lazarus2020nature; @kreps2020factors; @shekhar2021covid; @dror2020vaccine] and recommended by the WHO Data for Action guidance [@whodata], allowing for meaningful cross-study and cross-country comparisons. We measure trust in sources of information about COVID-19 vaccination using a measure similar to that used in the Vaccine Confidence Index (VCI), a widely used survey tool [@defigueiredo2020lancet].  




Consistent with other studies, we find higher average vaccine acceptance among men than women [@kreps2020factors; @shekhar2021covid; @dror2020vaccine;  @fisher2020attitudes; @wouters2021challenges]. In contrast to studies focused primarily on higher-income countries, we find no consistently significant differences with respect to age [@kreps2020factors; @dror2020vaccine] or education in our LMIC samples. 




A key contribution of our study relative to the existing literature is its focus on the reasons *why* respondents express intentions to take (or refuse) a COVID-19 vaccine. Other work has highlighted appeals to altruistic behavior or other prosocial motivations to promote vaccine acceptance [@chou2020considering]. Yet we find that the potential risks and benefits to personal well-being feature much more prominently in our respondents’ reasoning, suggesting that appeals about personal protection may be more effective in the countries under study here.


The most commonly stated reason for vaccine refusal is concern about safety (side effects). The vast majority (86%) of our surveys were conducted as reports from Phase 2 and 3 clinical trial data were emerging for the earliest commercially available vaccines, but prior to the first Emergency Use Authorization of any vaccine (Pfizer-BioNTech approved by the USA on December 11, 2020). Early trial data showed that severe adverse effects were extremely rare [@mulligan2020phase; @logunov2021safety; @folegatti2020safety; @voysey2021safety; @baden2021efficacy; @polack2020safety], occurring in fewer than 10\% of people in clinical trials [@wadman2020public]. Our respondents’ concern about side effects could reflect the rapid pace of vaccine development [@callaway2020russia] and limited information available about potential COVID-19 vaccine safety at the time of data collection. These concerns could also reflect worries about mild, yet common and transient side effects, such as fatigue, muscle pain, joint pain and headache.




Intensive media coverage of adverse events may exacerbate concerns about side effects [@stein2017golden]. In particular, new information about rare but severe cases of thrombosis associated with the Astra-Zeneca vaccine that appeared after our data collection period could affect hesitancy levels. This is of particular relevance to LMICs, which are likely to rely on the Astra-Zeneca vaccine in their immunization campaigns through initiatives such as COVAX.   


Concerns about vaccine efficacy, averaging approximately `r filter(effective_top, group=="All") %>% pull(estimate)`% in the LMIC samples, may also reflect a lack of information about vaccines at the time of our surveys. However, we note that respondents in our samples rarely cited conspiracy theories about ulterior motives on the part of corporations, politicians or the pharmaceutical industry, despite attention given to fears about these issues in higher-income countries [@loomba_measuring_2021]. 


<!-- Temp subheading: limitations -->
Our study has several limitations, which we address here. First, our data are not representative of all LMICs. They instead represent a sample of studies in countries where our organizations could quickly and safely mobilize coordinated data collection. Second, samples from the countries we include here are not fully nationally representative. Phone surveys, while necessary during a global pandemic, do not include individuals who reside outside coverage areas, lack access to a cell phone, or do not respond to calls. In addition, as shown in Table \ref{tab:sampling}, several studies focus on subnational populations of interest from pre-existing studies to which questions about COVID-19 vaccination were added. Particular care should be taken in extrapolating these findings to national populations. 


In spite of this variation in sample composition, our main findings---of high COVID-19 vaccine acceptance in our LMIC samples relative to the USA and Russia---are remarkably consistent across studies. We conduct several robustness checks to probe the sensitivity of our aggregate LMIC analysis to the inclusion of particular samples. First, as shown in \hyperref[fig:mainloo]{Supplementary Figure 3}, we re-estimate aggregate vaccine acceptance across our LMIC samples successively dropping one and two study samples at a time. Regardless of which samples are excluded, the average vaccine acceptance rate among LMIC samples remains consistent and considerably higher than in the USA and Russia, demonstrating that our results are not driven by the peculiarities of one or two studies. Second, we repeat the same analysis excluding all samples that are sub-national in scope, which yields a mean acceptance rate of `r ans_national %>% pull(estimate)`% (CI `r ans_national %>% pull(conf.low)`--`r ans_national %>% pull(conf.high)`%), as shown in the row “All LMICs (National samples)” in Figure \@ref(fig:mainfigure1) and \hyperref[tab:maintabledis]{Supplementary Table 1}. 


The expressed intentions to take a COVID-19 vaccine that we document in our LMIC samples, if translated into behavior, would meet or exceed the current herd immunity threshold for COVID-19 in many, but not all, countries (estimated to be between 70 and  80%, based on the predominant variant in circulation in different countries)[@ontario; @aschwanden2021five; @mcneil_2021]. Reported intentions may however not always translate into vaccine uptake [@mceachanetal2011]. The high salience of COVID-19 may have increased reported intentions. Conversely, reports about side effects and risks associated with expedited vaccine development may have increased hesitancy. The fast-moving pandemic and vaccine development context may change perceptions about vaccines by the time they are widely available in LMICs. 


Indeed, previous research on vaccine hesitancy has emphasized how concerns that arise surrounding vaccination campaigns are often case- and context-specific [@larson2011addressing], making it difficult to predict exactly how COVID-19 vaccines will be received in any given setting. The lower COVID-19 vaccine acceptance rates we observe in Russia and the USA, for example, may reflect the politicization of this specific pandemic and of vaccine development [@hornsey2020donald; @bokemper2021timing; @aaslund2020responses; @burki2020russian], in addition to generally greater vaccine skepticism.






<!-- Temp subheading: implications of results -->


Nonetheless, our findings suggest several concrete implications for policy relating to vaccine roll-out in LMICs. First and foremost, we document high levels of COVID-19 vaccine acceptance in our LMIC samples compared to Russia and the USA. While global vaccine distribution has skewed heavily toward higher-income countries to date [@wouters2021challenges],  our findings suggest that prioritizing distribution to LMICs is justified not only on equity grounds, but on the expectation of higher marginal returns in maximizing global coverage more quickly.


The high stated acceptance rates we document also imply that, once vaccine distribution to LMICs begins in earnest, interventions should focus on converting positive intentions into action. Straightforward, low-cost nudges may be effective in this regard. Two recent large-scale studies in the USA found that vaccination appointment reminder messages from healthcare providers increased influenza vaccine uptake [@milkmanetal2021a; @milkmanetal2021b]. Similar interventions have proven effective in increasing immunization in LMIC contexts. In Ghana and Kenya, vaccination reminders combined with small cash incentives increased childhood immunization coverage [@levine2021; @gibsonetal2017]. Cash and in-kind incentives programs were also effective in Nigeria and India [@idinsight; @Banerjee2010].


This recommendation is consistent with accepted frameworks, such as the WHO’s Behavioral and Social Drivers of vaccination (BeSD) model, which suggest leveraging favorable intentions through reminders and primes, and reducing access barriers when the vast majority of people intend to be vaccinated [@brewer2017increasing; @whodata]. Particularly since COVID-19 vaccination may be more collectively than individually optimal, ease of access is critical to achieve high coverage [@betsch2013inviting]. Availability of single-dose vaccines could be advantageous in settings with high vaccination demand but relatively low-capacity healthcare systems, as is the case in many LMICs.


Our findings also suggest directions for the design and delivery of messaging to address remaining COVID-19 vaccine hesitancy in the countries under study here. Persuasion campaigns may be particularly important in countries where acceptance rates are below herd immunity thresholds. We highlight three potential implications for messaging below, but urge policymakers and stakeholders to utilize country-specific results to develop further strategies that may work best in their particular context. We also echo calls for integrating rigorous impact evaluation of vaccine hesitancy interventions in all contexts, given limited evidence to date [@dube2015strategies].


First, our data strongly support the view that respondents from the included set of LMICs prefer to follow the guidance of actors with the most relevant knowledge and expertise. We find high levels of trust in health workers, which suggests that social and behavioral change communication (SBCC) strategies engaging local health workers may be particularly effective in combating remaining hesitancy [@who2020behavioral; @bokemper2021timing]. Health workers have also been the first group to receive the COVID-19 vaccine and are therefore best positioned to share locally credible experiences of vaccination [@katzman2021primary]. While celebrities were rarely identified as a most trustworthy source for COVID-19 advice in our study, celebrity endorsements have proven effective in other contexts and may complement a strategy that primarily focuses on health workers [@alatas2019celebrities].  


Second, our findings offer some guidance on the specific content of vaccine messaging that is likely to be most effective in persuading those who may be hesitant. Hesitant respondents were most concerned about side effects and vaccine efficacy. This suggests that proactive messaging, initiated before large-scale vaccination campaign roll-out, should highlight the high efficacy rates of the COVID-19 vaccines currently on the market in reducing or eliminating disease, hospitalizations, and death, and communicate accurate information about potential side effects, including the rarity of severe adverse events that may have contributed to hesitancy through widespread media coverage [@goldstein2015health; @puri2020social].


Third, consistent with previous studies on COVID-19 vaccination [@kreps2020factors; @shekhar2021covid; @dror2020vaccine;  @fisher2020attitudes; @wouters2021challenges] our study finds on average lower vaccine acceptance among women than men, suggesting that messaging strategies focusing on women may be important in addressing overall hesitancy. Recent work in Latin America on COVID-19 vaccine messaging found that the provision of basic information about the vaccines was particularly effective in persuading hesitant women [@argote2021messaging]. More generally, countries may consider tailoring their messaging campaigns to address concerns held by more hesitant sub-populations, which vary across our samples with respect to age, gender and education. Additional research is needed to identify and design effective messaging campaigns in order to overcome hesitancy among specific subpopulations in each setting [@goldstein2015health; @dube2015strategies].


Finally, high coverage rates of existing vaccines, coupled with respondents’ reliance on friends and family as information sources, suggest that the general pro-vaccination stance of many LMIC citizens could be leveraged to increase uptake of COVID-19 vaccines as they become available. This might yield particularly strong results in Colombia and Nepal, where family members and friends are seen as an important source of advice when deciding whether to take a COVID-19 vaccine. Social learning strategies and norm-setting are powerful drivers of behavior in many related sectors. Social signaling of positive attitudes towards vaccines may help shift social norms toward even greater immunization acceptance and uptake in the community at large [@van2020using; @karing2018social; @ashraf2014no]. As with messaging, policymakers should consider designing and evaluating social mobilization strategies targeted toward more hesitant subgroups [@jalloh2020mobilize].


\clearpage


# Methods {-}


## Survey questions and sample construction {-}




Survey data were collected between June 2020 and January 2021. Our main outcome measure is vaccine acceptance. Across studies, we asked respondents, “If a COVID-19 vaccine becomes available in [your country], would you take it?”. This measure aligns with widely reported COVID-19 vaccine acceptance measures [@fisher2020attitudes;  @Malik2020; @wouters2021challenges; @lazarus2020nature; @kreps2020factors; @shekhar2021covid; @dror2020vaccine]. If the respondent answered yes to this question, we followed up with the question, “Why would you take it? [the COVID-19 vaccine]”. If the respondent said they would not be willing to take the vaccine, we followed up with the question, “Why would you not take it? [the COVID-19 vaccine]”. Finally, regardless of their expressed willingness to take the vaccine, we asked about actors and institutions that would be most influential in their decision: “Which of the following people would you trust MOST to help you decide whether you would get a COVID-19 vaccine, if one becomes available?” following [@defigueiredo2020lancet]. To examine heterogeneity across demographic strata, we collected information about gender, age, and education. Slight variations in question wording and answer options across studies are documented in Supplementary Appendix.


Studies vary in terms of geographic scope, sampling methodology, and survey modality. Seven were national or nearly-national in scope. Studies from Burkina Faso, Colombia, Rwanda, and Sierra Leone ("Sierra Leone 1") used nationally-representative samples of active mobile phone numbers reached through Random Digit Dialing (RDD). Studies in the USA and Russia were conducted online using quota samples obtained from private survey companies.


The remaining eight studies targeted sub-national populations. One study from Pakistan ("Pakistan 2") used RDD in Punjab province. Respondents in Mozambique, Nigeria, Pakistan ("Pakistan 1"), Uganda ("Uganda 1","Uganda 2"), India, Nepal and Sierra Leone ("Sierra Leone 2”) were drawn from pre-existing studies to which COVID-19 vaccine questions were subsequently added. For example, Sierra Leone 2 has national coverage from a study on access to electricity and Uganda 1 sampled female caregivers of households in rural and semi-rural villages as part of a large ongoing cluster-RCT implemented across 13 districts. 


Table \ref{tab:sampling} summarizes the dates of data collection, geographic scope, sampling methodologies and survey modalities of all 15 studies.


All surveys were conducted remotely to minimize in-person contact and comply with social distancing guidelines. Interviews were conducted by local staff in each country in local language(s). Surveying by phone made rapid, large-scale data collection possible. In two samples, the USA and Russia, surveys occurred via online polling. All surveys lasted approximately 15 to 40 minutes.


Taken together, we have data from `r nice(nrow(filter(df, country!="USA" & country!="Russia")))` individuals from 10 LMICs and `r nice(nrow(filter(df, country=="USA" | country=="Russia")))` from the USA and Russia, for a total of `r nice(nrow(df))` respondents.


# Statistical Analysis {-}


Vaccine acceptance was defined as the percentage of respondents who answered “yes” to the question, “If a COVID-19 vaccine becomes available in [country], would you take it?”. This was calculated combining all other answer options (“No”, “Don’t Know” and “Refuse”) into a single reference category. We estimated average acceptance for each individual sample via ordinary least squares (OLS) weighted by respective study population weights and using robust standard errors clustered at the level of the sampling cluster. 


In addition to study-level estimates, we combined data from all studies other than the USA and Russia to calculate an aggregate “All LMIC studies” estimate. For these analyses, we estimated average acceptance by OLS with weights for each study normalized such that the total weight given to observations was constant across studies. Robust standard errors for these analyses were clustered at the study level. 


We note the core results would be virtually unchanged at `r round(100*unit_check$estimate[2], 1)`% (`r round(100*unit_check$conf.low[2], 1)`--`r round(100*unit_check$conf.high[2], 1)`) rather than `r round(100*unit_check$estimate[1], 1)`% (`r round(100*unit_check$conf.low[1], 1)` -`r round(100*unit_check$conf.high[1], 1)`) using countries rather than studies as groups in the pooled analysis, that is, if we set weights so that the sum of weights in each country (rather than in each study) sum to a constant and cluster standard errors at the country level (rather than the study level). 


In this combined analysis, we also estimated the underlying heterogeneity of vaccine acceptance across studies using the between studies variance estimator $\tau^2$ from a random effects model.


We conducted subgroup analyses by gender, age and education level and reported differences between groups. For age, we selected cut-offs below 25, between age 25 and 54, and above 55 years old, closely following the age breakdown proposed by recent work on COVID-19 vaccine acceptance [@lazarus2020nature]. However, the lower life expectancy (63 years on average [@wbage]) and younger-skewing populations (only 5% of the population is above 65 years old [@WorldBanksubSahara]) of low-income countries in particular, precluded further disaggregation at the upper end of the age distribution. For education, we divided the sample between respondents who had completed secondary school and those that had not. We defined these two groups to reflect broader schooling trends in LMICs, where out of every 100 students entering primary education, 61% complete lower secondary education [@world2017world]. The subgroup analyses estimates are calculated in exactly the same way as the overall acceptance rate---with weights again normalized to sum to a constant within each study---with the exception that the subsample used in the analysis is limited to those respondents fitting each demographic group.


We then investigated stated reasons for COVID-19 vaccine acceptance and hesitancy, and the types of actors respondents would trust most when making the decision about whether to take a COVID-19 vaccine. We report estimates of agreement with reasons for vaccine acceptance/hesitancy and trust in actors for individual studies and for the “All LMICs” group, which includes all study samples except Russia and the USA. Estimates were calculated with the same procedure as above, varying only the quantity of interest; i.e. one model is run for each reason why a respondent would (or would not) take the vaccine and each trusted actor. 




\newpage


# Data Availability Statement {-}
Individual participant data (de-identified) that underlie the results reported in this article, are available without restrictions at \url{https://github.com/wzb-ipi/covid_vaccines_nmed}. 


# Code Availability {-}
All code has been deposited into the publicly available GitHub repository at \url{https://github.com/wzb-ipi/covid_vaccines_nmed}. The code and output for all analyses can be easily inspected at \url{https://wzb-ipi.github.io/covid_vaccines_nmed/replication.html}.


# Acknowledgements {-}




This work would not have been possible without the hard work and ingenuity of many. For valuable contributions to data collection during the challenging context of the global pandemic, we would like to thank:


Burkina Faso: IPA Burkina Faso staff in Ouagadougou for their support of the RECOVR (Research for Effective Covid Response) survey from which the questions used in this study were drawn.


Colombia: IPA Colombia staff in Bogotá for their support of the RECOVR (Research for Effective Covid Response) survey from which the questions used in this study were drawn.


India: Pronab Sen (Country Director), Kumar Das (Programme Policy Manager) and Bijeta Mohanty (Country Economist) at IGC India and Morsel Research and Development’s field team. 


Mozambique:Frederica Mendonça and Simão Paiva from NOVAFRICA Mozambique office.


Nepal: Field staff at the Centre for the Study of Labour and Mobility (CESLAM) in Kathmandu.


Nigeria:  Segun Fayomi and Aishwarya Kumar at the Busara Center for Behavioral Economics, and Sarah Ryan (University of Illinois at Chicago) for research assistance. 


Pakistan 1: Staff at the Institute for Development and Economic Alternatives (IDEAS) in Lahore for their contributions to the Sheikhupura police study sample from which the questions used in the Pakistan 1 study were drawn.


Pakistan 2: Staff of the internal survey firm at the Centre for Economic Research Pakistan (CERP) for their contributions to the Economic Vulnerability Assessment (EVA) project from which the questions used in the Pakistan 2 study were drawn.


Rwanda: IPA Rwanda staff in Kigali for their support of the RECOVR (Research for Effective Covid Response) survey from which the questions used in this study were drawn.


Russia: Staff at Online Market Intelligence Survey Agency, Kirill Chmel (HSE University) and Vladimir Zabolotsky (HSE University and University of Bologna) for their intellectual contributions and research assistance.


Sierra Leone 1: IPA Sierra Leone staff in Freetown for their support of the RECOVR (Research for Effective Covid Response) survey from which the questions used in this study were drawn.


Sierra Leone 2: Junisa Nabieu (Field Supervisor), Momoh Flee (Field Supervisor), Hamid Kai, Mohamed Lamin (Field Supervisor) and all enumerators at IGC-Sierra Leone Freetown office and Wageningen University’s Project “Tracking the Economic Consequences of COVID-19 in Sierra Leone”. 


Uganda 1: Staff at IPA Uganda for their intellectual contributions, research assistance, and support throughout the study.


Uganda 2: Co-PIs Ana Garcia Hernandez, Paul Kiwanuka-Mukiibi, Leah Rosenzweig, Lily Tsai, and Elly Atuhumuza and the team of survey enumerators and support staff at IPA Uganda.


United States: Staff at Lucid (survey firm) for distributing the survey and providing the online survey sample.


IGC thanks their resident country teams in Sierra Leone, Uganda, Pakistan, and Mozambique for their continued guidance, coordination of activities with researchers and direct engagement and dissemination activities with national policymakers: 


* Herbert McLeod (Country Director) and Abou Bakarr Kamara (Country Economist) at IGC Sierra Leone
* Richard Newfarmer (Country Director), Priya Manwaring (Country Economist), and Jakob Rauschendorfer (Country Economist) at IGC - Uganda
* Ijaz Nabi (Country Director), Hina Shaikh (Country Economist), and Usman Naeem (Country Economist) at IGC - Pakistan
* Claudio Frischtak (Country Director)  and Egas Daniel (Country Economist) at IGC - Mozambique


IPA thanks staff from its global offices for their research assistance and support throughout the RECOVR survey. At IPA New York and Washington, DC, Michael Rosenbaum and Savanna Henderson provided research support.


NOVAFRICA thanks Cátia Batista (NOVA SBE), Sandra Sequeira (LSE) and Inês Vilela (Royal Holloway, University of London and NOVAFRICA) for their intellectual contributions. 


# Contributors {-}
JSo, SW, NMe, AS, NMc, GS, MV and AAM are co-first authors. DK, MC, MT, MH, AMM and SO are co-last authors. AMM and SO are also the corresponding authors. 
DK, AMM, MT, NMe, MC, and MV conceived of the study and provided overall guidance.
SAb and NMe led the literature search, with input from AS, NMc, SW, AMM, AAM and JSo.
SW, NMe, AS, NMc, MV, GS, AA, SA, BA, AB, EB, CMB, AC, EC, MF, AG, AK, SK, RL, MBN, MP, JQ, JSh, JSv, PV, LB, BZ, MC, SAs, AC, AF, AH, MC, MT, MH,CV, LW, BZ and BZa oversaw data collection as part of other research efforts.
OA, DA, MA, MAw, MCG, AC, FC, GE, MG, SJA, SKa, AK, AKh, SM, GM, LM, FM, AMu, IM, JN, IO, MJO, BWO, TP, LP,  MR, IR, TS, SS, , AT, AMT, HT, and BT implemented the surveys, including training and oversight of enumerators and management of sampling for the surveys.
SW, NMe, and MT coordinated the project across study samples.
The following verified the underlying data for individual study samples: EC (Burkina Faso, Colombia, Rwanda, Sierra Leone 1), BA and AB (India), AS and RL (Nigeria), AG, JSv and MBN (Uganda 1), CMB and MH (Uganda 2), NMe and MV (Sierra Leone 2), GS (Russia), MF (Mozambique), AF and JSh (Pakistan 1), SAs (Pakistan 2), CV (Nepal), and NMc (USA).
JSo, GS, MH and SA collated and processed all datasets used for the analysis.
NMe, MH, AMM, JSo, GS, SW, AS, EC, EB, MT, MV and NMc did the data interpretation with guidance from SO and AM.
JSo, GS, EC and MH verified final datasets and analysis.
JSo and GS did the data analysis and produced output figures with input from MH, AMM, DK, SW, EC, MV, NMe and MT.
MH supervised the data analysis.
JSo, SW, NMe, AMM, AS, NMc and MV wrote the first draft of the manuscript, with guidance from AAM and SO.
JSo, SW, NMe, AS, NMc, MV, SAb, EB, MP, JSh, PV, BZ, MC, MT, MH, AMM, AAM and SO revised the manuscript.
All authors approved the final version of the manuscript. All authors had full access to all the data used in this study and had final responsibility for the decision to submit for publication.


# Funding {-}
<!-- STROBE; 22- Source of funding and the role of the funders for the present study and, if applicable, for the original study on which the present article is based  -->


Funding was provided by Beyond Conflict, Bill and Melinda Gates Foundation, Columbia University, Givewell.org, Ghent University, HSE University Basic Research Program, International Growth Centre, Jameel Poverty Action Lab Crime and Violence Initiative, London School of Economics and Political Science, Mulago Foundation, NOVAFRICA at the Nova School of Business and Economics, NYU Abu Dhabi, Oxford Policy Management, Princeton University, Social Science Research Council, Trinity College Dublin COVID19 Response Funding, UK Aid, UKRI GCRF/Newton Fund, United Nations Office for Project Services, Weiss Family Fund, WZB Berlin Social Science Center, Yale Institute for Global Health, Yale Macmillan Center, and anonymous donors to IPA and Y-RISE. 




# Role of the funding source {-}


None of our funders played any role in the collection, analysis, interpretation, writing or decision to submit this article for publication.


# Declaration of interests {-}
We declare no competing interests.


---
nocite: '@*'
...


\newpage
# References {-}


<div id="refs"></div>








 
```{=latex}
\begin{table}[!h]
\caption{\label{tab:summary}Policy summary}
\centering
\resizebox{\linewidth}{!}{
\fontsize{10}{12}\selectfont
\begin{tabular}[t]{>{\raggedright\arraybackslash}p{8em}>{\raggedright\arraybackslash}p{30em}}
\toprule
Background & 
We analyze COVID-19 vaccine acceptance, hesitancy and their drivers across 15 survey samples covering ten low- and middle-income countries (LMICs) in Asia, Africa, and South America, as well as Russia and the United States, comprising `r nice(nrow(df))` individuals. \\
\\
Main findings and limitations & Willingness to take a COVID-19 vaccine is considerably higher in the LMICs in our sample than in the United States and Russia. The personal protective benefit of vaccination is the most frequently cited reason for vaccine acceptance. Concern about side effects is the most commonly cited reason for vaccine hesitancy. Health workers are considered the most trusted sources of guidance about COVID-19 vaccine choices.
Our data are not representative of all LMICs, and some individual samples are not nationally representative. However, our main findings---of high COVID-19 vaccine acceptance in our LMIC samples relative to the USA and Russia---are remarkably consistent across samples and specifications. \\
\\
Policy implications & While global vaccine distribution has skewed heavily toward higher income countries to date, the high levels of vaccine acceptance we identify suggest that prioritizing distribution to LMICs may be an efficient way to achieve immunity on a global scale, and prevent novel variants from emerging.
Vaccination campaigns should focus on converting positive intentions into uptake, which may require investment in ‘last-mile’ delivery. Engaging health workers to deliver vaccine information, leveraging pro-vaccine norms, and messaging focused on vaccine effectiveness and safety may be effective in addressing remaining hesitancy.\\
\bottomrule
\end{tabular}}
\end{table}
```


\newpage
```{r sampling}
tab_sampling
```


\newpage
```{r otherv}
tab_1b
```
\newpage


```{=latex}
\begin{figure}[!ht]
\caption{Acceptance rates overall and broken down by respondent characteristics \label{fig:mainfigure1}}
```
```{r mainfigure1, fig.height=7.5, fig.width = 9, warning = FALSE, dev='tikz', echo = FALSE}
fig_1_ages
```
```{=latex}
\scriptsize{Figure 1 presents average acceptance of the COVID-19 vaccine across studies and subgroups within studies. For each study, we summarize sampling information in parentheses in the following way: First, we indicate whether the geographic coverage of the sample is national or subnational. If the coverage is subnational we provide further details. Second, we list the number of observations included in the study. Data are presented as mean values +/- SD. In the plot, points represent the estimated percentage of individuals who would take the vaccine. ``No'', ``Don't know'' and ``Refuse'' are taken as a single reference category. Bars around each point indicate a 95\% confidence interval for the estimate. An estimate of average acceptance for all studies in LMICs (excluding USA and Russia) is also shown.}
\end{figure}
```




```{=latex}
\begin{figure}[!ht]
\caption{Reasons not to take the vaccine \label{fig:fig2paper}}
```
```{r fig2paper, fig.height=7.5, fig.width = 10, warning = FALSE, dev='tikz', echo = FALSE}
fig_2
```
```{=latex}
\scriptsize{Figure 2 shows the percentage of respondents mentioning reasons why they would not take the COVID-19 vaccine. For each study, we present the sample size in parentheses. Data are presented as mean values +/- SD. In the plot, points represent the estimated percentage of individuals that would not take the vaccine or do not know if they would take the vaccine for each possible response option. Bars around each point indicate a 95\% confidence interval for the estimate. An estimated average for all studies in LMICs is also shown. Studies India and Pakistan 2 are not included because they either did not include the question or were not properly harmonized with the other studies.}
\end{figure}
```


```{=latex}
\begin{figure}[!ht]
\caption{Trusted actors and institutions, broken down by expressed willingness to take a COVID-19 vaccine \label{fig:fig3paper}}
```


\newpage


```{r fig3paper, fig.height=9, fig.width = 10, dev='tikz', echo = FALSE}
fig_hist2
```
\newpage




```{=latex}
\scriptsize{Figure 3 shows histograms of actors and institutions respondents say they would trust most to help them decide whether to take the COVID-19 vaccine. Respondents were only permitted to select one most trusted actor or institution. Studies India, Mozambique, Pakistan 1, Pakistan 2 and Uganda 1 are not included because they either did not include the question or were not properly harmonized with the other studies.}
\end{figure}
```


\newpage


```{=latex}
\begin{figure}[!ht]
\caption{Trusted actors and institutions, broken down by gender \label{fig:genderhist}}
```
```{r genderhist, fig.height=9, fig.width = 10, warning = FALSE, dev='tikz'}
hist_gender
```
```{=latex}
\scriptsize{Figure S1 shows histograms of actors and institutions that respondents say they would trust most to help them decide whether or not to take the COVID-19 vaccine. Respondents were only permitted to select one most trusted  actor or institution. Responses are broken down by gender of respondent.}
\end{figure}
```


\pagebreak


```{=latex}
\begin{figure}[!ht]
\caption{Trusted actors and institutions, broken down by vaccine acceptance \label{fig:vaccacc}}
```
```{r histcattrust, fig.height=9, fig.width = 10, warning = FALSE, dev='tikz'}
fig_hist_categories
```
```{=latex}
\scriptsize{Figure S2 shows histograms of actors and institutions that respondents say they would trust most to help them decide whether or not to take the COVID-19 vaccine. Respondents were only permitted to select one most trusted  actor or institution. Responses are broken down by acceptance of the COVID-19 vaccine. The color of the bars reflect the answer given to the question ``If a COVID-19 vaccine becomes available in [country], would you take it?'' with ``No'' and ``Don't know'' pooled together and ``All'' combined average of ``Yes'' and ``No / Don't know''}
\end{figure}
```


\pagebreak


```{=latex}
\begin{figure}[!ht]
\caption{Average vaccine acceptance across all LMIC countries leaving one or two study samples out
 \label{fig:mainloo}}
```
```{r mainloo, fig.height=10, fig.width = 9, warning = FALSE, dev='tikz', echo = FALSE}
fig_hist_loo
```
```{=latex}
\scriptsize{Figure S3 shows distribution of estimates of average acceptance for all studies in LMICs (excluding USA and Russia) leaving one and two study samples out at a time. Figure also shows distributions of subgroup averages by gender, education and age leaving one and two study samples out at a time. To directly compare the resulting distributions to the estimates reported in Figure 1, we plot point estimates reported in Figure 1 for all LMIC studies, Russia and the US.}
\end{figure}
```
\pagebreak






\clearpage




# Supplementary Appendix {-}




\beginsupplement
\setcounter{page}{1}


\newpage


```{r, results = "asis"}
tab_fig1
```


\newpage




```{r yes, results = "asis"}
tab_reasons_y
```


\newpage




```{r yesall, results = "asis"}
tab_reasons_y_all
```


\newpage


```{r yesall1, results = "asis"}
tab_reasons_y_age
```


\newpage




```{r no, results = "asis"}
tab_fig2
```


\newpage




```{r trust, results = "asis"}
tab_trust
```




\pagebreak


```{r, results = "asis"}
dmeans
```


\pagebreak


```{r, results = "asis"}
nas
```


\newpage


```{r tcds}
t_country
```


\pagebreak




________________




\clearpage








```{r tabq1}
tab_question1
```


```{r tabq2}
tab_question2
```


```{r tabq3}
tab_question3
```


```{r tabq4}
tab_question4
```
